% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_classes_getters.R
\name{abundance}
\alias{abundance}
\alias{abundance,mgnet-method}
\alias{abundance,mgnetList-method}
\alias{abundance,mgnet,missing-method}
\alias{abundance,mgnet,character-method}
\alias{abundance,mgnetList,missing-method}
\alias{abundance,mgnetList,character-method}
\title{Retrieve Abundance Data at Specified Taxonomic Rank}
\usage{
abundance(object, rank = "missing")

\S4method{abundance}{mgnet,missing}(object)

\S4method{abundance}{mgnet,character}(object, rank = "missing")

\S4method{abundance}{mgnetList,missing}(object)

\S4method{abundance}{mgnetList,character}(object, rank = "missing")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{rank}{(Optional) A character string specifying the taxonomic rank of interest
for data aggregation. If not provided or missing, the function returns the
original abundance data without aggregation.}
}
\value{
For an \code{mgnet} object, a numeric matrix of abundance data, aggregated at the
specified taxonomic rank if provided. For an \code{mgnetList} object, a list of numeric
matrices, each representing the aggregated or original abundance data from each
\code{mgnet} object within the list.
}
\description{
Retrieves abundance data for an \code{mgnet} object, returning the abundance matrix
that represents the raw or processed counts of different taxa observed in each sample.
If a specific taxonomic rank is chosen, the function aggregates the abundance data
by summing the counts of taxa classified under the same category at the specified rank,
providing a summarized view of abundance data at that rank. If no rank is specified,
or if the rank is missing, the function returns the original abundance matrix without
aggregation. This allows for flexibility in analyzing data at various levels of
taxonomic resolution.
}
\details{
For an \code{mgnetList} object, it returns a list where each element corresponds to an
\code{mgnet} object within the list, containing either the aggregated abundance data at
the specified rank or the original abundance matrix if no rank is specified. This
enables consistent data retrieval across multiple samples or experiments encapsulated
within an \code{mgnetList}.

This function differs from \code{taxa_id}, which refers to the unique identifiers
associated with taxa at the finest taxonomic resolution (e.g., OTUs, species).
While \code{taxa_id} often corresponds to the finest resolution, \code{taxa_name} at the same
level might provide more descriptive names of the entities, and \code{abundance} at a
specified rank leverages this taxonomic information to aggregate data accordingly.
}
