% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_adjacency.R
\name{adjacency_absolute_value}
\alias{adjacency_absolute_value}
\title{Generate Adjacency Matrix by Absolute Value Threshold}
\usage{
adjacency_absolute_value(x = NULL, r = NULL, method = "pearson", th = 0.3)
}
\arguments{
\item{x}{Optional; a numeric matrix of data from which to calculate the correlation matrix.
Provide either 'x' or 'r', but not both. If 'x' is provided, 'r' should be NULL.}

\item{r}{Optional; a pre-computed numeric symmetric correlation matrix with elements in
the range [-1,1]. Provide either 'x' or 'r', but not both. If 'r' is provided, 'x' should be NULL.}

\item{method}{Character; the method for computing correlation if 'x' is provided.
Acceptable values are "pearson" (default), "spearman", or "kendall".}

\item{th}{Numeric; the threshold for determining significance in the adjacency matrix.
Must be a number in the range [0,1]. Correlations with absolute values greater than or equal
to this threshold are considered significant.}
}
\value{
A numeric adjacency matrix with the same dimensions as the correlation matrix 'r'.
Elements are set to the value of the correlation if the absolute value of the correlation
is greater than or equal to 'th', and set to zero otherwise.
}
\description{
This function creates an adjacency matrix by applying a threshold to the absolute
values of the correlations in a correlation matrix. Elements of the correlation
matrix greater than or equal to the threshold are preserved, while others are set to zero.
This approach is useful in network analysis, where an edge between two nodes (variables)
is considered significant if their correlation's absolute value exceeds the specified threshold.
}
