% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_adjacency.R
\name{adjacency_edge_density}
\alias{adjacency_edge_density}
\title{Generate Adjacency Matrix by Edge Density Threshold}
\usage{
adjacency_edge_density(x = NULL, r = NULL, method = "pearson", th = 0.05)
}
\arguments{
\item{x}{A numeric matrix of data, where each row represents a sample and each column represents a variable.
Used for computing the correlation matrix if \code{r} is not provided. Either \code{x} or \code{r} must be provided.}

\item{r}{A pre-computed correlation matrix. If provided, \code{x} should not be specified.}

\item{method}{Character string specifying the correlation method ("pearson", "spearman", "kendall").
Used only if \code{x} is provided. Default is "pearson".}

\item{th}{Numeric value specifying the desired edge density threshold in range [0,1]. Default is 0.05.}
}
\value{
An adjacency matrix where only correlations that meet the edge density threshold are retained.
}
\description{
Creates an adjacency matrix from abundance data or a pre-computed correlation matrix
by applying a threshold based on the desired edge density of the resulting network.
}
