% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{adjacency_list}
\alias{adjacency_list}
\alias{adjacency_list,mgnet-method}
\alias{adjacency_list,list-method}
\title{Retrieves adjacency edge list}
\usage{
adjacency_list(object)

\S4method{adjacency_list}{mgnet}(object)

\S4method{adjacency_list}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return the adjacency list associated to netw slot as data.frame using the
igraph function as_data_frame.
}
