% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{adjacency_matrix}
\alias{adjacency_matrix}
\alias{adjacency_matrix,mgnet-method}
\alias{adjacency_matrix,list-method}
\title{Retrieves adjacency matrix.}
\usage{
adjacency_matrix(object)

\S4method{adjacency_matrix}{mgnet}(object)

\S4method{adjacency_matrix}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return the adjacency matrix associated to netw slot using the igraph function
as_adjacency_matrix.
}
