% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{adjacency_matrix<-}
\alias{adjacency_matrix<-}
\title{Assign a new netw from adjacency matrix to \code{object}}
\usage{
adjacency_matrix(object) <- value
}
\arguments{
\item{object}{mgnet.}

\item{value}{\code{\link{igraph}}}
}
\description{
Assign a new netw from adjacency matrix to \code{object}
}
