% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency-tools.R
\name{adjacency_p_adjust}
\alias{adjacency_p_adjust}
\title{Get adjacency with corrected p-values}
\usage{
adjacency_p_adjust(
  x = NULL,
  r = NULL,
  n = NULL,
  method = "pearson",
  adjust = "holm",
  alpha = 0.05,
  keep.psych = FALSE
)
}
\arguments{
\item{x}{matrix of data.}

\item{r}{correlation matrix.}

\item{n}{number of observation.}

\item{method}{(Optional) correlation type, possible choices are "pearson",
"spearman" or "kendall" (default spearman).}

\item{adjust}{(Optional) multiple test choosen, possible choices are "holm",
"hochberg","hommel","bonferroni","BH","BY","fdr","none".}

\item{alpha}{(Optional) level of confidence interval.}

\item{keep.psych}{(Optional) logical paramater which indicates if return the
results of psych results (default FALSE).}
}
\description{
User wrapper of \code{\link{corr.test}} or \code{\link{corr.p}}
to get the adjacency from abundances or correlation matrix. The function apply a
threshold on correlation matrix using the adjusted p-value with multiple test
corrections.
}
