% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4g_classes_aggregate_taxa.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\alias{aggregate_taxa,mgnet,character-method}
\alias{aggregate_taxa,mgnetList,character-method}
\title{Aggregate Data at Higher Taxonomic Levels}
\usage{
aggregate_taxa(object, rank)
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{rank}{Taxonomic level chosen for aggregation.}
}
\value{
An updated \code{mgnet} object with aggregated data and taxa at the specified taxonomic level.
The slots \code{info_taxa}, \code{network}, and \code{community} are lost after the aggregation process.
}
\description{
This function aggregates abundance data and lineage at a higher taxonomic level specified
by the user. It calculates the sum of data abundances over all taxa that map to the same
higher-level group and removes any ambiguous levels from the taxonomy table.
}
