% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis_apply_verbs.R
\name{apply_filter_verb}
\alias{apply_filter_verb}
\title{Apply Filters with Abundance Check}
\usage{
apply_filter_verb(metadata, long_abun, expressions, .by, sample_or_taxa)
}
\arguments{
\item{long_abun}{A data frame containing abundance-related data,
used for abundance-related expressions.}

\item{expressions}{A list of expressions to apply to the data.}

\item{.by}{A character vector indicating the grouping variable(s).}

\item{sample_or_taxa}{...}

\item{info_mutated}{A data frame representing the metadata.}
}
\value{
A mutated version of the input \code{info_mutated} data frame.
}
\description{
This internal function applies filters to a data frame, handling both
abundance-related and non-abundance-related expressions.
}
\keyword{internal}
