% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_community.R
\name{are_isolated}
\alias{are_isolated}
\title{Check the isolated nodes}
\usage{
are_isolated(comm)
}
\arguments{
\item{comm}{An object of class \code{communities}.}
}
\value{
a logical vector that indicate isolated nodes.
}
\description{
Check the isolated nodes
}
