% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\name{are_lists_assign}
\alias{are_lists_assign}
\title{Check lists in assign methods}
\usage{
are_lists_assign(object, value)
}
\arguments{
\item{object}{mgnet list}

\item{value}{list}
}
\description{
Internal function to check if the two list in assign methods are
coherent.
}
\keyword{internal}
