% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{are_selected_links}
\alias{are_selected_links}
\alias{are_selected_links,mgnet-method}
\alias{are_selected_links,mgnets-method}
\title{Check whether links are selected}
\usage{
are_selected_links(object)

\S4method{are_selected_links}{mgnet}(object)

\S4method{are_selected_links}{mgnets}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
For an \code{mgnet}, \code{TRUE} if the \code{selected_links} attribute is set, \code{FALSE}
otherwise.

For an \code{mgnets}, \code{TRUE} only if \strong{all} contained \code{mgnet} objects currently
carry a link selection. This implements an all-or-none collection-level
semantics.
}
\description{
Determine whether an \code{mgnet} or \code{mgnets} object currently carries an active
link selection.
}
