% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{arrange_mgnet}
\alias{arrange_mgnet}
\alias{arrange_mgnet,mgnet-method}
\alias{arrange_mgnet,list-method}
\title{Re-arranging the samples or taxa positioning.}
\usage{
arrange_mgnet(object, i, j)

\S4method{arrange_mgnet}{mgnet}(object, i, j)

\S4method{arrange_mgnet}{list}(object, i, j)
}
\arguments{
\item{object}{mgnet.}

\item{i}{sample indices that are numeric or character vectors or empty (missing).}

\item{j}{taxa indices that are numeric or character vectors or empty (missing).}
}
\description{
The function permits to order differently the samples or the taxa
inside the mgnet object.
}
