% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{arrange_vertices}
\alias{arrange_vertices}
\alias{arrange_vertices,mgnet,matrix-method}
\alias{arrange_vertices,mgnet,matrix}
\alias{arrange_vertices,list,matrix-method}
\alias{arrange_vertices,list,matrix}
\title{Arrange the vertex number}
\usage{
arrange_vertices(obj, new_taxa)

\S4method{arrange_vertices}{mgnet,matrix}(obj, new_taxa)

\S4method{arrange_vertices}{list,matrix}(obj, new_taxa)
}
\arguments{
\item{obj}{\code{\link{mgnet}} object}

\item{new_taxa}{data.frame with taxonomic information about the vertices wanted in new graph.
The dataset must have the same form as taxa in MetaGenomic/MetaGenomicGraph object.
Naturally in new_taxa all the old vertices must be present.}
}
\description{
Edit the vertices/taxa of the \code{\link{mgnet}} object
(for example to compare two different networks)
}
