% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities.R
\name{as.list-mgnetList}
\alias{as.list-mgnetList}
\alias{as.list,mgnetList-method}
\title{Convert an mgnetList Object to a List}
\usage{
\S4method{as.list}{mgnetList}(x, ...)
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{...}{Additional arguments affecting the conversion (currently unused).}
}
\value{
A list where each element is an \code{mgnet} object previously contained
within the \code{mgnetList}. The names of the list elements correspond to
the names of the \code{mgnet} objects in the \code{mgnetList}, if they are named.
}
\description{
This method allows for the conversion of an \code{mgnetList} object into a list,
facilitating the application of list-based operations and functions. Each element
of the list represents an \code{mgnet} object contained within the \code{mgnetList}.
}
\seealso{
\link{mgnetList}, for details on the \code{mgnetList} class.
}
