% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_mgnetList_class.R
\name{as.list.mgnetList}
\alias{as.list.mgnetList}
\title{Convert an mgnetList Object to a List}
\usage{
\method{as.list}{mgnetList}(x, ...)
}
\arguments{
\item{x}{An \code{mgnetList} object to be converted to a list.}

\item{...}{Additional arguments (not currently used).}
}
\value{
A list containing the individual \code{mgnet} objects from the \code{mgnetList}.
}
\description{
This function converts an \code{mgnetList} object to a list by extracting the individual
\code{mgnet} objects contained within the list.
}
