% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{adjacency_list<-}
\alias{adjacency_list<-}
\alias{adjacency_list<-,mgnet,data.frame-method}
\alias{adjacency_list<-,list,list-method}
\title{Assign a new netw to \code{object} from an adjacency edge data.frame}
\usage{
adjacency_list(object) <- value

\S4method{adjacency_list}{mgnet,data.frame}(object) <- value

\S4method{adjacency_list}{list,list}(object) <- value
}
\arguments{
\item{object}{mgnet.}

\item{value}{\code{\link{data.frame}} with three named columns "from","to","weight"
like in the as_data_frame function of igraph.}
}
\description{
Assign a new netw to \code{object} from an adjacency edge data.frame
}
