% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{comm<-}
\alias{comm<-}
\alias{comm<-,mg,communities-method}
\alias{comm<-,mgnet,communities}
\title{Assign a new comm to \code{object}}
\usage{
comm(object) <- value

\S4method{comm}{mg,communities}(object) <- value
}
\arguments{
\item{object}{(Required) \code{\link{mgnet-class}}.}

\item{value}{(Required) \code{\link{communities}}}
}
\description{
Assign a new comm to \code{object}
}
