% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R, R/class-mgnet.R
\docType{methods}
\name{data<-}
\alias{data<-}
\alias{data<-,mg,matrix-method}
\alias{data<-,mg,matrix}
\alias{data<-,mgnet,matrix-method}
\alias{data<-,mgnet,matrix}
\title{Assign a new ngs matrix to \code{object}}
\usage{
data(object) <- value

\S4method{data}{mg,matrix}(object) <- value

\S4method{data}{mgnet,matrix}(object) <- value
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{value}{(Required) \code{\link{matrix}}}
}
\description{
Assign a new ngs matrix to \code{object}
}
