% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{log_data<-}
\alias{log_data<-}
\alias{log_data<-,mgnet,data.frame-method}
\title{Assign new log_data to \code{object}}
\usage{
log_data(object) <- value

\S4method{log_data}{mgnet,data.frame}(object) <- value
}
\arguments{
\item{object}{mgnet}

\item{value}{matrix}
}
\description{
Assign new log_data to \code{object}
}
