% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{meta_sample<-}
\alias{meta_sample<-}
\alias{meta_sample<-,mgnet,data.frame-method}
\title{Assign new samples metadata to \code{object}}
\usage{
meta_sample(object) <- value

\S4method{meta_sample}{mgnet,data.frame}(object) <- value
}
\arguments{
\item{object}{mgnet}

\item{value}{data.frame}
}
\description{
Assign new samples metadata to \code{object}
}
