% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{meta_taxa<-}
\alias{meta_taxa<-}
\alias{meta_taxa<-,mgnet,data.frame-method}
\title{Assign new meta taxa to \code{object}}
\usage{
meta_taxa(object) <- value

\S4method{meta_taxa}{mgnet,data.frame}(object) <- value
}
\arguments{
\item{object}{mgnet}

\item{value}{data.frame}
}
\description{
Assign new meta taxa to \code{object}
}
