% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-show.R
\docType{methods}
\name{data<-}
\alias{data<-}
\alias{assign-ngs_data}
\alias{data<-,mg,ngs_data-method}
\alias{data<-,mg,ngs_data-method,ngs_data}
\alias{data<-,mg,matrix-method}
\alias{data<-,mg,ngs_data-method,matrix}
\title{Assign a new ngs matrix to \code{object}}
\usage{
data(object) <- value

\S4method{data}{mg,ngs_data}(object) <- value

\S4method{data}{mg,matrix}(object) <- value
}
\arguments{
\item{object}{Required \code{\link{mg-class}}.}

\item{value}{Required \code{\link{ngs_data-class}} or \code{\link{matrix}}}
}
\description{
Assign a new ngs matrix to \code{object}
}
