% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-shows.R
\docType{methods}
\name{meta<-}
\alias{meta<-}
\alias{assign-sample_metadata}
\alias{meta<-,sample_metadata,sample_metadataOrDataFrameOrNULL-method}
\alias{meta<-,sample_metadata,sample_metadata-method}
\alias{meta<-,mg,ngs_dataOrMatrixOrNULL-method}
\alias{meta<-,mg,sample_metadata-method}
\title{Assign new samples metadata to \code{object}}
\usage{
meta(object) <- value

\S4method{meta}{sample_metadata,sample_metadataOrDataFrameOrNULL}(object) <- value

\S4method{meta}{mg,ngs_dataOrMatrixOrNULL}(object) <- value
}
\arguments{
\item{object}{Required \code{\link{mg-class}} or \code{\link{sample_metadata-class}}.}

\item{value}{Required \code{\link{sample_metadata-class}} or \code{\link{data.frame}}}
}
\description{
Assign new samples metadata to \code{object}
}
