% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors.R
\docType{methods}
\name{taxa<-}
\alias{taxa<-}
\alias{assign-taxonomy_table}
\alias{taxa<-,taxonomy_table,taxonomy_tableOrMatrixOrNULL-method}
\alias{taxa<-,taxonomy_table,taxonomy_table-method}
\alias{taxa<-,mg,taxonomy_tableOrMatrixOrNULL-method}
\alias{taxa<-,mg,taxonomy_table-method}
\title{Assign new taxonomy table to \code{object}}
\usage{
taxa(object) <- value

\S4method{taxa}{taxonomy_table,taxonomy_tableOrMatrixOrNULL}(object) <- value

\S4method{taxa}{mg,taxonomy_tableOrMatrixOrNULL}(object) <- value
}
\arguments{
\item{object}{Required \code{\link{mg-class}} or \code{\link{taxonomy_table-class}}.}

\item{value}{Required \code{\link{taxonomy_table-class}} or \code{\link{matrix}}}
}
\description{
Assign new taxonomy table to \code{object}
}
