% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_classes_bind.R
\name{bind_meta}
\alias{bind_meta}
\alias{bind_meta,mgnet-method}
\alias{bind_meta,mgnetList-method}
\title{Bind New Columns to Sample Information in mgnet or mgnetList Objects}
\usage{
bind_meta(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{...}{Data frames, tibbles, or vectors to bind as new columns to the \code{meta} data frame.
For \code{mgnet} objects, the additional columns should match the number of rows in the \code{meta} data frame.
For \code{mgnetList} objects, the additional columns should be provided as named lists,
where each name corresponds to an \code{mgnet} object in the list, and each element is a data frame,
tibble, or vector to bind to that \code{mgnet} object's \code{meta} data frame.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with an updated \code{meta} data frame.
}
\description{
Adds new columns to the \code{meta} data frame within \code{mgnet} objects by binding them
using the \code{bind_cols} function from the \code{dplyr} package. It updates the \code{meta} slot with
the new data frame. This method allows you to augment your sample metadata with additional
variables, enhancing the information contained within your \code{mgnet} object.
}
\details{
Leverages the \code{bind_cols} function from \code{dplyr} for intuitive and flexible column binding,
allowing you to augment the sample metadata with additional data.
}
