% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_classes_bind.R
\name{bind_taxa}
\alias{bind_taxa}
\alias{bind_taxa,mgnet-method}
\alias{bind_taxa,mgnetList-method}
\title{Bind New Columns to Taxa Information in mgnet or mgnetList Objects}
\usage{
bind_taxa(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{...}{Data frames, tibbles, or vectors to bind as new columns to the \code{taxa} data frame.
For \code{mgnet} objects, the additional columns should match the number of rows in the \code{taxa} data frame.
For \code{mgnetList} objects, the additional columns should be provided as named lists,
where each name corresponds to an \code{mgnet} object in the list, and each element is a data frame,
tibble, or vector to bind to that \code{mgnet} object's \code{taxa} data frame.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with an updated \code{taxa} data frame.
}
\description{
Adds new columns to the \code{taxa} data frame within \code{mgnet} objects by binding them
using the \code{bind_cols} function from the \code{dplyr} package. It updates the \code{taxa} slot with
the new data frame. This method allows you to augment your taxa metadata with additional
variables, enhancing the information contained within your \code{mgnet} object.
}
\details{
Leverages the \code{bind_cols} function from \code{dplyr} for intuitive and flexible column binding,
allowing you to augment the taxa metadata with additional data.
}
