% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_extractors.R
\name{$,mgnetList-method}
\alias{$,mgnetList-method}
\title{Access mgnet Objects by Name}
\usage{
\S4method{$}{mgnetList}(x, name)
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{name}{The name of the \code{mgnet} object to access.}
}
\value{
The \code{mgnet} object with the specified name.
}
\description{
This method allows for easy access to individual \code{mgnet} objects within an \code{mgnetList} by their names.
}
