% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_exclusive_keyword_usage}
\alias{check_exclusive_keyword_usage}
\title{Check for Exclusive Keyword Usage in Expressions}
\usage{
check_exclusive_keyword_usage(expressions)
}
\arguments{
\item{expressions}{A list of quosured expressions.}
}
\value{
Throws an error if any expression contains exclusive keywords along with other variables.
}
\description{
This internal function checks each expression to ensure that reserved keywords
('.', 'netw', 'comm') are used exclusively. If any of these keywords are present,
no other variables or keywords should be present in the expression.
}
\keyword{internal}
