% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_reserved_keywords}
\alias{check_reserved_keywords}
\title{Check for Reserved Keywords in Expression Names}
\usage{
check_reserved_keywords(expressions)
}
\arguments{
\item{expressions}{A list of expressions captured by \code{rlang::enquos(...)}.}
}
\value{
Throws an error if any reserved keywords are found in the expression names; otherwise, nothing is returned.
}
\description{
This internal function checks if any reserved keywords are used in the names of expressions.
Reserved keywords cannot be modified and should not be used as names in the expressions.
}
\keyword{internal}
