% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities.R
\name{check_sample_sum}
\alias{check_sample_sum}
\alias{check_sample_sum,mgnet-method}
\alias{check_sample_sum,mgnetList-method}
\title{Check for Presence of Sample Sum in \code{mgnet} or \code{mgnetList} Objects}
\usage{
check_sample_sum(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, returns a logical value (\code{TRUE} or \code{FALSE})
indicating the presence of the \code{sample_sum} column. For an \code{mgnetList} object,
returns a named logical vector with each element indicating the presence
of \code{sample_sum} for the corresponding \code{mgnet} object in the list.
}
\description{
This function checks whether the \code{sample_sum} column is present in the
\code{info_sample} slot of an \code{mgnet} object, or each \code{mgnet} object within
a \code{mgnetList}. It is crucial for analyses that rely on relative abundance
calculations.
}
\details{
The \code{sample_sum} column in the \code{info_sample} slot represents the total
abundance count for each sample, which is essential for calculating relative
abundances. This function allows users to quickly verify the inclusion of
this crucial piece of data in their metagenomic datasets.
}
