% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{CLR}
\alias{CLR}
\alias{CLR,mgnet-method}
\alias{CLR,list-method}
\title{Get CLR transformed abundances.}
\usage{
CLR(object,rank)

\S4method{CLR}{mgnet}(object)

\S4method{CLR}{list}(object)
}
\arguments{
\item{object}{mgnet.}
}
\description{
Retrieves the CLR transformed abundances of data. In mgnet object must be
save the variable geometric_mean in meta_sample slot ( see \code{\link{save_geometric_mean}}).
}
