% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{clr_abundance}
\alias{clr_abundance}
\alias{clr_abundance,mgnet,missing-method}
\alias{clr_abundance,mgnet,character-method}
\alias{clr_abundance,list,missing-method}
\alias{clr_abundance,list,character-method}
\title{Get clr abundances at taxonomy rank choosen.}
\usage{
relative(object,rank)

\S4method{clr_abundance}{mgnet,missing}(object)

\S4method{clr_abundance}{mgnet,character}(object, rank)

\S4method{clr_abundance}{list,missing}(object)

\S4method{clr_abundance}{list,character}(object, rank)
}
\arguments{
\item{object}{mgnet or list.}

\item{rank}{(Optional)}
}
\description{
Retrieves the clr abundances of data the taxonomic rank choosen.
In mgnet object must be save log_data slot.
}
