% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_CLR_zeros.R
\name{clr_zero_handle}
\alias{clr_zero_handle}
\title{CLR Transformation with Zero Replacement}
\usage{
clr_zero_handle(
  X,
  clr_variant = c("clr", "iclr"),
  zero_strategy = c("const", "unif")
)
}
\arguments{
\item{X}{A numeric matrix representing abundance or relative abundance data. The matrix should not be empty.}

\item{clr_variant}{A character string specifying the variant of log-ratio transformation to apply.
Either \code{"clr"} (centered log-ratio) or \code{"iclr"} (isometric centered log-ratio).
Default is \code{"clr"}.}

\item{zero_strategy}{A character string specifying the strategy for replacing zero values in the matrix.
Either \code{"const"} to replace zeros with a constant value or \code{"unif"} to replace zeros with a uniform random value.
Default is \code{"const"}.}
}
\value{
A matrix that has been transformed using the CLR or iCLR method with zero handling applied.
}
\description{
This function performs a centered log-ratio (CLR) or isometric CLR (iCLR) transformation
on an abundance matrix. It includes options for zero replacement strategies, allowing
for the substitution of zero values with either a constant or a uniform random value.
}
\details{
The function first replaces zero values in the matrix according to the chosen \code{zero_strategy}.
It then applies either the CLR or iCLR transformation.
The CLR transformation transforms the data into log-ratio form by dividing each value
by the geometric mean of its corresponding row, followed by taking the log.
The iCLR transformation is a variation of CLR that maintains isometry in the transformed space.
}
\references{
Lubbe S. et al, 2021. Comparison of zero replacement strategies for compositional data with large numbers of zeros. Chemometrics and Intelligent Laboratory Systems. 10.1016/j.chemolab.2021.104248
}
\seealso{
\code{\link[=clr]{clr}} for standard centered log-ratio transformation,
\code{\link[=iclr]{iclr}} for isometric centered log-ratio transformation,
and \code{\link[=zero_dealing]{zero_dealing}} for zero handling strategies.
}
