% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7a_classes_color.R
\name{colorize_taxa}
\alias{colorize_taxa}
\alias{colorize_taxa,mgnet-method}
\alias{colorize_taxa,mgnetList-method}
\title{Generate Colors Based on Factor Levels}
\usage{
colorize_taxa(object, var, color_to = NULL, colorspace = "pretty", alpha = 1)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object containing taxa data with a factor variable.}

\item{var}{The name of the column (as a string) in the taxa metadata representing the factor variable.}

\item{color_to}{The name of the column (as a string) where the generated colors will be stored.}

\item{colorspace}{Defines the color space for color generation. It can be one of the predefined
character options ('pretty', 'pretty_dark', 'rainbow', 'pastels') or a list specifying
the 'h', 's', and 'l' elements for custom color spaces. For a detailed explanation of
these options, see the \code{qualpalr} package documentation or enter \code{??qualpalr} in the R console.}

\item{alpha}{Numeric; the transparency level of the colors, where 1 is opaque and 0 is fully transparent.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the new color column added.
}
\description{
This function generates a color palette based on the levels of a specified factor variable in the taxa metadata. The colors are then stored as a new column.
}
\details{
Simplified method to assign colors based on the levels of a pre-existing factor variable in the taxa metadata of \code{mgnet} or \code{mgnetList} objects.

The function relies on the pre-existing factor levels in the specified column to generate distinct colors.
This makes it easy to apply a consistent color scheme to ordered or lumped categories.
}
