% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis_color.R
\name{colormap_categories}
\alias{colormap_categories}
\title{Create Color Palette for Categorical Data}
\usage{
colormap_categories(
  categories,
  distinct_colors,
  extraColor = "#FFFFFF",
  alpha = 1,
  colorspace = "pretty"
)
}
\arguments{
\item{categories}{A character vector of identifiers for the categories.}

\item{distinct_colors}{An integer specifying the maximum number of distinct colors to generate,
ranging from 1 to 99. Default is the length of categories.}

\item{extraColor}{The color used for identifiers exceeding the \code{distinct_colors} limit,
defaults to white. Accepts hexadecimal color codes.}

\item{alpha}{A numeric value specifying the transparency level of the colors.}

\item{colorspace}{Defines the color space for color generation. It can be one of the predefined
character options ('pretty', 'pretty_dark', 'rainbow', 'pastels') or a list specifying
the 'h', 's', and 'l' elements for custom color spaces. For a detailed explanation of
these options, see the \code{qualpalr} package documentation or enter \code{??qualpalr} in the R console.}
}
\value{
A named vector of hexadecimal color codes, with names corresponding to the category IDs.
}
\description{
Generates a distinct color palette for categorical data classifications. Each unique category
is assigned a color up to a specified limit (\code{distinct_colors}). Beyond this limit, \code{extraColor}
is used for the remaining categories. Duplicate identifiers are filtered out to ensure uniqueness.
}
