% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{comm}
\alias{comm}
\alias{comm,mgnet-method}
\alias{comm,list-method}
\title{Retrieves network communities.}
\usage{
comm(object)

\S4method{comm}{mgnet}(object)

\S4method{comm}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return the communities associated to netw slot.
}
