% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_setters.R
\name{comm<-}
\alias{comm<-}
\alias{comm<-,mgnet-method}
\alias{comm<-,mgnetList-method}
\title{Set Community Detection Results}
\usage{
comm(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new community detection results to be set, which should be compatible
with the structure expected by the \code{mgnet} object's community slot. For individual \code{mgnet}
objects, this is typically an object of class \code{communities} as returned by community detection
functions in the \code{igraph} package. For \code{mgnetList} objects, provide a named list where each element
is a \code{communities} object corresponding to the respective \code{mgnet} object within the list.}
}
\value{
The \code{mgnet} or \code{mgnetList} object with updated community detection results.
}
\description{
This setter function updates the community detection results for \code{mgnet} objects and
each \code{mgnet} object within a \code{mgnetList}. The community data must be an object of class \code{communities}
for \code{mgnet} objects. For \code{mgnetList} objects, the community data should be a named list of \code{communities} objects
corresponding to each \code{mgnet} object within the list.
}
