% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{comm}
\alias{comm}
\alias{comm,mgnet-method}
\alias{comm,mgnets-method}
\title{Access community detection results}
\usage{
comm(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
For \code{mgnet}, the community detection object stored in the \code{comm} slot.

For \code{mgnets}, a named list of community detection objects.
}
\description{
Retrieve the community detection object stored in the \code{comm} slot of an
\code{mgnet} object, or from each element of an \code{mgnets} object.
}
