% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{comm_id}
\alias{comm_id}
\alias{comm_id,mgnet-method}
\alias{comm_id,mgnetList-method}
\title{Retrieve Community IDs from mgnet or mgnetList Objects}
\usage{
comm_id(object, .fmt = "list")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{.fmt}{The format of the output. Possible values are:
\itemize{
\item \code{"list"}: Returns a named list where each element is a character vector of community IDs.
\item \code{"df"}: Returns a data frame where each row corresponds to a taxa ID and each column to a community ID.
\item \code{"tbl"}: Returns a tibble where each row corresponds to a taxa ID and each column to a community ID.
}
Default is \code{"list"}.}
}
\value{
For a single \code{mgnet} object, the function returns the community IDs in the specified format.
For an \code{mgnetList} object, it returns a list, data frame, or tibble, depending on the specified format,
where each entry corresponds to the community IDs for each \code{mgnet} object in the list.
}
\description{
This function retrieves the community IDs from an \code{mgnet} or \code{mgnetList} object. The community IDs are
derived from the \code{community} slot of the \code{mgnet} object or from each \code{mgnet} object within an \code{mgnetList}.
The function supports multiple output formats including list, data frame, and tibble.
}
