% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_classes_getters.R
\name{community}
\alias{community}
\alias{community,mgnet-method}
\alias{community,mgnetList-method}
\title{Get Community Detection Results}
\usage{
community(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
The community detection result object stored in the \code{community} slot for an \code{mgnet} object,
or a list of such objects for an \code{mgnetList} object, each representing the community
detection results of a contained \code{mgnet} object.
}
\description{
Retrieves the community detection results stored in the \code{community} slot of an \code{mgnet} object
or each \code{mgnet} object within an \code{mgnetList}. These results are typically derived from network
analysis methods and indicate the grouping or clustering of taxa into communities based on their
network interactions.
}
