% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{community_members}
\alias{community_members}
\alias{community_members,mgnet-method}
\alias{community_members,mgnetList-method}
\title{Retrieve Community Memberships for Network Vertices}
\usage{
community_members(object, .fmt = "list")
}
\arguments{
\item{object}{An object of class \code{mgnet} or \code{mgnetList}. For \code{mgnet}, the function returns the community membership IDs
of all vertices in the network. For \code{mgnetList}, the function operates on each network in the list and compiles the results.}

\item{.fmt}{A character string indicating the desired output format for the community membership IDs when the input
object is an \code{mgnetList}. Options include "list" for a list of vectors, "df" for a data frame, and "tbl" for a tibble.
The default is "list". This parameter is ignored if the input is an \code{mgnet} object.}
}
\value{
Depending on the input object and the \code{.fmt} parameter:
\itemize{
\item For an \code{mgnet} object, returns a named vector where each name is a vertex ID and each value is the corresponding community membership ID.
\item For an \code{mgnetList} object with \code{.fmt} set to "list", returns a list of named vectors, each corresponding to one network in the list.
\item For an \code{mgnetList} object with \code{.fmt} set to "df", returns a data frame where each row corresponds to a vertex (across all networks) and each column corresponds to one network.
\item For an \code{mgnetList} object with \code{.fmt} set to "tbl", returns a tibble similar to the data frame format, but with \code{taxa_id} as the first column.
}
}
\description{
This function fetches the community membership IDs for each vertex in a network or a list of networks.
It supports \code{mgnet} objects, representing single networks, and \code{mgnetList} objects, representing collections of networks.
The function can return the results in various formats, including lists, data frames, or tibbles, depending on user preference.
}
\details{
The function leverages the \code{membership} function from the \code{igraph} package to determine the community membership of vertices.
For \code{mgnetList} objects, the function applies the operation to each network in the list and organizes the results according to the specified format.
When dealing with \code{mgnetList} objects and choosing "df" or "tbl" for \code{.fmt}, the function uniquely identifies each vertex across all networks to compile the community memberships coherently.
}
