% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{data}
\alias{data}
\alias{data,mgnet-method}
\alias{data,list-method}
\title{Retrieves ngs data.}
\usage{
data(object)

\S4method{data}{mgnet}(object)

\S4method{data}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Return the numeric matrix associated with the ngs data from mgnet class.
}
