% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{default_decoration}
\alias{default_decoration}
\alias{default_decoration,mgnet-method}
\alias{default_decoration,mgnet}
\alias{default_decoration,list-method}
\alias{default_decoration-methods,list,numeric,logical}
\title{Add graphical decorations}
\usage{
default_decoration(obj)

\S4method{default_decoration}{mgnet}(obj)

\S4method{default_decoration}{list}(obj)
}
\arguments{
\item{obj}{\code{\link{mgnet-class}}}
}
\description{
Add graphical descriptions to network:
\itemize{
\item Vertex size proportional to mean clr abundances over samples.
\item Link color could be red (+) or blue (-) respect the weights sign.
}
}
