% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{depth}
\alias{depth}
\alias{depth,mgnet-method}
\alias{depth,list-method}
\title{Get samples depth.}
\usage{
depth(object)

\S4method{depth}{mgnet}(object)

\S4method{depth}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Retrieves the depth of each sample using the Depth variable in in meta slot.
Depth is automatically generated if not present during the construction of
the mg object.
}
