% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{deselect_link}
\alias{deselect_link}
\alias{deselect_link,mgnet-method}
\alias{deselect_link,mgnets-method}
\title{Clear link selection}
\usage{
deselect_link(object)

\S4method{deselect_link}{mgnet}(object)

\S4method{deselect_link}{mgnets}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
The same object, with no active link selection.
}
\description{
Remove the \code{selected_links} attribute from an \code{mgnet} or \code{mgnets} object.
}
