% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities_link.R
\name{deselect_links}
\alias{deselect_links}
\title{Deselect Links in an mgnet Object}
\usage{
deselect_links(object)
}
\arguments{
\item{object}{An \code{mgnet} object.}
}
\value{
The \code{mgnet} object without the \code{selected_links} attribute, reflecting no current link selection.
}
\description{
Clears the \code{selected_links} attribute from the \code{mgnet} object, effectively removing the selection of links.
This method resets the state of link selection within the object, allowing for new operations without the constraints of the previous link selection.
}
