% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_abun_zero_structure}
\alias{.assert_abun_zero_structure}
\title{Assert: identical zero position in the two matrices}
\usage{
.assert_abun_zero_structure(x1, field1, x2, field2)
}
\description{
Assert: identical zero position in the two matrices
}
\keyword{internal}
