% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_disjoint_feature_names}
\alias{.assert_disjoint_feature_names}
\title{Assert: disjoint feature names between meta and taxa}
\usage{
.assert_disjoint_feature_names(meta, taxa)
}
\description{
Assert: disjoint feature names between meta and taxa
}
\keyword{internal}
