% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_identical_samples}
\alias{.assert_identical_samples}
\title{Assert: reciprocal number and names of rows}
\usage{
.assert_identical_samples(x1, field1, x2, field2)
}
\description{
Assert: reciprocal number and names of rows
}
\keyword{internal}
