% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_identical_taxa}
\alias{.assert_identical_taxa}
\title{Assert: reciprocal number and names of cols of x1 and rows of x2}
\usage{
.assert_identical_taxa(x1, field1, x2, field2)
}
\description{
Assert: reciprocal number and names of cols of x1 and rows of x2
}
\keyword{internal}
