% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_identical_taxa_nodes}
\alias{.assert_identical_taxa_nodes}
\title{Assert: reciprocal number and names of rows of x1 and nodes in netw}
\usage{
.assert_identical_taxa_nodes(x1, field1, netw)
}
\description{
Assert: reciprocal number and names of rows of x1 and nodes in netw
}
\keyword{internal}
