% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_named_and_uniqueness}
\alias{.assert_named_and_uniqueness}
\title{Assert: unique row and column names}
\usage{
.assert_named_and_uniqueness(x, field)
}
\description{
Assert: unique row and column names
}
\keyword{internal}
