% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_netw_comm_consistency}
\alias{.assert_netw_comm_consistency}
\title{Assert: consistency between network and community}
\usage{
.assert_netw_comm_consistency(netw, comm)
}
\description{
Assert: consistency between network and community
}
\keyword{internal}
