% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_no_reserved_names}
\alias{.assert_no_reserved_names}
\title{Assert: no reserved keywords in column names}
\usage{
.assert_no_reserved_names(x, field)
}
\description{
Assert: no reserved keywords in column names
}
\keyword{internal}
