% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.assert_reciprocal_dimensions}
\alias{.assert_reciprocal_dimensions}
\title{Assert: reciprocal names of rows and columns must be identical}
\usage{
.assert_reciprocal_dimensions(x1, field1, x2, field2)
}
\description{
Assert: reciprocal names of rows and columns must be identical
}
\keyword{internal}
