% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_mgnet_validity.R
\name{.assertCharacterMatrix}
\alias{.assertCharacterMatrix}
\title{Internal: Assert Character Matrix}
\usage{
.assertCharacterMatrix(obj, errors)
}
\arguments{
\item{obj}{The object to check.}

\item{errors}{A character vector for accumulating error messages from validation checks.}
}
\value{
An updated character vector containing any new error messages if \code{obj}
is not a character matrix; otherwise, the original \code{errors} vector is returned.
}
\description{
Validates that a given object is a character matrix. This function ensures that
slots intended to hold character data, such as taxonomic classifications, adhere
to the expected matrix format with character data types.
}
\keyword{internal}
