% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_mgnet_validity.R
\name{.assertDataFrame}
\alias{.assertDataFrame}
\title{Internal: Assert Data Frame}
\usage{
.assertDataFrame(obj, errors)
}
\arguments{
\item{obj}{The object to check.}

\item{errors}{A character vector for accumulating error messages from validation checks.}
}
\value{
An updated character vector containing any new error messages if \code{obj}
is not a data frame; otherwise, the original \code{errors} vector is returned.
}
\description{
Validates that the given object is a data frame. This function is essential for
ensuring that slots expected to contain metadata or other tabular data are correctly
formatted as data frames.
}
\keyword{internal}
