% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_mgnet_validity.R
\name{.assertMatchingNames}
\alias{.assertMatchingNames}
\title{Internal: Assert Matching Row and Column Names Between Two Slots}
\usage{
.assertMatchingNames(obj, slotName1, slotName2, errors)
}
\arguments{
\item{obj}{The S4 object containing the slots to be compared.}

\item{slotName1}{The name of the first slot to compare.}

\item{slotName2}{The name of the second slot to compare.}

\item{errors}{A pre-existing character vector where any new error messages identified
during the validation process will be appended. This allows for the accumulation
of error messages across multiple validation checks.}
}
\value{
An updated character vector of error messages. If the number of columns or
the column names between the two slots do not match, appropriate error messages
are added to the vector. If there are no discrepancies, the original vector is
returned unchanged.
}
\description{
Validates that two specified slots within an S4 object have the same number of
rows and columns and matching row and column names.
This function is crucial for ensuring consistency and
alignment between related datasets within the \code{mgnet} class, such as ensuring that
matrices representing different types of data (e.g., raw and processed data) are
compatible in terms of their structure and labeling.
}
\keyword{internal}
