% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_mgnet_validity.R
\name{.assertMatchingNamesVertices}
\alias{.assertMatchingNamesVertices}
\title{Internal: Assert Matching Dimension and Network Vertices Names}
\usage{
.assertMatchingNamesVertices(obj, slotName, which, errors)
}
\arguments{
\item{obj}{The S4 object containing the slots to be compared.}

\item{which}{rows or columns of the slot to compare with the network.}

\item{errors}{A pre-existing character vector where any new error messages identified
during the validation process will be appended. This allows for the accumulation
of error messages across multiple validation checks.}

\item{slot}{The name of the slot to compare.}
}
\value{
An updated character vector of error messages.
}
\description{
Validates that the specified slots within an S4 object match the names of the
network slot.
This function is crucial for ensuring consistency and alignment between
related datasets within the \code{mgnet} class, such as ensuring that
matrices representing different types of data (e.g., raw and processed data) are
compatible in terms of their structure and labeling.
}
\keyword{internal}
