% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_mgnet_validity.R
\name{.assertMatchingZeroPositions}
\alias{.assertMatchingZeroPositions}
\title{Internal: Assert Matching Zero Positions Between Two Slots}
\usage{
.assertMatchingZeroPositions(obj, slotName1, slotName2, errors)
}
\arguments{
\item{obj}{The S4 object containing the slots to be compared.}

\item{slotName1}{The name of the first slot containing a matrix to check for zero positions.}

\item{slotName2}{The name of the second slot containing a matrix to check for zero positions.}

\item{errors}{A pre-existing character vector where any new error messages identified
during the validation process will be appended. This allows for the accumulation
of error messages across multiple validation checks.}
}
\value{
An updated character vector of error messages. If zero positions do not match,
an appropriate error message is added to the vector. If there are no discrepancies,
the original vector is returned unchanged.
}
\description{
Validates that the zero positions in two specified slots within an S4 object are identical.
This function is crucial for ensuring data integrity, particularly in biological or ecological
datasets where the absence (zero) of a measurement in one type of data must correspond to the absence
in another, ensuring aligned analytical outputs.
}
\keyword{internal}
