% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_mgnet_validity.R
\name{.assertNumericMatrix}
\alias{.assertNumericMatrix}
\title{Internal: Assert Numeric Matrix}
\usage{
.assertNumericMatrix(obj, errors)
}
\arguments{
\item{obj}{The object to check.}

\item{errors}{A character vector accumulating error messages from validation checks.}
}
\value{
An updated character vector containing any new error messages if \code{obj}
is not a numeric matrix; otherwise, the original \code{errors} vector is returned.
}
\description{
Validates that the given object is a numeric matrix. This function is used
internally to ensure data slots intended to be numeric matrices, such as abundance
matrices, meet the required data type specifications.
}
\keyword{internal}
