% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_mgnet_validity.R
\name{.assertSampleSumValidity}
\alias{.assertSampleSumValidity}
\title{Internal: Validate Sample Sum in info_sample}
\usage{
.assertSampleSumValidity(object, errors)
}
\arguments{
\item{object}{The \code{mgnet} object being validated.}

\item{errors}{A character vector accumulating any error messages identified during the validation process.
This vector is updated in-place with messages pertaining to the \code{sample_sum} validations.}
}
\value{
A possibly updated character vector containing accumulated validation error messages. If new errors
related to the \code{sample_sum} column are identified, they are appended to the vector.
}
\description{
This internal function performs validation checks on the \code{sample_sum} column within the \code{info_sample} slot
of an \code{mgnet} object. It ensures that if \code{sample_sum} is provided, it adheres to several conditions:
\itemize{
\item The \code{abundance} slot must be present and contain data.
\item The length of \code{sample_sum} must match the number of rows in \code{abundance}, representing the total number of samples.
\item All values in \code{sample_sum} must be positive, reflecting the non-negativity constraint of abundance measures.
\item Each value in \code{sample_sum} must be greater than or equal to the corresponding row sum of the \code{abundance} matrix, ensuring accuracy in representing total sample abundance.
}
}
\details{
This function is part of the validation process for \code{mgnet} objects, called within the class's
set validity method. It focuses specifically on the integrity and logical consistency of the
\code{sample_sum} column when provided. As an internal function, it is not intended for direct use
outside of the \code{mgnet} object validation context.
}
\keyword{internal}
