% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-link.R
\name{.ensure_link_id}
\alias{.ensure_link_id}
\title{Ensure that a graph has a \code{link_id} edge attribute}
\usage{
.ensure_link_id(g)
}
\arguments{
\item{g}{An \code{igraph} object.}
}
\value{
The input graph, with a guaranteed \code{link_id} edge attribute whenever edges
are present.
}
\description{
Check whether an \code{igraph} object already has an edge attribute named
\code{link_id}. If not, generate one with \code{\link[=.make_link_id]{.make_link_id()}} and attach it to the
graph.
}
\keyword{internal}
