% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.validate_link_groups}
\alias{.validate_link_groups}
\title{Validate link_groups attribute (edge-level grouping)}
\usage{
.validate_link_groups(object)
}
\description{
Validate link_groups attribute (edge-level grouping)
}
\keyword{internal}
