% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.validate_mgnet_groups}
\alias{.validate_mgnet_groups}
\title{Validate mgnet_groups attribute (grouping on meta/taxa)}
\usage{
.validate_mgnet_groups(object)
}
\description{
Validate mgnet_groups attribute (grouping on meta/taxa)
}
\keyword{internal}
