% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet-validity.R
\name{.validate_selected_links}
\alias{.validate_selected_links}
\title{Validate selected_links attribute (edge selection)}
\usage{
.validate_selected_links(object)
}
\description{
Validate selected_links attribute (edge selection)
}
\keyword{internal}
