% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_classes_utilities.R
\name{ensure_mgnet}
\alias{ensure_mgnet}
\title{Ensure Object is an \code{mgnet} Instance}
\usage{
ensure_mgnet(object)
}
\arguments{
\item{object}{The object to check.}
}
\value{
Invisible \code{TRUE} if the object is an instance of the \code{mgnet} class.
}
\description{
Checks if the provided object is an instance of the \code{mgnet} class
and stops execution with an error message if not.
}
