% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_classes_utilities.R
\name{ensure_mgnetList}
\alias{ensure_mgnetList}
\title{Ensure Object is an \code{mgnetList} Instance}
\usage{
ensure_mgnetList(object)
}
\arguments{
\item{object}{The object to check.}
}
\value{
Invisible \code{TRUE} if the object is an instance of the \code{mgnetList} class.
}
\description{
Checks if the provided object is an instance of the \code{mgnetList} class
and stops execution with an error message if not.
}
