% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-show.R
\name{[,ngs_data-method}
\alias{[,ngs_data-method}
\alias{[,sample_metadata-method}
\title{Method extensions to extraction operator.}
\usage{
\S4method{[}{ngs_data}(x, i, j, ..., drop = TRUE)

\S4method{[}{sample_metadata}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{j}{See \code{\link[base]{Extract}}}

\item{...}{See \code{\link[base]{Extract}}}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }
}
\description{
See the documentation for the \code{\link[base]{Extract}} generic,
defined in the R \code{\link[base]{base-package}} for the expected behavior.
}
\seealso{
\code{\link[base]{Extract}}
}
