% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{extract_required_keys}
\alias{extract_required_keys}
\title{Extract Required Keys from Expressions}
\usage{
extract_required_keys(expressions)
}
\arguments{
\item{expressions}{A list of expressions captured using \code{rlang::enquos}. These
expressions contain variable names that need to be extracted and evaluated.}
}
\value{
A character vector of unique variable names (keys) required by the expressions.
}
\description{
This internal function extracts the variable names from a list of expressions.
It identifies all the variables used within the expressions provided and returns
a unique set of required keys. This function is helpful in determining which
abundance or metadata variables are needed for processing.
}
\seealso{
\code{\link[=rlang]{enquos}} for capturing expressions in quosures.
}
\keyword{internal}
