% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_classes_filter_info.R
\name{filter_info_sample}
\alias{filter_info_sample}
\alias{filter_info_sample,mgnet-method}
\alias{filter_info_sample,mgnetList-method}
\title{Filter Samples in mgnet Objects Based on info_sample Conditions}
\usage{
filter_info_sample(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{...}{Conditions to filter samples, which are passed unquoted and can
include dplyr-style filtering expressions. These conditions are applied to the
columns of the \code{info_sample} data frame, which is accessed as a tibble with
\code{sample_id} serving as the key column.}
}
\value{
An \code{mgnet} or \code{mgnetList} object with samples filtered according to the
specified conditions. The filtered object(s) retain only those samples that meet
the filtering criteria specified by the \code{...} arguments.
}
\description{
This function filters samples in an \code{mgnet} object or each \code{mgnet} object within
an \code{mgnetList} based on specified conditions applied to the \code{info_sample} data frame.
The conditions for filtering are based on expressions applied to columns within
the \code{info_sample} data frame.
}
\details{
The \code{filter_info_sample} function leverages the \code{info_sample} getter function
to access the metadata of samples as a tibble, where the row names are transformed
into a \code{sample_id} column. This enables seamless integration with dplyr's \code{filter}
function for specifying complex filtering conditions. The function then uses the
\code{sample_id} column to identify which samples should be retained in the filtered \code{mgnet}
or \code{mgnetList} object.
}
\seealso{
\code{\link[dplyr]{filter}} for details on filter conditions.
}
