% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mutate-filter.R
\name{filter_mgnet}
\alias{filter_mgnet}
\alias{filter_meta}
\alias{filter_taxa}
\alias{filter_meta,mgnet-method}
\alias{filter_meta,mgnets-method}
\alias{filter_taxa,mgnet-method}
\alias{filter_taxa,mgnets-method}
\title{Filter \code{meta} or \code{taxa} in \code{mgnet} / \code{mgnets}}
\usage{
filter_meta(object, ..., .ungroup = FALSE)

\S4method{filter_meta}{mgnet}(object, ..., .ungroup = FALSE)

\S4method{filter_meta}{mgnets}(object, ..., .ungroup = FALSE)

filter_taxa(object, ..., .ungroup = FALSE)

\S4method{filter_taxa}{mgnet}(object, ..., .ungroup = FALSE)

\S4method{filter_taxa}{mgnets}(object, ..., .ungroup = FALSE)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{...}{One or more tidy-evaluated filtering expressions.}

\item{.ungroup}{Logical, default \code{FALSE}. If \code{TRUE}, remove the mgnet grouping
from the object at the end.}
}
\value{
A filtered \code{mgnet} or \code{mgnets} object.
}
\description{
Filters samples (\code{filter_meta()}) or taxa (\code{filter_taxa()}) in an \code{mgnet} object
using one or more tidy-evaluated logical expressions (like \code{dplyr::filter()}).
For \code{mgnets}, filtering is applied independently to each contained \code{mgnet}.
}
\details{
Expressions may reference columns in the target table (\code{meta} or \code{taxa}).
If abundance fields are referenced (\code{abun}, \code{rela}, \code{norm}), the relevant
long-format abundance table is automatically joined to provide the necessary
context.

Multiple expressions are combined using logical AND.

See \code{mutate_mgnet} for grouping and cross-slot semantics.
}
