% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{filter_sample}
\alias{filter_sample}
\alias{filter_sample,mgnet,list-method}
\alias{filter_sample,list,list-method}
\title{Filter samples based on across-sample abundance criteria.}
\usage{
filter_sample(object, flist)

\S4method{filter_sample}{mgnet,list}(object, flist)

\S4method{filter_sample}{list,list}(object, flist)
}
\arguments{
\item{object}{mgnet.}

\item{flist}{list. Each element of flist it must be a function.}
}
\description{
It applies an arbitrary set of functions list on each samples taking into account
their abundances. The function takes as input a mgnet object, and returns its
trimmed mgnet version.
}
