% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4c_classes_filter_info.R
\name{filter_sample_info}
\alias{filter_sample_info}
\alias{filter_taxa_info,mgnetList-method}
\title{Filter Sample Metadata in mgnet or mgnetList Objects}
\usage{
filter_sample_info(object, ..., .by)
}
\arguments{
\item{object}{An object of class \code{mgnetList} containing the data to be filtered.}

\item{...}{Expressions that define the filtering criteria using variables available
in the samples metadata of the object.}

\item{.by}{A character vector specifying the grouping variables for filtering, which defaults to
\code{"sample_id"} for \code{mgnet} objects and \code{"mgnet", "sample_id"} for \code{mgnetList} objects.
This parameter should not include 'taxa_id' as it focuses solely on sample metadata.}
}
\value{
Returns a modified \code{mgnetList} object containing only the samples that meet the specified
filtering criteria. The method operates on samples metadata within each \code{mgnet} object in the list.
}
\description{
This function applies specified filtering criteria exclusively to the metadata of samples within
an \code{mgnetList} object. It allows for advanced querying based on metadata fields and supports
tidyverse-style syntax for manipulating these data.
}
\details{
The function uses several reserved keywords:
\itemize{
\item \code{mgnet}: Refers to each individual \code{mgnet} object within an \code{mgnetList}, used for handling
data across multiple datasets.
\item \code{sample_id}: Unique identifier for each taxon.
\item \code{comm_id}: Identifies community membership; used if community data is present.
Additional metadata columns within the taxa structure can also be utilized for detailed filtering.
This function facilitates complex filtering operations tailored to specific study requirements using
these reserved keywords.
}
}
