% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{filter_taxa}
\alias{filter_taxa}
\alias{filter_taxa,mgnet,list-method}
\alias{filter_taxa,list,list-method}
\title{Filter taxa based on across-sample abundance criteria.}
\usage{
filter_taxa(object, flist)

\S4method{filter_taxa}{mgnet,list}(object, flist)

\S4method{filter_taxa}{list,list}(object, flist)
}
\arguments{
\item{object}{mgnet-class.}

\item{flist}{list. Each element of flist it must be a function.}
}
\description{
It applies an arbitrary set of functions list as across-sample criteria,
one taxa at a time. The function takes as input a mgnet object, and returns its
trimmed mgnet version.
}
