% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4c_classes_filter_data.R
\name{filter_taxa_data}
\alias{filter_taxa_data}
\alias{filter_taxa_data,mgnet-method}
\alias{filter_taxa_data,mgnetList-method}
\title{Filter Taxa Data in mgnet or mgnetList Objects}
\usage{
filter_taxa_data(object, ...)
}
\arguments{
\item{object}{An object of class \code{mgnet} or \code{mgnetList} containing the
data to be filtered.}

\item{...}{Expressions that define the filtering criteria using variables available
in the taxa metadata and abundances of the object. These expressions can
include tidyverse functions and operators, allowing for sophisticated
data manipulation directly within the function call.}
}
\value{
Returns a modified \code{mgnet} or \code{mgnetList} object containing only
the taxa that meet the specified filtering criteria. The method operates on each sample
within an \code{mgnet} object individually. If the input is an \code{mgnetList}, the filtering
operation is applied independently to each \code{mgnet} object within the list, ensuring
that results are consistent and isolated across different datasets or experimental conditions.

Returns a modified \code{mgnet} or \code{mgnetList} object containing only
the taxa that meet the specified filtering criteria.
}
\description{
This function applies specified filtering criteria to taxa data within an \code{mgnet}
or \code{mgnetList} object. It intelligently uses reserved keywords to manage and manipulate
taxa data based on abundance metrics and taxa metadata. The function is designed
to integrate seamlessly with the tidyverse, supporting dplyr-style syntax and operations,
which facilitates complex data manipulation and filtering within a familiar framework.
}
\details{
The function utilizes reserved keywords integral to the \code{mgnet} and \code{mgnetList} structures:
\itemize{
\item \code{abun}: raw abundance data.
\item \code{rela}: relative abundance data.
\item \code{norm}: normalized abundance data.
\item \code{taxa_id}: A unique identifier for each taxon.
\item \code{comm_id}: Community membership identifiers, used if the \code{comm} slot is present.
Additionally, all other metadata columns within the taxa data can be used for filtering.
Internally, the method constructs a long-format data frame where each row represents a
combination of sample and taxa data. This format facilitates complex filtering operations
across multiple dimensions of the data.
}
}
