% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2g_classes_gather.R
\name{gather_meta}
\alias{gather_meta}
\alias{gather_meta,mgnet-method}
\alias{gather_meta,mgnetList-method}
\title{Summarize Sample Metadata in \code{mgnet} and \code{mgnetList} Objects}
\usage{
gather_meta(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.
For \code{mgnet}, this method consolidates all sample metadata into a long-format tibble.
For \code{mgnetList}, it applies the transformation to each \code{mgnet} object within the list and combines the results
into a single tibble, incorporating an additional column named \code{mgnet} that identifies the source
\code{mgnet} object.}
}
\value{
A tibble in a long format that consolidates all the relevant sample metadata under sample identifiers.
For \code{mgnetList} objects, the tibble includes an extra column \code{mgnet} that indicates the source \code{mgnet} object.
}
\description{
Converts \code{mgnet} or \code{mgnetList} objects into a longer format, focusing exclusively on
sample metadata. This transformation facilitates the use of tidyverse tools for
subsequent data manipulation and analysis, especially when only sample-specific details are required.
}
\details{
The \code{gather_meta} method is tailored to prepare \code{mgnet} data for easy integration with tidyverse
workflows by:
\itemize{
\item Transforming sample metadata into a long format based on \code{sample_id}, making it more accessible for
various tidyverse functions such as \code{mutate}, \code{filter}, and \code{summarize}.
\item For \code{mgnetList} objects: The transformation is applied to each individual \code{mgnet} in the list.
The resulting longer format data for each \code{mgnet} is then combined into a single tibble, with
an additional column using the reserved keyword \code{mgnet} to track the original \code{mgnet} object each row of data pertains to.
}
}
