% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2g_classes_gather.R
\name{gather_mgnet}
\alias{gather_mgnet}
\alias{gather_mgnet,mgnet-method}
\alias{gather_mgnet,mgnetList-method}
\title{Summarize \code{mgnet} and \code{mgnetList} Objects in Long Format}
\usage{
gather_mgnet(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.
For \code{mgnet}, this method consolidates sample and taxa information along with
abundance data and associated metadata into a long-format tibble. For \code{mgnetList},
it applies the transformation to each \code{mgnet} object within the list and combines the results
into a single tibble, incorporating an additional column named \code{mgnet} that identifies the source
\code{mgnet} object.}
}
\value{
A tibble in a long format that consolidates all the relevant data fields under sample and taxon identifiers.
For \code{mgnetList} objects, the tibble includes an extra column \code{mgnet} that indicates the source \code{mgnet} object.
}
\description{
Converts \code{mgnet} or \code{mgnetList} objects into a longer format, gathering all available
information including abundances (absolute, relative, and normalized) and metadata into a
single structured tibble. This transformation facilitates the use of tidyverse tools for
subsequent data manipulation and analysis.
}
\details{
The \code{gather_mgnet} method is tailored to prepare \code{mgnet} data for easy integration with tidyverse
workflows by:
\itemize{
\item Transforming data into a long format based on \code{sample_id} and \code{taxa_id}, making it more accessible for
various tidyverse functions such as \code{mutate}, \code{filter}, and \code{summarize}.
\item Aggregating both abundance metrics (e.g., \code{abun}, \code{rela}, \code{norm}) and metadata associated (e.g meta, taxa, comm_id)
with samples and taxa, which allows comprehensive data analysis within a single data frame structure.
\item In the context of \code{mgnetList} objects, besides the standard data transformation applied to each \code{mgnet},
an additional column using the reserved keyword \code{mgnet} is included to keep track of the original \code{mgnet}
object each row of data pertains to, facilitating analysis across multiple datasets.
}
}
