% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{geometric_mean}
\alias{geometric_mean}
\alias{geometric_mean,mgnet-method}
\alias{geometric_mean,list-method}
\title{Get samples geometric mean.}
\usage{
geometric(object)

\S4method{geometric_mean}{mgnet}(object)

\S4method{geometric_mean}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Retrieves the geometric mean of each sample using the geometric_mean variable in in
meta_sample slot if it's present (see \code{\link{save_geometric_mean}}).
}
