% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-verbs-check-expressions.R
\name{get_abundance_keys}
\alias{get_abundance_keys}
\title{Extract Abundance-Related Keys from Expressions}
\usage{
get_abundance_keys(expressions)
}
\arguments{
\item{expressions}{A list of expressions, typically created using \code{rlang::enquos()} in a user-facing function.
These expressions are meant to manipulate or assess the abundance data stored in an \code{mgnet} object.}
}
\value{
A character vector containing unique abundance-related keys found within the provided expressions.
These keys are essential for further processing or filtering operations in metagenomic data analysis.
}
\description{
This function analyzes expressions to identify any references to specific abundance-related keys within an \code{mgnet} object.
It checks for mentions of "abun", "rela", or "norm", which represent different types of abundance data.
}
\details{
The function parses each expression to extract variables, which are then checked against known abundance-related
keys. It's primarily used internally within functions that need to process or manipulate abundance data based on
user-defined criteria.
}
\keyword{internal}
