% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{get_grouped_link}
\alias{get_grouped_link}
\alias{get_grouped_link,mgnet-method}
\alias{get_grouped_link,mgnets-method}
\title{Retrieve link grouping}
\usage{
get_grouped_link(object)

\S4method{get_grouped_link}{mgnet}(object)

\S4method{get_grouped_link}{mgnets}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
For an \code{mgnet}, an integer vector giving the group ID of each edge, or \code{NULL}
if no grouping is currently set.

For an \code{mgnets}, a named list of integer vectors, one per contained \code{mgnet}.
}
\description{
Return the values currently stored in the \code{link_groups} attribute.
}
\seealso{
\code{\link[=group_link]{group_link()}} for assigning link groups and \code{\link[=ungroup_link]{ungroup_link()}} for removing
them.
}
