% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP2}
\alias{HMP2}
\title{Example \code{mgnet} object derived from the Human Microbiome Project}
\format{
An object of class \code{mgnet} with \code{1122} samples and \code{1953} taxa.
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(HMP2)
}
\description{
This dataset contains OTU counts, sample metadata, and taxonomic information
from the HMP2 gut microbiome dataset, stored as an \code{mgnet} object.
}
\keyword{datasets}
