% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_CLR_transformations.R
\name{iclr}
\alias{iclr}
\title{Interquartile Centered Log-Ratio Transformation}
\usage{
iclr(X)
}
\arguments{
\item{X}{A numeric matrix with all elements greater than 0, where rows are samples
and columns are components or taxa.}
}
\value{
A numeric matrix of the same dimension as \code{X}, representing the iclr-transformed data.
}
\description{
Performs the Interquartile Centered Log-Ratio (iclr) transformation on a given dataset.
This transformation is designed to reduce the influence of highly variable components
by normalizing the data using a subset of components with lower variability.
The transformation is applied to compositional data where rows represent samples
and columns represent components or taxa.
}
