% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_classes_base_methods.R
\name{info_sample_vars}
\alias{info_sample_vars}
\alias{info_sample_vars,mgnet-method}
\alias{info_sample_vars,mgnetList-method}
\title{Get Sample Metadata Variables}
\usage{
info_sample_vars(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, a character vector of metadata variable names.
For an \code{mgnetList} object, a named list of character vectors, with each list item representing
the metadata variable names in the corresponding \code{mgnet} objects.
}
\description{
Retrieves the names of metadata variables available in the \code{info_sample} slot of an \code{mgnet} object,
or for each \code{mgnet} object within an \code{mgnetList}.
}
