% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_setters.R
\name{info_taxa<-}
\alias{info_taxa<-}
\alias{info_taxa<-,mgnet-method}
\alias{info_taxa<-,mgnetList-method}
\title{Set Taxa Metadata}
\usage{
info_taxa(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new taxa metadata to be set.}
}
\value{
The \code{mgnet} or \code{mgnetList} object with updated taxa metadata.
}
\description{
This setter function updates the taxa metadata for \code{mgnet} objects and
each \code{mgnet} object within a \code{mgnetList}. The taxa metadata must be provided as a dataframe
for \code{mgnet} objects. For \code{mgnetList} objects, the taxa metadata should be a named list of dataframes
corresponding to each \code{mgnet} object within the list.
}
