% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape.R
\name{initialize_meta}
\alias{initialize_meta}
\title{Initialize Sample Information}
\usage{
initialize_meta(object)
}
\arguments{
\item{object}{An \code{mgnet} object.}
}
\value{
A tibble containing either the sample metadata or just the \code{sample_id} column.
}
\description{
This internal function initializes the sample information for an \code{mgnet} object.
If metadata is present in the object, it returns the metadata in tibble format.
If no metadata is available, it creates a tibble using the \code{sample_id} of the object.
}
\keyword{internal}
