% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape.R
\name{initialize_taxa}
\alias{initialize_taxa}
\title{Initialize Taxa Information}
\usage{
initialize_taxa(object)
}
\arguments{
\item{object}{An \code{mgnet} object.}
}
\value{
A tibble containing either the sample metadata or just the \code{taxa_id} column.
}
\description{
This internal function initializes the taxa information for an \code{mgnet} object.
If metadata is present in the object, it returns the metadata in tibble format.
If no metadata is available, it creates a tibble using the \code{taxa_id} of the object.
}
\keyword{internal}
