% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities.R
\name{is_assign_tbl}
\alias{is_assign_tbl}
\title{Check Tibble in Assign Methods for mgnetList}
\usage{
is_assign_tbl(object, value, sample_or_taxa)
}
\arguments{
\item{object}{An \code{mgnetList} object.}

\item{value}{A tibble or data frame that will be split into different tables, one for each taxa metadata.
It is necessary that the columns \code{mgnet} and \code{sample_id}/\code{taxa_id} exist in the value.}

\item{sample_or_taxa}{Character that indicates if the assign tbl is for the sample or taxa metadata. Possible choices are "sample" or "taxa".}
}
\value{
Logical; \code{TRUE} if all checks pass, otherwise an error is thrown.
}
\description{
Internal function to check if the tibble provided as a value in assign methods
is coherent with the \code{mgnetList} object. The value will replace all the taxa metadata in the \code{mgnetList}.
This function ensures that the provided value meets necessary conditions before assignment.
}
\keyword{internal}
