% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_group.R
\name{is_grouped_taxa}
\alias{is_grouped_taxa}
\alias{is_grouped_taxa,mgnet-method}
\alias{is_grouped_taxa,mgnetList-method}
\title{Check if an mgnet or mgnetList Object is Grouped by Taxa}
\usage{
is_grouped_taxa(object)

\S4method{is_grouped_taxa}{mgnet}(object)

\S4method{is_grouped_taxa}{mgnetList}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
Logical value indicating if the object is grouped.
}
\description{
Returns \code{TRUE} if the \code{mgnet} or \code{mgnetList} object has non-empty taxa grouping,
and \code{FALSE} otherwise.
}
