% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{is_link_grouped}
\alias{is_link_grouped}
\alias{is_link_grouped,mgnet-method}
\alias{is_link_grouped,mgnets-method}
\title{Check whether links are grouped}
\usage{
is_link_grouped(object)

\S4method{is_link_grouped}{mgnet}(object)

\S4method{is_link_grouped}{mgnets}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
For an \code{mgnet}, \code{TRUE} if the \code{link_groups} attribute is set, \code{FALSE}
otherwise.

For an \code{mgnets}, \code{TRUE} only if \strong{all} contained \code{mgnet} objects currently
carry link grouping information. This implements an all-or-none
collection-level semantics.
}
\description{
Determine whether an \code{mgnet} or \code{mgnets} object currently carries link
grouping information.
}
