% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_classes_utilities.R
\name{is_mgnet}
\alias{is_mgnet}
\title{Check if an object is an \code{mgnet} Instance}
\usage{
is_mgnet(object)
}
\arguments{
\item{object}{The object to check.}
}
\value{
\code{TRUE} if the object is an instance of the \code{mgnet} class, \code{FALSE} otherwise.
}
\description{
This function checks whether a given object is an instance of the \code{mgnet} class.
}
