% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\name{is_mgnet_list}
\alias{is_mgnet_list}
\title{Check mgnet list}
\usage{
is_mgnet_list(L)
}
\arguments{
\item{object}{list}
}
\description{
Internal function to check if a list is compose from only
mgnet objects.
}
\keyword{internal}
