% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities.R
\name{is_mgnetList}
\alias{is_mgnetList}
\title{Check if an Object is an \code{mgnetList} Instance}
\usage{
is_mgnetList(object)
}
\arguments{
\item{object}{The object to check.}
}
\value{
\code{TRUE} if the object is an instance of the \code{mgnetList} class, \code{FALSE} otherwise.
}
\description{
This function checks whether a given object is an instance of the \code{mgnetList} class.
}
