% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{isolate}
\alias{isolate}
\alias{isolate,mgnet-method}
\alias{isolate,list-method}
\title{Isolate Vertices}
\usage{
isolate(object, ..., condition="AND")

\S4method{isolate}{mgnet}(object, ..., condition = "AND")

\S4method{isolate}{list}(object, ..., condition = "AND")
}
\arguments{
\item{object}{mgnet.}

\item{...}{arbitrary set of vectors of functions. If vectors they must contain
numeric or name or logical indices to be preserved. If functions they must return
indices as just explained.}

\item{condition}{(Optional default "AND") character indicates the logical operation between indices.}
}
\description{
Remove vertices from their communities.
}
