% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{layout_signed}
\alias{layout_signed}
\title{Signed-Weighted Graph Layout}
\usage{
layout_signed(obj)
}
\arguments{
\item{obj}{network belong to igraph or mgnet class.}
}
\description{
It elaborates the coordinates for the representation of
the vertices of the graph considering only the links with a positive sign.
}
