% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utility.R
\name{layout_signed}
\alias{layout_signed}
\title{Signed-Weighted Graph Layout}
\usage{
layout_signed(graph, seed = 123)
}
\arguments{
\item{graph}{network belong to igraph class}

\item{seed}{random seed for reprodubility}
}
\description{
It elaborates the coordinates for the representation of
the vertices of the graph considering only the links with a positive sign.
}
