% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{length,mgnetList-method}
\alias{length,mgnetList-method}
\title{Length of an mgnetList}
\usage{
\S4method{length}{mgnetList}(x)
}
\arguments{
\item{x}{An \code{mgnetList} object.}
}
\value{
Integer value representing the number of \code{mgnet} objects in the \code{mgnetList}.
}
\description{
Returns the number of \code{mgnet} objects contained within an \code{mgnetList} object.
}
