% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_setters.R
\name{lineage<-}
\alias{lineage<-}
\alias{lineage<-,mgnet-method}
\alias{lineage<-,mgnetList-method}
\title{Set Lineage Data}
\usage{
lineage(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new lineage data to be set.}
}
\value{
The \code{mgnet} or \code{mgnetList} object with updated lineage data.
}
\description{
This setter function updates the lineage data for \code{mgnet} objects and
each \code{mgnet} object within a \code{mgnetList}. The lineage data must be provided as a character matrix
for \code{mgnet} objects. For \code{mgnetList} objects, the lineage data should be a named list of character matrices
corresponding to each \code{mgnet} object within the list.
}
