% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineage_HMP2}
\alias{lineage_HMP2}
\title{Lineage Information for OTUs in HMP2 Gut Dataset}
\format{
A data frame with 1953 rows and columns for each taxonomic rank from phylum to OTU. The row names are the taxa IDs.
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(lineage_HMP2)
}
\description{
This dataset contains the lineage information for each operational taxonomic
unit (OTU) identified in the \code{otu_HMP2} dataset from the Human Microbiome
Project Phase 2 (HMP2), focusing on gut microbiome samples.
Each row represents an OTU (1953 in total), and each column represents a
different level of taxonomic classification, from phylum to OTU.
The taxa IDs, which are equivalent to the OTU names in this dataset
(but can differ in general, such as with NCBI where names and IDs are distinct),
are provided as the row names.
}
\keyword{datasets}
