% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{link}
\alias{link}
\alias{link,mgnet-method}
\alias{link,mgnets-method}
\title{Get link-level data from \code{mgnet} and \code{mgnets} objects}
\usage{
link(object, selected = TRUE, .fmt = c("list", "tbl"))

\S4method{link}{mgnet}(object, selected = TRUE, .fmt = c("list", "tbl"))

\S4method{link}{mgnets}(object, selected = TRUE, .fmt = c("list", "tbl"))
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{selected}{Logical. Whether to return only selected links. Default is
\code{TRUE}.}

\item{.fmt}{For \code{mgnets}, output format. One of \code{"list"} or \code{"tbl"}.}
}
\value{
For \code{mgnet}, a data.frame of links.

For \code{mgnets}, either a named list of link tables or a single collapsed tibble
with an additional \code{mgnet} column.
}
\description{
Retrieve edge tables from the network stored in an \code{mgnet} or \code{mgnets}
object.

By default, only selected links are returned.
}
\details{
This method calls \code{\link[=netw]{netw()}} and then converts the graph(s) to edge tables with
\code{igraph::as_data_frame(..., what = "edges")}.
}
