% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_link.R
\name{linkwise}
\alias{linkwise}
\alias{linkwise,mgnet-method}
\alias{linkwise,mgnetList-method}
\title{Apply Functions Link-wise in an mgnet Object}
\usage{
linkwise(object)

\S4method{linkwise}{mgnet}(object)

\S4method{linkwise}{mgnetList}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
The same \code{mgnet} object or \code{mgnetList} with the operation mode set for link-wise manipulation.
}
\description{
Sets the operation mode of an \code{mgnet} or \code{mgnetList} object to be link-wise,
allowing subsequent operations to consider each link (or edge) individually
or in specified groups.
}
\details{
For \code{mgnet} objects, the function sets the operation mode so that subsequent
manipulations with \code{\link[=mutate_link]{mutate_link()}} can apply to individual links or defined groups of links.
For \code{mgnetList} objects, the function applies the settings to each contained \code{mgnet}.
}
