% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_setters.R
\name{log_abundance<-}
\alias{log_abundance<-}
\alias{log_abundance<-,mgnet-method}
\alias{log_abundance<-,mgnetList-method}
\title{Set Log-Transformed Abundance Data}
\usage{
log_abundance(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new log-transformed abundance data to be set.}
}
\value{
The \code{mgnet} or \code{mgnetList} object with updated log-transformed abundance data.
}
\description{
This setter function allows you to update the log-transformed abundance data for \code{mgnet} objects
and each \code{mgnet} object within a \code{mgnetList}. The log-transformed abundance data must be a numeric matrix
for \code{mgnet} objects. For \code{mgnetList} objects, the data should be a named list of numeric matrices
corresponding to each \code{mgnet} object within the list.
}
