% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_classes_getters.R
\name{log_abundance}
\alias{log_abundance}
\alias{log_abundance,mgnet-method}
\alias{log_abundance,mgnetList-method}
\title{Retrieve Log_Abundance Data at Specified Taxonomic Rank}
\usage{
log_abundance(object, rank = "missing", .fmt = "mat")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{rank}{A character string specifying the taxonomic rank of interest for data aggregation.
If not provided or if the rank is "missing", the function returns the original abundance
data without aggregation.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are "mat" for matrix, "df" for data.frame, and "tbl" for tibble.
When ".fmt" is set to "tbl", for \code{mgnet} objects, the row names of the abundance matrix
are moved into a new column named \code{sample_id}, aligning with the reserved keyword in
\code{mgnet}.}
}
\value{
For an \code{mgnet} object, abundance data formatted as specified by the \code{.fmt} parameter,
aggregated at the specified taxonomic rank if provided. For an \code{mgnetList} object,
a list of such formatted data from each \code{mgnet} object within the list, allowing
for batch processing and analysis of multiple datasets simultaneously.
}
\description{
Retrieves log-ratio transformed abundance data for an \code{mgnet} or \code{mgnetList} object,
optionally aggregated at a specified taxonomic rank, and allows the output to
be formatted as a matrix, data frame, or tibble.
}
\details{
This function returns the log-ratio transformed abundance data, which is useful
for compositional data analysis.
}
