% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{log_data}
\alias{log_data}
\alias{log_data,mgnet-method}
\alias{log_data,list-method}
\title{Retrieves log-transformed data.}
\usage{
log_data(object)

\S4method{log_data}{mgnet}(object)

\S4method{log_data}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Return the matrix associated with the log-transformed data
mg class.
}
