% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mgnet.R
\docType{methods}
\name{make_mgnet}
\alias{make_mgnet}
<<<<<<< HEAD
\alias{make_mgnet,mgnet,character,character,numeric-method}
\alias{make_mgnet,list,character,character,numeric-method}
\title{Build a meta-genomic network.}
\usage{
make_mgnet(mgnet, cor.method, thresh.method, thresh, adjust = NA)

\S4method{make_mgnet}{mgnet,character,character,numeric}(mgnet, cor.method, thresh.method, thresh, adjust = NULL)

\S4method{make_mgnet}{list,character,character,numeric}(mgnet, cor.method, thresh.method, thresh, adjust = NULL)
}
\arguments{
\item{mgnet}{mgnet object}
=======
\alias{make_mgnet,mg,character,character,character,numeric-method}
\alias{make_mgnet,mg,character,character,character,numeric}
\alias{make_mgnet,mg,character,character,missing,numeric-method}
\alias{make_mgnet,mg,character,character,missing,numeric}
\alias{make_mgnet,list,character,character,character,numeric-method}
\alias{make_mgnet,list,character,character,character,numeric}
\alias{make_mgnet,list,character,character,missing,numeric-method}
\alias{make_mgnet,list,character,character,missing,numeric}
\title{Build a meta-genomic network.}
\usage{
make_mgnet(mg, cor.method, thresh.method, adjust, thresh)

\S4method{make_mgnet}{mg,character,character,character,numeric}(mg, cor.method, thresh.method, adjust, thresh)

\S4method{make_mgnet}{mg,character,character,missing,numeric}(mg, cor.method, thresh.method, adjust, thresh)

\S4method{make_mgnet}{list,character,character,character,numeric}(mg, cor.method, thresh.method, adjust, thresh)

\S4method{make_mgnet}{list,character,character,missing,numeric}(mg, cor.method, thresh.method, adjust, thresh)
}
\arguments{
\item{mg}{mg object}
>>>>>>> main

\item{cor.method}{correlation type, possible choices are "pearson","spearman","kendall".}

\item{thresh.method}{threshold method to retrieve the adjacency matrix. Possible
choices are "absolute","density","p-value".}

<<<<<<< HEAD
\item{thresh}{numeric indicates the threshold value.}

\item{adjust}{multiple test criteria, possible choices are "holm",
"hochberg","hommel","bonferroni","BH","BY","fdr","none".}
=======
\item{adjust}{multiple test criteria, possible choices are "holm",
"hochberg","hommel","bonferroni","BH","BY","fdr","none".}

\item{thresh}{numeric indicates the threshold value.}
>>>>>>> main
}
\description{
The function has the purpose of building a meta-genomic network
starting from a mg object following the following steps:
\itemize{
\item calculates the clr abundances from data and if there are zeros,
add the value 1 to all elements.
\item Evaluate the correlations between all possible pairs of taxa being able
to choose between Pearson Kendall and Spearman.
\item derives the adjacency matrix by applying a threshold to the lowest
correlations with three possible criteria. The simplest by giving it the
absolute value, based on the density of the links or considering the p-value
corrected with multiple tests.
\item constructs the undirected, weighted, and signed network of class igraph.
}
}
