% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,ANY-method}
\alias{meta,sample_metadata-method}
\alias{meta,mg-method}
\title{Retrieves sample metadata.}
\usage{
meta(object)

\S4method{meta}{ANY}(object)

\S4method{meta}{sample_metadata}(object)

\S4method{meta}{mg}(object)
}
\arguments{
\item{object}{Required \code{\link{mg-class}} or \code{\link{sample_metadata-class}}.}
}
\description{
Return the data.frame associated with the sample metadata with experimental
variables from mg or sample_metadata classes.
}
