% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{meta<-}
\alias{meta<-}
\alias{meta<-,mgnet-method}
\alias{meta<-,mgnets-method}
\title{Update Sample Metadata for mgnet and mgnets Objects}
\usage{
meta(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object to be updated.}

\item{value}{The new sample metadata to set, either a dataframe or a named list
of dataframes as per the object type.}
}
\value{
The updated \code{mgnet} or \code{mgnets} object with new sample metadata.
}
\description{
This function sets the sample metadata for \code{mgnet} objects and also updates
each \code{mgnet} object within a \code{mgnets}. The metadata for \code{mgnet} objects must be
provided as a dataframe. For \code{mgnets} objects, the metadata should be either
a named list of dataframes corresponding to each \code{mgnet} object within the list,
or a single dataframe containing a 'mgnet' and 'sample_id' columns to split the
metadata accordingly.
}
