% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.HMP2}
\alias{meta.HMP2}
\title{Samples metadata of otu.HMP2}
\format{
\subsection{\code{meta.HMP2}}{

A data.frame with 1122 rows/sample and 11 columns/info.
The main focus are the Subjects ID, different columns with the health status,
the sensibility or the resistance to the insulin and the time for each samples.
}
}
\source{
<<<<<<< HEAD
George Weinstock
=======
Jackson Laboratory
>>>>>>> main
}
\usage{
meta.HMP2
}
\description{
Samples metadata of otu.HMP2
}
\keyword{datasets}
