% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_HMP2}
\alias{meta_HMP2}
\title{Sample Metadata for HMP2 Gut Dataset}
\format{
A data frame with 1122 rows and 10 variables:
\describe{
\item{SubjectID}{Identifier for the subject from whom the sample was taken.}
\item{DateZero}{The number of days passed from the first specimen collected in the study.}
\item{DateSubject}{The number of days passed from the first specimen collected from the subject.}
\item{IRIS}{Indicator of insulin resistance or sensitivity.}
\item{CL4_1, CL4_2, CL1, CL2, CL3, CL4}{Various indicators of the health status of the patient at the time of the specimen.}
}
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(meta_HMP2)
}
\description{
This dataset provides the experimental and metadata information corresponding to the samples in the \code{otu_HMP2} dataset.
Each row matches a sample from \code{otu_HMP2}, and columns include various metadata attributes such as Subject ID, dates,
insulin resistance status, and health status indicators.
}
\keyword{datasets}
