% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{meta_sample}
\alias{meta_sample}
\alias{meta_sample,mgnet-method}
\alias{meta_sample,list-method}
\title{Retrieves sample metadata.}
\usage{
meta_sample(object)

\S4method{meta_sample}{mgnet}(object)

\S4method{meta_sample}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Return the data.frame associated with the sample metadata with experimental
variables from mg class.
}
