% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{meta_taxa}
\alias{meta_taxa}
\alias{meta_taxa,mgnet-method}
\alias{meta_taxa,list-method}
\title{Retrieves taxa additional info.}
\usage{
meta_taxa(object)

\S4method{meta_taxa}{mgnet}(object)

\S4method{meta_taxa}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Return the data.frame associated with the additional taxa information from
mg class.
}
