% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_classes_getters.R
\name{meta_vars}
\alias{meta_vars}
\alias{meta_vars,mgnet-method}
\alias{meta_vars,mgnetList-method}
\title{Get Sample Metadata Variables}
\usage{
meta_vars(object, .fmt = "list")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{.fmt}{Character; specifies the output format when checking an \code{mgnetList}.
Accepted values are \code{"list"} for a list of sample metadata variables
in each \code{mgnet} and \code{"unique"} for a single array with the unique elemnts.
Default is \code{"list"}.}
}
\value{
For an \code{mgnet} object, a character vector of metadata variable names.
For an \code{mgnetList} object, a named list of character vectors, with each list item representing
the metadata variable names in the corresponding \code{mgnet} objects.
}
\description{
Retrieves the names of metadata variables available in the \code{meta} slot of an \code{mgnet} object,
or for each \code{mgnet} object within an \code{mgnetList}.
}
\examples{
data(mg, package = "mgnet")
meta_vars(mg)  

data(mgl, package = "mgnet")
meta_vars(mgl, .fmt = "list)  
meta_vars(mgl, .fmt = "unique) 

}
