% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_group.R
\name{metawise}
\alias{metawise}
\alias{metawise,mgnet-method}
\alias{metawise,mgnetList-method}
\title{Set Meta-wise Operation Mode for an mgnet Object}
\usage{
metawise(object)

\S4method{metawise}{mgnet}(object)

\S4method{metawise}{mgnetList}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
The modified \code{mgnet} object or \code{mgnetList} with updated grouping attributes.
}
\description{
This function sets the operation mode of an \code{mgnet} or \code{mgnetList} object to be sample-wise,
similar to dplyr's rowwise but specific to meta elements.
}
\details{
\itemize{
\item For \code{mgnet} objects, sets the grouping variable to \code{sample_id}.
\item For \code{mgnetList} objects, sets the grouping variable to \code{mgnet}, \code{sample_id}.
}
}
