% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg-class.R
\docType{class}
\name{mg-class}
\alias{mg-class}
\alias{mg}
\title{S4 class to manage metagenomics data}
\description{
S4 class to manage metagenomics data
}
\section{Slots}{

\describe{
\item{\code{data}}{numeric matrix with data}

\item{\code{meta}}{sample features as data.frame}

\item{\code{taxa}}{character matrix with taxonomic information}
}}

