% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-definitions-validators-constructors.R
\name{mg}
\alias{mg}
\title{User constructor for mg s4 class.}
\usage{
mg(data = NULL, meta = NULL, taxa = NULL)
}
\arguments{
\item{data}{numeric matrix or ngs_data s4 obj with all elements >=0.}

\item{meta}{data.frame or sample_metadata s4 obj with experimental variables.}

\item{taxa}{character matrix or taxonomy_table s4 obj with taxonomic classification.}
}
\description{
User constructor to create an object belonging to formal s4 class mg
avoiding the new function.
}
