% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{mgmelt}
\alias{mgmelt}
\alias{mgmelt,mgnet-method}
\alias{mgmelt,list-method}
=======
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{mgmelt}
\alias{mgmelt}
\alias{mgmelt,mg-method}
\alias{mgmelt,list-method}
\alias{mgmelt,list}
>>>>>>> main
\title{Melt mg object into form suitable for easy casting.}
\usage{
mgmelt(object)

<<<<<<< HEAD
\S4method{mgmelt}{mgnet}(object)
=======
\S4method{mgmelt}{mg}(object)
>>>>>>> main

\S4method{mgmelt}{list}(object)
}
\arguments{
<<<<<<< HEAD
\item{object}{mgnet.}
=======
\item{object}{(Required) \code{\link{mg-class}}.}
>>>>>>> main
}
\description{
Summarize all elements present in an mg object in a single data frame.
The functioning is similar to melt function of reshape2 package and it will
be useful as a preprocess for ggplot2.
}
