% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\name{mgnet}
\alias{mgnet}
\alias{mgnet,constructor}
\title{User constructor for mgnet s4 class.}
\usage{
mgnet(
  data = matrix(nrow = 0, ncol = 0),
  meta = data.frame(),
  taxa = matrix(nrow = 0, ncol = 0),
  netw = make_empty_graph(n = 0, directed = FALSE),
  comm = cluster_fast_greedy(make_empty_graph(n = 0, directed = FALSE)),
  mg = new("mg")
)
}
\arguments{
\item{data}{numeric matrix with all elements >=0.}

\item{meta}{data.frame with experimental variables.}

\item{taxa}{character matrix with taxonomic classification.}

\item{netw}{un-directed weighted igraph network.}

\item{comm}{communities object associated to netw.}

\item{mg}{object belong to mg class.}
}
\description{
User constructor to create an object belonging to formal s4
class mgnet avoiding the new function.
}
