% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-merge.R
\name{mgnet_collapse}
\alias{mgnet_collapse}
\title{Collapse groups of \code{mgnet} objects within an \code{mgnets}}
\usage{
mgnet_collapse(
  object,
  by,
  abun = NULL,
  rela = NULL,
  norm = NULL,
  meta = NULL,
  taxa = NULL,
  netw = NULL,
  comm = NULL
)
}
\arguments{
\item{object}{An object of class \code{mgnets}.}

\item{by}{A named or unnamed list defining groups of \code{mgnet} names to merge.
Each element must be a character vector of names found in \code{object}.}

\item{abun, rela, norm, meta, taxa, netw, comm}{Functions used to merge the
corresponding slots.}
}
\value{
An object of class \code{mgnets}, where each output element is the merge of one
group specified in \code{by}.
}
\description{
Collapse subsets of an \code{mgnets} object into merged \code{mgnet} objects according
to a user-specified grouping.
}
\details{
If \code{by} is unnamed, names are generated by concatenating the grouped element
names with \code{"-"}.

Output group names must be unique.
}
