% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_mgnetList_class.R
\name{mgnetList-access}
\alias{mgnetList-access}
\alias{$,mgnetList-method}
\alias{$<-,mgnetList-method}
\title{Access or Assign Elements in an mgnetList Object}
\usage{
\S4method{$}{mgnetList}(x, name)

\S4method{$}{mgnetList}(x, name) <- value
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{name}{A character string representing the name of the \code{mgnet} object to extract or assign.}

\item{value}{A \code{mgnet} object to be assigned to the \code{mgnetList}.}
}
\value{
\itemize{
\item For \code{$}, the method returns the \code{mgnet} object corresponding to \code{name}.
\item For \verb{$<-}, the method assigns a new \code{mgnet} object to the list and returns the modified \code{mgnetList}.
}
}
\description{
These methods allow accessing or assigning individual \code{mgnet} objects within an \code{mgnetList}
using the \code{$} and \verb{$<-} operators, providing list-like behavior for the \code{mgnetList} class.
}
\details{
\code{mgl$name} extracts the \code{mgnet} object from the \code{mgnetList} named \code{name}. If no
such object exists, an error is thrown.

\code{mgl$name <- value} assigns a new \code{mgnet} object to the element named \code{name} in
the \code{mgnetList}. If the value is not a valid \code{mgnet} object, an error is thrown. After the
assignment, the validity of the entire \code{mgnetList} object is checked to ensure it adheres
to the defined constraints.
}
\section{Usage}{

\preformatted{
mgl$name
mgl$name <- value
}
}

\examples{
data(mgl, package = "mgnet")

# Access a specific mgnet object
mgnet_obj <- mgl$`69-001`

# Assign a new mgnet object to the mgnetList
mgl$new_sample <- mgnet()

}
