% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mgnetList_class.R
\docType{class}
\name{mgnetList-class}
\alias{mgnetList-class}
\title{MetaGenomic NETwork List Class (mgnetList)}
\description{
An S4 class designed to hold a list of \code{mgnet} objects. This class ensures that all elements
within the list are instances of the \code{mgnet} class, providing a structured and type-safe way
to manage collections of metagenomic network data.
}
\section{Slots}{

\describe{
\item{\code{mgnets}}{A list where each element is an instance of the \code{mgnet} class.}
}}

\seealso{
\code{\link{mgnet}} for details on the \code{mgnet} class.
}
