% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mgnetList_class.R
\name{mgnetList}
\alias{mgnetList}
\title{Create an mgnetList Object}
\usage{
mgnetList(mgnets = list())
}
\arguments{
\item{mgnets}{An optional list of \code{mgnet} objects to initialize the \code{mgnetList}.
Defaults to an empty list. Each element must be an instance of the \code{mgnet} class.}
}
\value{
An \code{mgnets} object.
}
\description{
This function constructs an \code{mgnetList} object, which is designed to encapsulate and manage
a collection of \code{mgnet} objects in a structured list.
}
