% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnets.R
\name{mgnets-access}
\alias{mgnets-access}
\alias{$,mgnets-method}
\alias{$<-,mgnets,mgnet-method}
\alias{[[,mgnets-method}
\alias{[[<-,mgnets,ANY,ANY,mgnet-method}
\alias{[,mgnets-method}
\alias{[<-,mgnets,ANY,ANY,list-method}
\title{Access or assign \code{mgnet} elements in an \code{mgnets} object}
\usage{
\S4method{$}{mgnets}(x, name)

\S4method{$}{mgnets,mgnet}(x, name) <- value

\S4method{[[}{mgnets}(x, i, j, ...)

\S4method{[[}{mgnets,ANY,ANY,mgnet}(x, i, j) <- value

\S4method{[}{mgnets}(x, i, j, ..., drop = TRUE)

\S4method{[}{mgnets,ANY,ANY,list}(x, i, j) <- value
}
\arguments{
\item{x}{An object of class \code{mgnets}.}

\item{name}{A single character string giving the element name for \code{$} and \verb{$<-}.}

\item{value}{An object of class \code{mgnet} to assign.}

\item{i}{An index or name for \code{[[} and \verb{[[<-}.}

\item{j}{Ignored. Included only for signature compatibility.}

\item{...}{Not used.}
}
\value{
\itemize{
\item \code{$} and \code{[[} return a single \code{mgnet} object.
\item \verb{$<-} and \verb{[[<-} return the updated \code{mgnets} object.
}
}
\description{
List-like accessors for \code{mgnets} objects using \code{$}, \verb{$<-}, \code{[}, \verb{[<-}, \code{[[}, and \verb{[[<-}.

These methods allow convenient extraction and replacement of individual
\code{mgnet} objects stored in the collection.
}
