% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnets.R
\name{mgnets-methods}
\alias{mgnets-methods}
\alias{as.list.mgnets}
\alias{coerce,mgnets,list-method}
\alias{length,mgnets-method}
\alias{names,mgnets-method}
\alias{names<-,mgnets,character-method}
\title{General methods for \code{mgnets} objects}
\usage{
\method{as.list}{mgnets}(x, ...)

\S4method{length}{mgnets}(x)

\S4method{names}{mgnets}(x)

\S4method{names}{mgnets,character}(x) <- value
}
\arguments{
\item{x}{An object of class \code{mgnets}.}

\item{...}{Not used.}

\item{value}{A character vector used in the replacement form \verb{names<-}.}
}
\value{
\itemize{
\item \code{as.list()} and \code{as(., "list")} return a list of \code{mgnet} objects.
\item \code{length()} returns an integer of length 1.
\item \code{names()} returns a character vector.
\item \verb{names<-()} returns the updated \code{mgnets} object.
}
}
\description{
Basic utilities for interacting with \code{mgnets} objects, including coercion to
a base list, retrieval of length, and access to element names.
}
\section{Methods}{

\describe{
\item{\code{as.list(x)}}{Convert an \code{mgnets} object to a plain list of \code{mgnet} objects.}
\item{\code{as(x, "list")}}{Coerce an \code{mgnets} object to a base list.}
\item{\code{length(x)}}{Return the number of contained \code{mgnet} objects.}
\item{\code{names(x)}}{Return the names of the contained \code{mgnet} objects.}
\item{\code{names(x) <- value}}{Replace the names of the contained \code{mgnet} objects.}
}
}

