% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{miss_sample}
\alias{miss_sample}
\alias{miss_sample,mgnet-method}
\alias{miss_sample-mgnetList}
\title{Check if Samples are Missing in \code{mgnet} or \code{mgnetList} Objects}
\usage{
miss_sample(object, .fmt = "list")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{.fmt}{Character; specifies the output format when checking an \code{mgnetList}.
Accepted values are \code{"list"} for a list of logical values, \code{"any"} for
a single logical indicating if any \code{mgnet} is missing samples, and \code{"all"}
for a single logical indicating if all \code{mgnet} objects are missing samples.
Default is \code{"list"}.}
}
\value{
For an \code{mgnet} object, a logical indicating whether samples are missing.
For an \code{mgnetList} object, the output depends on the value of \code{.fmt}.
If \code{.fmt = "list"}, returns a named logical vector. If \code{.fmt = "any"},
returns a single logical value indicating if any \code{mgnet} is missing samples.
If \code{.fmt = "all"}, returns a single logical value indicating if all \code{mgnet}
objects are missing samples.
}
\description{
This function checks whether samples are missing (i.e., absent or empty) in an \code{mgnet} object or
in each \code{mgnet} object within an \code{mgnetList}. For \code{mgnetList} objects, the output format can
be a list, or a single logical value indicating if any or all objects are missing samples.
}
\examples{
data(mg, package = "mgnet")
miss_sample(mg)  

data(mgl, package = "mgnet")
miss_sample(mgl, .fmt = "list")  
miss_sample(mgl, .fmt = "any")  
miss_sample(mgl, .fmt = "all")  

}
