% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_classes_mutate.R
\name{mutate_sample}
\alias{mutate_sample}
\alias{mutate_sample,mgnet-method}
\alias{mutate_sample,mgnetList-method}
\title{Modify and Augment \code{mgnet} Objects by Transforming \code{sample} Slot}
\usage{
mutate_sample(object, ..., .by)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object from which data will be modified.
This function specifically targets the \code{sample} slot, which stores metadata for each sample.}

\item{...}{Dynamic expressions or functions to be applied to the data, utilizing
\code{\link[dplyr]{mutate}} directly. This integration supports all standard and custom
\code{tidyverse} manipulation techniques, ensuring versatile and complex data transformations.}

\item{.by}{Optional; a character vector indicating the columns used for grouping data
before applying transformations. For \code{mgnet} objects, defaults to "sample_id".
For \code{mgnetList} objects, defaults to \code{c("mgnet", "sample_id")}, where "mgnet"
refers to individual \code{mgnet} objects within the list, ensuring transformations are
applied contextually within each sub-object. The use of "taxa_id" as a grouping variable
is explicitly prohibited to maintain consistency and prevent data misinterpretation.}
}
\value{
Returns the \code{mgnet} or \code{mgnetList} object with updated \code{sample} slots reflecting
the applied transformations. All other structures within the object remain unchanged, ensuring
that only the targeted metadata is modified.
}
\description{
This function dynamically manipulates the \code{sample} data slot within \code{mgnet} objects,
applying user-defined transformations. It harnesses the comprehensive functionality of the
\code{tidyverse}, particularly \code{dplyr}, to facilitate these data transformations.
}
\details{
This function is designed to enhance and modify the metadata of samples stored within
the \code{sample} slot using conditional logic and statistical summaries tailored to the needs
of the analysis. It is built to integrate seamlessly with the \code{tidyverse} methods, promoting
familiar and powerful data manipulation capabilities directly within this package. Users are
encouraged to apply any \code{tidyverse}-compatible expressions or \code{dplyr} verbs to achieve the desired
data transformation, including but not limited to conditional operations, summarizations, or the
use of \code{dplyr::across}.
}
