% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_classes_mutate_info.R
\name{mutate_sample_info}
\alias{mutate_sample_info}
\alias{mutate_sample_info,mgnet-method}
\title{Modify and Augment Sample Metadata in \code{mgnet} Objects}
\usage{
mutate_sample_info(object, ..., .by)
}
\arguments{
\item{object}{An \code{mgnet} object. This function targets the \code{meta} slot, which contains
metadata for each sample in the object.}

\item{...}{Dynamic expressions or functions to be applied to the metadata.
These expressions can modify metadata fields in the \code{meta} slot.}

\item{.by}{A character vector specifying the columns to group the data by before applying
transformations. The default is \code{"sample_id"}. Grouping ensures that transformations
are contextually applied within each subgroup defined by \code{.by}. The use of \code{"taxa_id"}
as a grouping variable is strictly prohibited to maintain consistency.}
}
\value{
Returns the modified \code{mgnet} object with the transformed \code{meta} slot reflecting
the applied transformations. All other slots and structure within the \code{mgnet}
object remain unchanged.
}
\description{
This function dynamically manipulates the \code{meta} slot within \code{mgnet} objects by applying
user-defined transformations to the sample metadata. It leverages the \code{tidyverse} tools
such as \code{dplyr} to enable flexible data transformations.
}
