% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mutate.R
\name{mutate_taxa}
\alias{mutate_taxa}
\alias{mutate_taxa,mgnet-method}
\alias{mutate_taxa,mgnets-method}
\title{Modify and Augment \code{mgnet} Objects by Transforming the \code{taxa} Slot}
\usage{
mutate_taxa(object, ..., .ungroup = FALSE)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.
The function targets the \code{taxa} slot, which contains metadata for each taxon.}

\item{...}{Dynamic expressions or functions to be applied to the data.
These expressions can manipulate both abundance data (e.g., 'abun', 'rela', 'norm') and
metadata within the taxa slot. This allows for a comprehensive data transformation
experience that supports all standard and custom \code{tidyverse} manipulation techniques.}

\item{.ungroup}{Logical, default \code{FALSE}. If \code{TRUE}, remove the mgnet grouping
from the object at the end (i.e., drop the grouping attribute).

\if{html}{\out{<div class="sourceCode">}}\preformatted{     ### Keywords in `mgnet` and `mgnets`:
     - **abun, rela, norm**: Slots within `mgnet` objects that store abundance data, which can be
       directly manipulated or used in conjunction with metadata to perform advanced analyses.
     - **taxa_id**: An essential identifier used to uniquely reference individual taxon within an `mgnet` object. 
     - **mgnet**: Used exclusively within `mgnets` objects to differentiate between multiple `mgnet` objects 
       contained in the list.
}\if{html}{\out{</div>}}}
}
\value{
Returns the \code{mgnet} or \code{mgnets} object with updated \code{taxa} slots reflecting the applied transformations.
All other structures within the object remain unchanged, ensuring that only the targeted metadata is modified.
}
\description{
This function dynamically manipulates the \code{taxa} slot within \code{mgnet} or \code{mgnets} objects,
applying user-defined transformations. It leverages the full suite of \code{tidyverse} tools, particularly
\code{dplyr}, to enable powerful and flexible data transformations.
}
