% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_classes_mutate_data.R
\name{mutate_taxa_data}
\alias{mutate_taxa_data}
\alias{mutate_taxa_data,mgnet-method}
\alias{mutate_taxa_data,mgnetList-method}
\title{Modify and Augment Taxa Data in \code{mgnet} and \code{mgnetList} Objects}
\usage{
mutate_taxa_data(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object. For \code{mgnet}, the function modifies taxa
metadata and abundance data within the object. For \code{mgnetList}, the same transformation
is applied to each \code{mgnet} object in the list.}

\item{...}{Dynamic expressions or functions to be applied to the taxa data and metadata.
These expressions can include both abundance-related fields (e.g., 'abun', 'rela', 'norm')
and metadata fields.}
}
\value{
Returns the modified \code{mgnet} or \code{mgnetList} object with the transformed \code{meta} slot
reflecting the applied transformations. All other structures within the object remain
unchanged.
}
\description{
This function dynamically manipulates both the taxa metadata and abundance data within
\code{mgnet} and \code{mgnetList} objects, applying user-defined transformations. It leverages the \code{tidyverse} tools
such as \code{dplyr} to enable flexible data transformations.
}
\section{Reserved Keywords}{

Reserved keywords such as \code{sample_id}, \code{taxa_id}, \code{comm_id}, \code{abun}, \code{rela}, \code{norm}, and \code{mgnet}
are essential to the structure of the \code{mgnet} and \code{mgnetList} objects. They should not be
overwritten or modified by user-defined transformations. Any attempt to modify these keywords
will result in an error.
}

