% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_classes_utilities.R
\name{names,mgnetList-method}
\alias{names,mgnetList-method}
\title{Names of mgnet Objects in an mgnetList}
\usage{
\S4method{names}{mgnetList}(x)
}
\arguments{
\item{x}{An \code{mgnetList} object.}
}
\value{
A character vector of names of the \code{mgnet} objects.
}
\description{
Retrieves the names of \code{mgnet} objects contained within an \code{mgnetList}.
Names provide a convenient way to reference and manage individual \code{mgnet} objects.
}
