% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{names<-,mgnetList-method}
\alias{names<-,mgnetList-method}
\title{Set Names of mgnet Objects in an mgnetList}
\usage{
\S4method{names}{mgnetList}(x) <- value
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{value}{A character vector representing the new names to be assigned to the \code{mgnet} objects.}
}
\value{
The modified \code{mgnetList} object with updated names.
}
\description{
Sets the names of \code{mgnet} objects contained within an \code{mgnetList}.
Names provide a convenient way to reference and manage individual \code{mgnet} objects.
}
