% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{ncomm}
\alias{ncomm}
\alias{ncomm,mgnet-method}
\alias{ncomm,list-method}
\title{Get communities number.}
\usage{
ncomm(object)

\S4method{ncomm}{mgnet}(object)

\S4method{ncomm}{list}(object)
}
\arguments{
\item{object}{mgnet or mgnet-list.}
}
\description{
Return an integer indicates the communities present in comm slot.
}
