% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{ncomm}
\alias{ncomm}
\alias{ncomm,mgnet-method}
\alias{ncomm,mgnetList-method}
\title{Get the Number of Communities}
\usage{
ncomm(object)
}
\arguments{
\item{object}{An object of class \code{mgnet} or \code{mgnetList}. For \code{mgnet}, it calculates the number of communities directly from the \code{comm} slot.
For \code{mgnetList}, it applies the calculation to each \code{mgnet} object within the list and returns a vector of results.}
}
\value{
For an \code{mgnet} object, returns an integer representing the number of communities.
For an \code{mgnetList} object, returns a numeric named vector with each element representing the number of communities in each \code{mgnet} object.
}
\description{
This function returns an integer indicating the number of communities detected in a network analysis represented by an \code{mgnet} object,
or a named vector of integers for an \code{mgnetList} object, where each element corresponds to the number of communities in each respective \code{mgnet} object.
For \code{mgnet} objects without any communities, the function returns 0. For \code{mgnetList} objects, it iteratively applies this logic to each contained \code{mgnet} object.
}
\details{
Retrieves the number of communities present in the \code{comm} slot of an \code{mgnet} object or each \code{mgnet} object within an \code{mgnetList}.
}
\examples{
data(mg, package = "mgnet")
ncomm(mg)  

data(mgl, package = "mgnet")
ncomm(mgl)  

}
