% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{netw<-}
\alias{netw<-}
\alias{netw<-,mgnet-method}
\alias{netw<-,mgnets-method}
\title{Set Network Data}
\usage{
netw(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{value}{The new network data to be set.}
}
\value{
The \code{mgnet} or \code{mgnets} object with updated network data.
}
\description{
This setter function updates the network data for \code{mgnet} objects and
each \code{mgnet} object within a \code{mgnets}. The network data must be an \code{igraph} object
for \code{mgnet} objects. For \code{mgnets} objects, the network data should be a named list of \code{igraph} objects
corresponding to each \code{mgnet} object within the list.
}
