% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg-class.R
\docType{class}
\name{NGS_data-class}
\alias{NGS_data-class}
\alias{NGS_data}
\title{The S4 class to store NGS data.}
\description{
The information is saved in a  \code{\link{matrix}} where the rows express the samples and
the columns the taxa.

\describe{
\item{.Data}{This slot is the \code{\link{matrix}} containing NGS data.}
}
}
