% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-definitions-validators-constructors.R
\name{ngs_data}
\alias{ngs_data}
\title{User constructor for ngs_data s4 class.}
\usage{
ngs_data(data = NULL)
}
\arguments{
\item{data}{numeric matrix with all elements >=0.}
}
\description{
User constructor to create an object belonging to formal s4 class ngs_data
avoiding the new function.
}
