% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{norm<-}
\alias{norm<-}
\alias{norm<-,mgnet-method}
\alias{norm<-,mgnets-method}
\title{Set Log-Transformed Abundance Data}
\usage{
norm(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{value}{The new log-transformed abundance data to be set.}
}
\value{
The \code{mgnet} or \code{mgnets} object with updated log-transformed abundance data.
}
\description{
This setter function allows you to update the log-transformed abundance data for \code{mgnet} objects
and each \code{mgnet} object within a \code{mgnets}. The log-transformed abundance data must be a numeric matrix
for \code{mgnet} objects. For \code{mgnets} objects, the data should be a named list of numeric matrices
corresponding to each \code{mgnet} object within the list.
}
