% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_classes_getters.R
\name{norm}
\alias{norm}
\alias{norm,mgnet-method}
\alias{norm,mgnetList-method}
\title{Retrieve Normalized Abundance Data Based on Taxonomic Information}
\usage{
norm(object, .var = NULL, .fmt = "mat", .fun = sum)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object containing normalized abundance and informational data.}

\item{.var}{A character string specifying the column name from the \code{taxa} slot to be used
for grouping and aggregation. If this parameter is omitted, the function will return
the original normalized abundance data without any aggregation.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are:
- "mat": returns a matrix
- "df": returns a data.frame
- "tbl": returns a tibble, where for \code{mgnet} objects, the row names of the normalized abundance
matrix are moved into a new column named \code{sample_id}.
The default format is "mat".}

\item{.fun}{A function to specify how the relative abundance data should be aggregated.
This can be any function that takes a numeric vector and returns a single number (e.g., sum, mean).
The default aggregation function is sum.}
}
\value{
Depending on the '.fmt' parameter:
- For an \code{mgnet} object, returns normalized abundance data formatted as specified, aggregated
based on the provided column if specified.
- For an \code{mgnetList} object, returns a list of such formatted data from each \code{mgnet}
object within the list, enabling batch processing and analysis of multiple datasets.
}
\description{
Retrieves and aggregates normalized abundance data for an \code{mgnet} or \code{mgnetList} object based on
a specified column in the \code{taxa} data frame. This function allows for custom aggregation
methods such as sum or mean. If no column is specified, it returns the original normalized abundance data
in the selected format. Although \code{taxa} can contain various types of information, it is
commonly used to obtain normalized abundances at different taxonomic ranks or other categorical variables.
}
