% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{nrank}
\alias{nrank}
\alias{nrank,mgnet-method}
\alias{nrank,list-method}
\title{Get taxonomic ranks number.}
\usage{
nrank(object)

\S4method{nrank}{mgnet}(object)

\S4method{nrank}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return taxonomy ranks number number as integer.
}
