% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{nsample}
\alias{nsample}
\alias{nsample,mgnet-method}
\alias{nsample,list-method}
\title{Get number of samples.}
\usage{
nsample(object)

\S4method{nsample}{mgnet}(object)

\S4method{nsample}{list}(object)
}
\arguments{
\item{object}{mgnet ot list.}
}
\description{
Return an integer indicating the number of sample.
}
