% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{nsample}
\alias{nsample}
\alias{nsample,mg-method}
\alias{nsample,mg}
\title{Get number of samples.}
\usage{
nsample(object)

\S4method{nsample}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return an integer indicating the number of sample.
}
