% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{ntaxa}
\alias{ntaxa}
\alias{ntaxa,mgnet-method}
\alias{ntaxa,list-method}
\title{Get number of taxa}
\usage{
ntaxa(object)

\S4method{ntaxa}{mgnet}(object)

\S4method{ntaxa}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return an integer indicating the number of taxa
}
