% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-base-methods.R
\name{ntaxa}
\alias{ntaxa}
\alias{ntaxa,mgnet-method}
\alias{ntaxa,mgnets-method}
\title{Get Number of Taxa}
\usage{
ntaxa(object)
}
\arguments{
\item{object}{An object of class \code{mgnet} or \code{mgnets}.}
}
\value{
Integer (for \code{mgnet}) or named integer vector (for \code{mgnets}).
}
\description{
Returns the number of taxa in an \code{mgnet} or \code{mgnets} object.
}
\details{
For an empty object, returns \code{0L}. For \code{mgnets}, returns a named integer
vector with one value per element.
}
