% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu.HMP2}
\alias{otu.HMP2}
\title{Collection of 16s Gut Bacteria part of Human Microbiome Project 2 (HMP2)}
\format{
\subsection{\code{otu.HMP2}}{

A matrix with 1122 rows/sample and 1953 columns/OTUs:
}
}
\source{
Jackson Laboratory
}
\usage{
otu.HMP2
}
\description{
Collection of 16s Gut Bacteria part of Human Microbiome Project 2 (HMP2)
}
\keyword{datasets}
