% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_HMP2}
\alias{otu_HMP2}
\title{OTU Data from HMP2 Gut Dataset}
\format{
A data frame with 1122 rows and 1953 columns.
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(otu_HMP2)
}
\description{
This dataset contains operational taxonomic unit (OTU) abundance data from the Human Microbiome Project Phase 2 (HMP2),
focusing on gut microbiome samples. The dataset features 1122 samples (rows) across 1953 OTUs (columns).
}
\keyword{datasets}
