% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9b_classes_plot.R
\name{plot.mgnet}
\alias{plot.mgnet}
\title{Plot mgnet Network}
\usage{
\method{plot}{mgnet}(
  x,
  layout,
  ...,
  expFactor = 1,
  multConst = 1,
  sumConst = 0,
  thickness = TRUE,
  alphaFactor = 0.5,
  widthFactor = 1,
  posCol = rgb(0, 0, 1),
  negCol = rgb(1, 0, 0),
  maxSize = NULL,
  maxWidth = NULL,
  force.positive = TRUE
)
}
\arguments{

\item{x}{An \code{mgnet} object containing the network to be plotted.}

\item{layout}{An igraph layout function or coordinates for vertex positions.
If missing, \code{layout_signed} is applied, focusing on the structure formed by positive edges.}

\item{...}{Additional arguments passed to \code{igraph::plot.igraph}, allowing further
customization of the plot. This includes all parameters supported by \code{plot.igraph},
such as \code{vertex.label}, \code{edge.arrow.size}, among others.}

\item{expFactor}{A numeric value that adjusts the relative sizes of vertices.
Higher values increase the size difference between smaller and larger vertices.}

\item{multConst}{A positive numeric value applied as a multiplicative factor to all vertex sizes.}

\item{sumConst}{A positive numeric value added to the size of all vertices,
allowing for baseline size adjustment.}

\item{thickness}{A logical value; when \code{TRUE}, scales edge widths based on their absolute weights.}

\item{alphaFactor}{A numeric value in [0,1], adjusting the transparency of edge colors.}

\item{widthFactor}{A positive numeric value that scales the width of all edges uniformly.}

\item{posCol}{The color used for positive edges, specified as an RGB value.}

\item{negCol}{The color used for negative edges, specified as an RGB value.}

\item{maxSize}{An optional positive integer specifying the maximum size for vertices.
If set, vertex sizes are scaled such that the largest vertex has this size.}

\item{maxWidth}{An optional positive integer specifying the maximum width for edges.
If set, edge widths are scaled such that the widest edge has this width.}

\item{force.positive}{A logical indicating whether to adjust vertex sizes to be strictly positive
by offsetting them by the minimum size value found in the network.}
}
\value{
Generates a plot of the specified \code{mgnet} network with the given aesthetic parameters.
}
\description{
\code{plot.mgnet} focuses on the visualization of signed networks. It highlights the positive
connections while still allowing adjustments to graphical representations of both vertices and
edges. The function is highly customizable, supporting various igraph layout options and
additional aesthetic parameters.

Vertex sizes are dynamically calculated using the formula:
\deqn{v=\text{\footnotesize{sumConst}}+\text{\footnotesize{multConst}}\left(\text{\footnotesize{max}}(v_0)\left(\frac{v_0}{\text{\footnotesize{max}}(v_0)}\right)^{\text{expFactor}}\right)}
where \eqn{v_0} is the initial vertex size. This formula allows for the enhancement of
size differences between vertices, making the network's structure more apparent.
}
\details{
Visualizes an \code{mgnet} network object, emphasizing the signed properties of edges by
color-coding them and adjusting vertex sizes and edge widths for enhanced readability.
Positive edges are colored differently from negative edges, and vertex sizes can be dynamically
scaled based on their properties.
}
