% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-filter.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence}
\usage{
prevalence(X, mar = 2)
}
\arguments{
\item{X}{vector or matrix with all elements >=0.}

\item{mar}{a vector giving the subscripts which the function will be applied over; 1 indicates rows, 2 indicates columns, 0 for all elements.}
}
\description{
Return the portion of the elements greater than zero of a vector
or a matrix.
}
