% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pull.R
\name{pull_taxa}
\alias{pull_taxa}
\alias{pull_taxa,mgnet-method}
\alias{pull_taxa,mgnets-method}
\title{Pull Taxa Information from mgnet or mgnets Objects}
\usage{
pull_taxa(object, var = -1)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{var}{The name or position of the column to retrieve from the \code{taxa} data frame.
If -1 (default), the last column of the data frame is returned. You can specify the column name unquoted due to non-standard evaluation.}
}
\value{
For a single \code{mgnet} object, a vector containing the data from the specified column
of the \code{taxa} data frame is returned. For an \code{mgnets} object, a list of such vectors
is returned, each corresponding to one \code{mgnet} object in the list.
}
\description{
Retrieves specific taxonmgnet information from the \code{taxa} data frame within \code{mgnet} objects,
or each \code{mgnet} object within an \code{mgnets}. This function simplifies direct access to specific columns of interest
using dynamic column name handling.
}
\details{
This function supports dynamic evaluation of the column name using \code{rlang} for unquoted names, allowing more
flexible and intuitive usage within data manipulation workflows.
}
