% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-filter.R
\name{quantile_nozero}
\alias{quantile_nozero}
\title{Quantile Non Zero Elements}
\usage{
quantile_nozero(X, probs = 0.5, mar = 2, ...)
}
\arguments{
\item{X}{vector or matrix with all elements >=0.}

\item{probs}{numeric vector of probabilities with values in [0,1] with default
set to 0.5, thus the median values (see \code{\link{quantile}}).}

\item{mar}{a vector giving the subscripts which the function will be applied over; 1 indicates rows, 2 indicates columns, 0 for all elements.}

\item{...}{additional arguments of quantile.}
}
\description{
Return the choosen quantile of the elements greater than zero from a vector
or a matrix.
}
