% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{ranks}
\alias{ranks}
\alias{ranks,mgnet-method}
\alias{ranks,list-method}
\title{Get taxonomic ranks.}
\usage{
ranks(object)

\S4method{ranks}{mgnet}(object)

\S4method{ranks}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return taxonomy ranks as a character vector.
}
