% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{ranks}
\alias{ranks}
\alias{ranks,mg-method}
\alias{ranks,mg}
\alias{ranks,list-method}
\alias{ranks,list}
\title{Get taxonomic ranks.}
\usage{
ranks(object)

\S4method{ranks}{mg}(object)

\S4method{ranks}{list}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return taxonomy ranks as a character vector.
}
