% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4a_classes_refine.R
\name{refine_taxa}
\alias{refine_taxa}
\alias{refine_taxa,mgnet-method}
\alias{refine_taxa,mgnetList-method}
\title{Refines Taxa Selection in \code{mgnet} or \code{mgnetList} Objects}
\usage{
refine_taxa(object, ..., condition = "AND", trim = "yes")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{...}{Criteria for taxa selection: logical vectors, numeric indices, taxa names,
or functions returning such indices. These criteria determine which taxa are retained.}

\item{condition}{Logical condition to combine multiple criteria: "AND" (default) or "OR".
"AND" requires all criteria to be met, "OR" requires any criterion to be met.}

\item{trim}{A parameter to handle taxa not meeting criteria: "yes" to remove them,
"no" to set their abundance to zero, or "aggregate" to combine them into a 'merged' category.}
}
\value{
A modified \code{mgnet} object with only the selected taxa, or an \code{mgnetList} with
each contained \code{mgnet} object filtered accordingly.
}
\description{
This function refines the selection of taxa within an \code{mgnet} object or across
\code{mgnet} objects within an \code{mgnetList}. It supports the specification of criteria
for which taxa to retain, using logical vectors, numeric indices, taxa names,
or functions that dynamically specify taxa selection criteria.
}
\details{
This function provides a flexible approach to taxa refinement, allowing users to easily
include or exclude specific taxa based on custom logic. The \code{trim} parameter offers
additional control over how unselected taxa are treated, enabling tailored data cleaning
and preparation steps prior to further analysis.
}
\seealso{
\link{mgnet}, \link{mgnetList}
}
