% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_setters.R
\name{rel_abundance<-}
\alias{rel_abundance<-}
\alias{rel_abundance<-,mgnet-method}
\alias{rel_abundance<-,mgnetList-method}
\title{Set rel_abundance Data}
\usage{
rel_abundance(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new rel_abundance data to be set.}
}
\value{
The \code{mgnet} or \code{mgnetList} object with updated abundance data.
}
\description{
This setter function allows you to update the rel_abundance data for \code{mgnet} objects
and each \code{mgnet} object within a \code{mgnetList}. The rel_abundance data must be a numeric matrix
for \code{mgnet} objects. For \code{mgnetList} objects, the rel_abundance data should be a named list
of numeric matrices corresponding to each \code{mgnet} object within the list.
}
