% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{rela<-}
\alias{rela<-}
\alias{rela<-,mgnet-method}
\alias{rela<-,mgnets-method}
\title{Set rela Data}
\usage{
rela(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{value}{The new rela data to be set.}
}
\value{
The \code{mgnet} or \code{mgnets} object with updated abundance data.
}
\description{
This setter function allows you to update the rela data for \code{mgnet} objects
and each \code{mgnet} object within a \code{mgnets}. The rela data must be a numeric matrix
for \code{mgnet} objects. For \code{mgnets} objects, the rela data should be a named list
of numeric matrices corresponding to each \code{mgnet} object within the list.
}
