% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{relative}
\alias{relative}
\alias{relative,mg,missing-method}
\alias{relative,mg,missing}
\alias{relative,mg,character-method}
\alias{relative,mg,character}
\alias{relative,list,missing-method}
\alias{relative,list,missing}
\alias{relative,list,character-method}
\alias{relative,list,character}
\title{Get relative abundances.}
\usage{
relative(object,rank)

\S4method{relative}{mg,missing}(object)

\S4method{relative}{mg,character}(object, rank)

\S4method{relative}{list,missing}(object)

\S4method{relative}{list,character}(object, rank)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{rank}{(Optional)}
}
\description{
Retrieves the relative abundances of data, normalized by their depths at
the taxonomic rank choosen.
}
