% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7b_classes_manage_community.R
\name{remove_smaller_communities}
\alias{remove_smaller_communities}
\alias{remove_smaller_communities,mgnet-method}
\title{Remove or Isolate Smaller Communities}
\usage{
remove_smaller_communities(
  object,
  size,
  trim = TRUE,
  membership_format = "mgnet"
)
}
\arguments{
\item{object}{An object of class \code{igraph}, \code{mgnet}, or \code{mgnetList} representing the network.}

\item{size}{A positive integer indicating the minimum number of vertices required for a community to be retained.}

\item{trim}{A logical value indicating whether vertices from smaller communities should be removed (\code{TRUE})
or simply isolated within the network (\code{FALSE}).}

\item{membership_format}{A character string specifying the format of the output regarding isolated nodes.
\code{"mgnet"} will label isolated nodes as \code{"0"}, whereas \code{"igraph"} retains the default behavior
where isolated nodes are not specifically labeled. This parameter is particularly relevant when \code{trim} is \code{FALSE}.}
}
\value{
The function returns an object of the same class as the input (\code{object}) but with smaller communities either removed or isolated
based on the specified \code{size} and \code{trim} parameters. The output format of isolated nodes can be adjusted with the \code{membership_format} parameter.
}
\description{
The function \code{remove_smaller_comm} removes or isolates vertices from communities that are below a specified size threshold.
It allows further customization of the output by optionally labeling isolated nodes in a specified format.
}
\details{
This function processes an object (either \code{igraph}, \code{mgnet}, or \code{mgnetList}) to remove or isolate communities based on their size,
with additional control over the output format regarding how isolated nodes are labeled.
}
