% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_classes_reorder.R
\name{reorder_taxa}
\alias{reorder_taxa}
\alias{reorder_taxa,mgnet-method}
\alias{reorder_taxa,mgnetList-method}
\title{Reorder Taxa in mgnet Object}
\usage{
reorder_taxa(object, taxa_order)
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{taxa_order}{A numeric vector of indices, a character vector of taxa labels, or
a function that takes \code{object} as its input and returns a vector of new ordering indices.}
}
\value{
The \code{mgnet} object with reordered taxa.
}
\description{
This function reorders the taxa in an \code{mgnet} object according to a specified order.
The order can be specified using taxa indices, taxa identifiers, or a function that
returns the new order when applied to the object.
}
