% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{sample_info}
\alias{sample_info}
\alias{sample_info,mgnet-method}
\alias{sample_info,list-method}
\title{Get sample metadata variables.}
\usage{
sample_info(object)

\S4method{sample_info}{mgnet}(object)

\S4method{sample_info}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return sample metadata variables number as character vector.
}
