% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{sample_metadata-class}
\alias{sample_metadata-class}
\title{The S4 class for storing sample metadata}
\description{
Rows represent samples, while column indices represent experimental variables
that describe the samples.

\describe{
\item{.Data}{Slot containing \code{\link{data.frame}} with sample metadata.}
}
}
