% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-definitions-validators-constructors.R
\name{sample_metadata}
\alias{sample_metadata}
\title{User constructor for sample_metadata s4 class.}
\usage{
sample_metadata(meta)
}
\arguments{
\item{meta}{data.frame with experimental variables.}
}
\description{
User constructor to create an object belonging to formal s4 class sample_metadata
avoiding the new function.
}
