% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{sample_name}
\alias{sample_name}
\alias{sample_name,mgnet-method}
\alias{sample_name,list-method}
\title{Get samples names.}
\usage{
sample_name(object)

\S4method{sample_name}{mgnet}(object)

\S4method{sample_name}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return names of samples as character vector.
}
