% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{sample_name}
\alias{sample_name}
\alias{sample_name,mg-method}
\alias{sample_name,mg}
\title{Get samples names.}
\usage{
sample_name(object)

\S4method{sample_name}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return names of samples as character vector.
}
