% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{sample_sum}
\alias{sample_sum}
\alias{sample_sum,mgnet-method}
\alias{sample_sum,list-method}
\title{Get samples sum.}
\usage{
sample_sum(object)

\S4method{sample_sum}{mgnet}(object)

\S4method{sample_sum}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Retrieves the sum on each sample using the sample_sum variable in in
meta_sample slot if it's present (see \code{\link{save_sample_sum}}).
}
