% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{sample_sum}
\alias{sample_sum}
\alias{sample_sum,mgnet-method}
\alias{sample_sum,mgnetList-method}
\title{Calculate Sample Sum of Abundance Data in mgnet Objects}
\usage{
sample_sum(object, na.rm = FALSE)
}
\arguments{
\item{object}{An \code{mgnet} object containing an abundance matrix.}

\item{na.rm}{Logical indicating whether NA values should be removed before
summing the abundance values. Defaults to FALSE.}
}
\value{
A numeric vector with the sum of abundance values for each sample.
}
\description{
This function calculates the sum of abundance values for each sample within the
abundance matrix of an \code{mgnet} object. It returns a numeric vector where each element
corresponds to the sum of abundance values for a sample.
}
