% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{sampleID}
\alias{sampleID}
\alias{sampleID,mgnet-method}
\alias{sampleID,list-method}
\title{Get samples IDs.}
\usage{
sampleID(object)

\S4method{sampleID}{mgnet}(object)

\S4method{sampleID}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return names of samples as character vector.
}
