% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{save_geometric_mean}
\alias{save_geometric_mean}
\alias{save_geometric_mean,mgnet-method}
\alias{save_geometric_mean,list-method}
\title{Save sample geometric mean in mgnet.}
\usage{
save_geometric_mean(object)

\S4method{save_geometric_mean}{mgnet}(object)

\S4method{save_geometric_mean}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Store in mgnet object the variable geometric_mean in meta_sample in which are
saved the geometric mean for each sample. It is necessary to retrieves clr
transformed abundances of samples.
}
