% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{save_sample_sum}
\alias{save_sample_sum}
\alias{save_sample_sum,mgnet-method}
\alias{save_sample_sum,list-method}
\title{Save sample sum in mgnet.}
\usage{
save_sample_sum(object)

\S4method{save_sample_sum}{mgnet}(object)

\S4method{save_sample_sum}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Store in mgnet object the variable sample_sum in meta_sample in which are
saved the count sum for each sample. It's necessary to retrieves relative
abundances of samples.
}
