% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_select.R
\name{select_meta}
\alias{select_meta}
\alias{select_meta,mgnet-method}
\alias{select_meta,mgnetList-method}
\title{Modify and Update Sample Information in mgnet or mgnetList Objects}
\usage{
select_meta(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{...}{Columns to select from the \code{meta} data frame, specified unquoted due to
non-standard evaluation supported by \code{dplyr}.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with an updated \code{meta} data frame.
}
\description{
Modifies the \code{meta} data frame within \code{mgnet} objects by selecting specific columns
using the \code{select} function from the \code{dplyr} package, and updates the \code{meta} slot with
the new data frame. This method provides dynamic manipulation of sample metadata,
allowing selection of columns based on specific analysis needs.
}
\details{
Leverages the \code{select} function from \code{dplyr} for intuitive and flexible column selection,
ensuring that only relevant metadata are retained in the \code{meta} slot of the \code{mgnet} object.
}
