% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_classes_select.R
\name{select_taxa}
\alias{select_taxa}
\alias{select_taxa,mgnet-method}
\alias{select_taxa,mgnetList-method}
\title{Modify and Update taxa Information in mgnet or mgnetList Objects}
\usage{
select_taxa(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{...}{Columns to select from the \code{taxa} data frame, specified unquoted due to
non-standard evaluation supported by \code{dplyr}.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with an updated \code{taxa} data frame.
}
\description{
Modifies the \code{taxa} data frame within \code{mgnet} objects by selecting specific columns
using the \code{select} function from the \code{dplyr} package, and updates the \code{taxa} slot with
the new data frame. This method provides dynamic manipulation of taxa taxadata,
allowing selection of columns based on specific analysis needs.
}
\details{
Leverages the \code{select} function from \code{dplyr} for intuitive and flexible column selection,
ensuring that only relevant taxadata are retained in the \code{taxa} slot of the \code{mgnet} object.
}
