% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_sample}
\alias{selection_sample}
\alias{selection_sample,mgnet,vector-method}
\title{Selection a subset of samples from mgnet object.}
\usage{
selection_sample(object, idx)

\S4method{selection_sample}{mgnet,vector}(object, idx)
}
\arguments{
\item{object}{mgnet.}

\item{idx}{Vector of integer position indices or logical (like to
\code{[} extractor function).}
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which samples to keep.
}
