% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_taxa}
\alias{selection_taxa}
\alias{selection_taxa,mgnet-method}
\alias{selection_taxa,list-method}
\title{Selection a subset of taxa from mgnet object.}
\usage{
selection_taxa(object, ..., condition="AND",trim=TRUE)

\S4method{selection_taxa}{mgnet}(object, ..., condition = "AND", trim = TRUE)

\S4method{selection_taxa}{list}(object, ..., condition = "AND", trim = TRUE)
}
\arguments{
\item{object}{mgnet.}

\item{...}{arbitrary set of vectors containing the (numeric or name) or logical indices containing the samples to be preserved}

\item{condition}{(Optional default "AND") character indicates the logical operation between indices.}

\item{trim}{(Option default TRUE) logical, if true the discarded taxa are removed from obj otherwise they are only set to zero in data.}
=======
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{selection_taxa}
\alias{selection_taxa}
\alias{selection_taxa,mg,vector-method}
\alias{selection_taxa,mg,vector}
\alias{selection_taxa,list,vector-method}
\alias{selection_taxa,list,vector}
\title{Selection a subset of taxa from mg object.}
\usage{
selection_taxa(object, idx)

\S4method{selection_taxa}{mg,vector}(object, idx)

\S4method{selection_taxa}{list,vector}(object, idx)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{idx}{(Required) Vector of integer position indices or logical or names
of taxa (like to \code{[} extractor function).}
>>>>>>> main
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which taxa to keep.
}
