% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_vertices}
\alias{selection_vertices}
\alias{selection_vertices,mgnet,vector-method}
\alias{selection_vertices,mgnet,vector}
\alias{selection_vertices,list,vector-method}
\alias{selection_vertices,list,vector}
\title{Selection a subset of taxa from mgnet object.}
\usage{
selection_vertices(object, idx)

\S4method{selection_vertices}{mgnet,vector}(object, idx)

\S4method{selection_vertices}{list,vector}(object, idx)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{idx}{(Required) Vector of integer position indices or logical or names
of taxa (like to \code{[} extractor function).}
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which taxa to keep. The function make a subset of the mgnet
object cutting the columns of data and the rows of taxa, making a subset of
the network and creating a new communities object with only the conserved taxa.
}
