% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7a_classes_color.R
\name{set_categorical_color}
\alias{set_categorical_color}
\alias{set_categorical_color,mgnet-method}
\alias{set_categorical_color,mgnetList-method}
\title{Set Categorical Colors in \code{mgnet} or \code{mgnetList} Objects}
\usage{
set_categorical_color(
  object,
  ...,
  vars,
  decreasing = TRUE,
  mode = "merged",
  distinct_colors = 20,
  colorspace = "pretty",
  alpha = 1,
  extraColor = "#FFFFFF"
)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object containing the taxa data.}

\item{...}{The expressions to evaluate, which determine how colors are assigned within each grouping variable's context.
Each expression must be named as it dictates the resulting column name in the taxa metadata.}

\item{vars}{Character vector specifying one or more grouping variables by which taxa are categorized before coloring.
Each variable specified must be present in the taxa data.}

\item{decreasing}{Logical; whether to sort the categories within each group in decreasing order based on the expression evaluation.}

\item{distinct_colors}{Integer; the maximum number of distinct colors to generate for the categories. Defaults to 20.}

\item{colorspace}{Character or list; the color palette to use, which can be a predefined palette name or a list specifying
hue, saturation, and lightness ranges. Supported names are 'pretty', 'pretty_dark', 'rainbow', and 'pastels'.
For a detailed explanation of these options, see the \code{qualpalr} package documentation or enter \code{??qualpalr} in the R console.}

\item{alpha}{Numeric; transparency level of the colors, where 1 is opaque and 0 is fully transparent. Defaults to 1.}

\item{extraColor}{Character; the color used for any additional categories beyond \code{distinct_colors}. Defaults to white ("#FFFFFF").}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with added color columns in the taxa data.
}
\description{
This function assigns colors to taxa based on the results of specified expressions evaluated within the context
of defined grouping variables. It allows for the dynamic creation of color-coded categories for visual representation
in subsequent visualizations.
}
\details{
\code{set_categorical_color} dynamically generates colors based on the results of specified expressions, applying these colors
to the taxa metadata. This method is useful for preparing data for visualization, where colors can help in distinguishing
taxa based on categorical attributes derived from the data.

The function modifies the original \code{mgnet} or \code{mgnetList} object by adding new columns that represent color codes,
corresponding to each grouping variable and expression combination.

If any existing columns in the taxa data conflict with the newly generated color column names, they will be overwritten
with the new color data, and a warning will be issued.
}
\examples{
# Assuming `mg` is an `mgnet` object with taxa metadata containing `genus` and abundance data
mg <- set_categorical_color(mg, sum_abun = sum(abun), vars = "genus")
}
\seealso{
\code{\link[qualpalr]{qualpalr}} for details on generating qualitative color palettes.
}
