% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_classes_set.R
\name{set_CLR_abundance}
\alias{set_CLR_abundance}
\alias{set_CLR_abundance,mgnet-method}
\alias{set_CLR_abundance,mgnetList-method}
\title{Store clr-transformed data in \code{norm_abundance} Slot in \code{mgnet} Objects}
\usage{
set_CLR_abundance(object, clr_variant = "clr", zero_strategy = "unif")
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{clr_variant}{The method to calculate norm_abundance. Options are 'clr' for centered log-ratio and 'iclr' for
inter-quantile log-ratio. Default is 'clr'.}

\item{zero_strategy}{The zero replacement strategy before log-ratio transformation.
'const' replaces zeros with a small constant, and 'unif' replaces zeros with a uniform random value
between 6.5\% and 65\% of the detection limit. Default is 'unif'.}
}
\value{
The \code{mgnet} object with updated \code{norm_abundance} data.
}
\description{
Applies centered log-ratio (CLR) transformation or inter-quantile log-ratio (ICLR) transformation to the abundance data
in \code{mgnet} objects, after handling zeros with specified strategies. This transformation is useful for compositional data analysis,
making the data suitable for statistical modeling and comparison.
}
\seealso{
\code{\link{clr}}, \code{\link{iclr}}, \code{\link{zero_dealing}}
See also:
\code{\link[=zero_dealing]{zero_dealing}} for zero replacement strategies,
\code{\link[=clr]{clr}} for details on centered log-ratio transformation,
\code{\link[=iclr]{iclr}} for details on inter-quantile log-ratio transformation.
}
