% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_classes_set.R
\name{set_comm}
\alias{set_comm}
\alias{set_comm,mgnet-method}
\alias{set_comm,mgnetList-method}
\title{Set \code{comm} Slot in \code{mgnet} Objects}
\usage{
set_comm(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new community data to be set, a numeric matrix for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated community data.
}
\description{
This function sets the community data for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
