% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_classes_set.R
\name{set_info_sample}
\alias{set_info_sample}
\alias{set_info_sample,mgnet-method}
\alias{set_info_sample,mgnetList-method}
\title{Set \code{info_sample} Slot in \code{mgnet} Objects}
\usage{
set_info_sample(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new info_sample data to be set, a numeric matrix for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated info_sample data.
}
\description{
This function sets the info_sample data for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
