% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7c_classes_taxa_color.R
\name{set_lineage_color}
\alias{set_lineage_color}
\alias{set_lineage_color,mgnet-method}
\alias{set_lineage_color,mgnetList-method}
\title{Set Color Mapping to Top N Taxa in an \code{mgnet} Object}
\usage{
set_lineage_color(
  object,
  n,
  field,
  rank = NULL,
  order_fun = sum,
  decreasing = TRUE,
  extraColor = "#FFFFFF",
  alpha = 1,
  colorspace = "pretty",
  color_to
)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{n}{Integer; the number of top taxa to be distinctly colored.}

\item{field}{Character; specifies the data field (\code{"abundance"}, \code{"rel_abundance"}, or \code{"norm_abundance"}) used for ranking taxa.}

\item{rank}{Optional character string specifying the taxonomic rank for aggregation and coloring.}

\item{order_fun}{Function to compute rankings, defaulting to \code{sum}. Other possibilities include \code{mean}, \code{median} or costum functions.}

\item{decreasing}{Logical indicating if sorting should be in decreasing order. When \code{TRUE} (default),
samples with the highest metric values are considered top.}

\item{extraColor}{String; hexadecimal color code for taxa outside the top set. This color is applied to taxa that do not make
into the top \code{n} taxa across all \code{mgnet} objects in a \code{mgnetList} or within the single \code{mgnet} object. Default is "#FFFFFF" (white).}

\item{alpha}{Numeric; opacity level for colors, ranging from 0 (transparent) to 1 (opaque), with a default of 1.}

\item{colorspace}{Specifies the color space for generating the palette, with default 'pretty'. Options include 'pretty_dark', 'rainbow', 'pastels', or a custom list.}

\item{color_to}{Character; the name of the new column in \code{info_taxa} where color codes will be stored.}
}
\value{
The updated \code{mgnet} or \code{mgnetList} object with \code{info_taxa} including color mappings.
}
\description{
This function assigns distinctive color codes to the top \code{n} taxa based on an aggregate metric computed from a specified data field
in an \code{mgnet} object. For \code{mgnetList} objects, it applies this logic individually to each contained \code{mgnet} object and
aggregates a unique set of top \code{n} taxa across all objects to ensure distinctive coloring. Taxa not in the top set across the list
are assigned a default or specified \code{extraColor}.
}
\seealso{
\link[mgnet]{colormap_categories}, \link[mgnet]{top_taxa}
}
