% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_classes_set.R
\name{set_meta}
\alias{set_meta}
\alias{set_meta,mgnet-method}
\alias{set_meta,mgnetList-method}
\title{Set \code{meta} Slot in \code{mgnet} Objects}
\usage{
set_meta(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new sample metadata data to be set, a data.frame for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated sample data.
}
\description{
This function sets the sample metadata for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
