% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_classes_set.R
\name{set_netw}
\alias{set_netw}
\alias{set_netw,mgnet-method}
\alias{set_netw,mgnetList-method}
\title{Set \code{netw} Slot in \code{mgnet} Objects}
\usage{
set_netw(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new network data to be set, a numeric matrix for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated network data.
}
\description{
This function sets the network data for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
