% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_classes_set.R
\name{set_norm}
\alias{set_norm}
\alias{set_norm,mgnet-method}
\alias{set_norm,mgnetList-method}
\title{Set \code{norm} Slot in \code{mgnet} Objects}
\usage{
set_norm(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new norm data to be set, a numeric matrix for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated rela data.
}
\description{
This function sets the rela data for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
