% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_classes_set.R
\name{set_norm_abundance}
\alias{set_norm_abundance}
\alias{set_norm_abundance,mgnet-method}
\alias{set_norm_abundance,mgnetList-method}
\title{Set \code{norm_abundance} Slot in \code{mgnet} Objects}
\usage{
set_norm_abundance(
  object,
  value = NULL,
  clr_variant = NULL,
  zero_strategy = "unif"
)
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{value}{Optional. A numeric matrix to set as norm_abundance data directly.}

\item{clr_variant}{Optional. A string specifying the method to calculate norm_abundance ('clr' or 'iclr').}

\item{zero_strategy}{A string specifying the zero replacement strategy applied before log-ratio transformation.}
}
\value{
The \code{mgnet} object with updated norm_abundance data.
}
\description{
This function allows setting the norm_abundance data directly or calculating it
from the abundance data using specified methods ('clr' or 'iclr') after applying
zero replacement strategies ('unif' or 'const'). The zero replacement is applied
to the abundance data before the log-ratio transformation when 'method' is specified.
}
\seealso{
\code{\link{clr}}, \code{\link{iclr}}, \code{\link{zero_dealing}}
}
