% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_classes_set.R
\name{set_norm_CLR}
\alias{set_norm_CLR}
\alias{set_norm_CLR,mgnet-method}
\alias{set_norm_CLR,mgnetList-method}
\title{Store clr-transformed data in \code{norm} Slot in \code{mgnet} Objects}
\usage{
set_norm_CLR(object, clr_variant = "clr", zero_strategy = "const")
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{clr_variant}{The method to calculate norm. Options are 'clr' for centered log-ratio and 'iclr' for
inter-quantile log-ratio. Default is 'clr'.}

\item{zero_strategy}{The zero replacement strategy before log-ratio transformation.
'const' replaces zeros with a small constant, and 'unif' replaces zeros with a uniform random value
between 6.5\% and 65\% of the detection limit. Default is 'const'.}
}
\value{
The \code{mgnet} object with updated \code{norm} data.
}
\description{
Applies centered log-ratio (CLR) transformation or inter-quantile log-ratio (ICLR) transformation to the abundance data
in \code{mgnet} objects, after handling zeros with specified strategies. This transformation is useful for compositional data analysis,
making the data suitable for statistical modeling and comparison.
}
\seealso{
\code{\link{clr}}, \code{\link{iclr}}, \code{\link{zero_dealing}}
See also:
\code{\link[=zero_dealing]{zero_dealing}} for zero replacement strategies,
\code{\link[=clr]{clr}} for details on centered log-ratio transformation,
\code{\link[=iclr]{iclr}} for details on inter-quantile log-ratio transformation.
}
