% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_classes_set.R
\name{set_rel_abundance}
\alias{set_rel_abundance}
\alias{set_rel_abundance,mgnet-method}
\alias{set_rel_abundance,mgnetList-method}
\title{Set \code{rel_abundance} Slot in \code{mgnet} Objects}
\usage{
set_rel_abundance(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new rel_abundance data to be set, a numeric matrix for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated rel_abundance data.
}
\description{
This function sets the rel_abundance data for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
