% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_classes_set.R
\name{set_taxa}
\alias{set_taxa}
\alias{set_taxa,mgnet-method}
\alias{set_taxa,mgnetList-method}
\title{Set \code{taxa} Slot in \code{mgnet} Objects}
\usage{
set_taxa(object, value)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new taxa data to be set, a numeric matrix for \code{mgnet} objects
or a list of numeric matrices for \code{mgnetList} objects.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with the updated taxa data.
}
\description{
This function sets the taxa data for an \code{mgnet} object or each \code{mgnet} object
within an \code{mgnetList}.
}
