% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-filter.R
\name{shannon_diversity}
\alias{shannon_diversity}
\title{Shannon alpha diversity}
\usage{
shannon_diversity(X, mar = 2, norm = TRUE)
}
\arguments{
\item{X}{vector or matrix with all elements >=0.}

\item{mar}{a vector giving the subscripts which the function will be applied over; 1 indicates rows, 2 indicates columns, 0 for all elements.}

\item{norm}{logical elements indicates if the entropy must be normalize to 1 dividing
by the logarithm of the dimension (default TRUE).}
}
\description{
Returns the Shannon alpha diversity from the margins of a matrix.
}
