% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape.R
\name{split_arrange_merged_meta}
\alias{split_arrange_merged_meta}
\title{Split and Arrange Sample Info by mgnet}
\usage{
split_arrange_merged_meta(merged_meta, object)
}
\arguments{
\item{merged_meta}{A tibble containing the merged sample info with \code{mgnet} column.}

\item{object}{An \code{mgnetList} object to match the sample order.}
}
\value{
A list of tibbles with sample IDs as row names, one for each \code{mgnet} object.
}
\description{
This internal function splits the mutated sample info by \code{mgnet}, arranges
the samples to match their order in the corresponding \code{mgnet} object, and
returns the result with sample IDs as row names.
}
\keyword{internal}
