% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape.R
\name{split_arrange_merged_taxa}
\alias{split_arrange_merged_taxa}
\title{Split and Arrange Taxa Info by mgnet}
\usage{
split_arrange_merged_taxa(merged_taxa, object)
}
\arguments{
\item{merged_taxa}{A tibble containing the merged taxa info with \code{mgnet} column.}

\item{object}{An \code{mgnetList} object to match the sample order.}
}
\value{
A list of tibbles with taxa IDs as row names, one for each \code{mgnet} object.
}
\description{
This internal function splits the mutated taxa info by \code{mgnet}, arranges
the taxa to match their order in the corresponding \code{mgnet} object, and
returns the result with taxa IDs as row names.
}
\keyword{internal}
