% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-split.R
\name{split_mgnet}
\alias{split_mgnet}
\alias{split_meta}
\alias{split_meta,mgnet-method}
\alias{split_taxa}
\alias{split_taxa,mgnet-method}
\title{Split an \code{mgnet} object into multiple \code{mgnet} objects}
\usage{
split_meta(object, ...)

\S4method{split_meta}{mgnet}(object, ...)

split_taxa(object, ...)

\S4method{split_taxa}{mgnet}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mgnet}.}

\item{...}{One or more metadata columns used to define the split.}
}
\value{
An object of class \code{mgnets} containing one \code{mgnet} per group.
}
\description{
Split an \code{mgnet} object into multiple \code{mgnet} objects according to values of
selected metadata columns. The resulting objects are returned inside an
\code{mgnets} container.

\code{split_meta()} performs the split using \strong{sample-level metadata} returned
by \code{\link[=meta]{meta()}}, while \code{split_taxa()} performs the split using \strong{taxa-level
metadata} returned by \code{\link[=taxa]{taxa()}}.
}
\details{
Each unique combination of the selected columns defines a group. A new
\code{mgnet} object is created for each group.

The names of the resulting objects are constructed by concatenating the
values of the selected columns using \code{"-"}.

Group names must be unique.
}
\examples{
\dontrun{
split_meta(x, Country)
split_meta(x, Country, Surface)

split_taxa(x, Phylum)
split_taxa(x, Phylum, Genus)
}

}
