% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_classes_extractors.R
\name{[,mgnetList-method}
\alias{[,mgnetList-method}
\title{Extract Subsets from mgnetList}
\usage{
\S4method{[}{mgnetList}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{i}{Numeric indices or character vector of names specifying the \code{mgnet} objects to extract.
If missing, the entire \code{mgnetList} is returned.}

\item{j}{Ignored in this method, included for compatibility with the generic method.}

\item{...}{Additional arguments, currently ignored.}

\item{drop}{Ignored in this method, included for compatibility with the generic method.}
}
\value{
Depending on the type and number of indices or names provided in \code{i}:
- A single \code{mgnet} object if a single index or name is provided.
- An \code{mgnetList} containing the selected \code{mgnet} objects if multiple indices or names are provided.
- The full \code{mgnetList} if \code{i} is missing.
}
\description{
This method enables the extraction of sub-lists or individual \code{mgnet} objects from an \code{mgnetList}
container by specifying numeric indices or names. Use \code{[} to extract sub-lists or individual
objects and maintain the \code{mgnetList} structure.
}
\seealso{
\code{\link[base]{Extract}} for base R extraction operations.
}
