% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_classes_extractors.R
\name{[[,mgnetList-method}
\alias{[[,mgnetList-method}
\title{Extract a Single mgnet Object}
\usage{
\S4method{[[}{mgnetList}(x, i, ..., exact = TRUE)
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{i}{Index or name specifying the \code{mgnet} object to extract.}

\item{...}{Additional arguments (currently unused).}

\item{exact}{Logical indicating if exact matching is required (default is \code{TRUE}).}
}
\value{
The extracted \code{mgnet} object.
}
\description{
This method extracts a single \code{mgnet} object from an \code{mgnetList} by index or name.
}
