% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_extractors.R
\name{[[<-,mgnetList-method}
\alias{[[<-,mgnetList-method}
\title{Replace a Single mgnet Object}
\usage{
\S4method{[[}{mgnetList}(x, i) <- value
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{i}{Index or name specifying which \code{mgnet} object to replace.}

\item{value}{The new \code{mgnet} object to be inserted into the list.}
}
\value{
The \code{mgnetList} object with the specified \code{mgnet} object replaced.
}
\description{
This method replaces a single \code{mgnet} object within an \code{mgnetList} by index or name.
}
