% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subject_HMP2_netw}
\alias{subject_HMP2_netw}
\title{MGNet Object with Network Data for Subject HMP2}
\format{
An S4 object of class \code{mgnet} with:
\itemize{
\item \strong{abun}: abundance matrix for 376 taxa across 51 samples.
\item \strong{rela}: Associated relative abundances.
\item \strong{norm}: CLR transformed abundance matrix.
\item \strong{meta}: Sample metadata including SubjectID, DateZero, DateSubject, IRIS, etc.
\item \strong{taxa}: Detailed taxonomic information for each taxon (domain, phylum, class, order, etc.).
\item \strong{netw}: An \code{igraph} object representing microbial interactions with 2128 edges and a 3.02\% density.
\item \strong{comm}: Community detection results identifying 12 communities with detailed community sizes and 121 isolated nodes.
}
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(subject_HMP2_netw)
}
\description{
This \code{mgnet} object contains metagenomic data for a specific subject from the Human Microbiome Project Phase 2 (HMP2),
including normalized abundance data, sample metadata, taxonomic information, network interactions, and community
detection results. It is designed for comprehensive analysis of microbial communities within the gut microbiome of the subject.
}
\keyword{datasets}
