% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subjects_HMP2}
\alias{subjects_HMP2}
\title{Filtered Subject-specific MGNet List from HMP2 Dataset}
\format{
A \code{mgnetList} object containing 2 \code{mgnet} instances:
\describe{
\item{69-001}{An \code{mgnet} object with 185 samples and 977 taxa.}
\item{69-053}{An \code{mgnet} object with 40 samples and 793 taxa.}
}
}
\usage{
data(subjects_HMP2)
}
\description{
This \code{mgnetList} object contains filtered \code{mgnet} objects for two specific subjects
from the Human Microbiome Project Phase 2 (HMP2), focusing on their gut microbiome samples.
The dataset is specifically filtered to include only subjects "69-053" and "69-001"
and further processed to exclude taxa with zero abundance across samples.
}
\details{
Each \code{mgnet} object in the list represents comprehensive metagenomic data for one subject,
including abundance data, sample metadata, and taxonomic information. This structure is
ideal for detailed, subject-specific analysis of microbial communities. The list simplifies
comparative studies between subjects under similar experimental conditions.

The \code{mgnet} objects within the list do not contain network or community detection results,
which allows for custom analyses to be conducted by the user. This data can be used to explore
microbial diversity, abundance patterns, and potential interactions within and between the microbial
communities of the two subjects.
}
\keyword{datasets}
