% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subjects_HMP2_netw}
\alias{subjects_HMP2_netw}
\title{MGNet List with Network Data for Selected Subjects from HMP2 Dataset}
\format{
A \code{mgnetList} object containing 2 \code{mgnet} instances:
\describe{
\item{69-001}{An \code{mgnet} object with 185 samples, 377 taxa, a network of 3460 edges (~4.88\% density), and 7 detected communities (21.75\% isolated nodes).}
\item{69-053}{An \code{mgnet} object with 40 samples, 375 taxa, a network of 174 edges (~0.25\% density), and 18 detected communities (74.40\% isolated nodes).}
}
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(subjects_HMP2_netw)
}
\description{
This \code{mgnetList} object contains \code{mgnet} instances for two selected subjects from the Human Microbiome Project Phase 2 (HMP2),
focusing on their gut microbiome samples with network data and community analyses. The list facilitates comparative
metagenomic studies, highlighting interactions and community structures unique to each subject.
}
\details{
Each \code{mgnet} object within the list is structured to include normalized abundance data, comprehensive taxa and sample metadata,
networks with their respective density and detected communities indicating how many taxa are isolated or grouped. This setup provides
an in-depth look at the microbial ecosystem of each subject under study.
}
\keyword{datasets}
