% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa.HMP2}
\alias{taxa.HMP2}
\title{Samples metadata of taxa.HMP2}
\format{
\subsection{\code{taxa.HMP2}}{

A character matrix with 1953 rows/taxa and 7 columns/taxonomic rank.
It contains the taxonomy classification of the OTUs registered in otu.HMP2.
}
}
\source{
Jackson Laboratory
}
\usage{
taxa.HMP2
}
\description{
Samples metadata of taxa.HMP2
}
\keyword{datasets}
