% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{taxa_info}
\alias{taxa_info}
\alias{taxa_info,mgnet-method}
\alias{taxa_info,list-method}
\title{Get taxa additional info.}
\usage{
taxa_info(object)

\S4method{taxa_info}{mgnet}(object)

\S4method{taxa_info}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return taxa additional info as character vector.
}
