% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{taxa_vars}
\alias{taxa_vars}
\alias{taxa_vars,mgnet-method}
\alias{taxa_vars,mgnetList-method}
\title{Get Taxa Metadata Variables}
\usage{
taxa_vars(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, a character vector of metadata variable names.
For an \code{mgnetList} object, a named list of character vectors, with each list item representing
the metadata variable names in the corresponding \code{mgnet} objects.
}
\description{
Retrieves the names of metadata variables available in the \code{taxa} slot of an \code{mgnet} object,
or for each \code{mgnet} object within an \code{mgnetList}.
}
\examples{
data(HMP2, package = "mgnet")
taxa_vars(HMP2)  

data(subjects_HMP2, package = "mgnet")
taxa_vars(subjects_HMP2) 

}
