% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{taxaID}
\alias{taxaID}
\alias{taxaID,mgnet-method}
\alias{taxaID,list-method}
\title{Get taxa ID.}
\usage{
taxaID(object)

\S4method{taxaID}{mgnet}(object)

\S4method{taxaID}{list}(object)
}
\arguments{
\item{object}{mgnet or list.}
}
\description{
Return taxonomy id as a character vector.
}
