% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_group.R
\name{taxawise}
\alias{taxawise}
\alias{taxawise,mgnet-method}
\alias{taxawise,mgnetList-method}
\title{Set Taxa-wise Operation Mode for an mgnet Object}
\usage{
taxawise(object)

\S4method{taxawise}{mgnet}(object)

\S4method{taxawise}{mgnetList}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
The modified \code{mgnet} object or \code{mgnetList} with updated grouping attributes.
}
\description{
This function sets the operation mode of an \code{mgnet} or \code{mgnetList} object to be taxa-wise,
similar to dplyr's rowwise but specific to taxa elements.
}
\details{
\itemize{
\item For \code{mgnet} objects, sets the grouping variable to \code{taxa_id}.
\item For \code{mgnetList} objects, sets the grouping variable to \code{mgnet}, \code{taxa_id}.
}
}
