% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-definitions-validators-constructors.R
\docType{class}
\name{taxonomy_table-class}
\alias{taxonomy_table-class}
\title{An S4 class for storing taxonomic classification of taxa as a character
matrix.}
\description{
Row indices represent taxa, columns represent taxonomic ranks.
}
\details{
\describe{
\item{value}{Slot containing \code{\link{matrix}} class.}
}
}
