% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-definitions-validators-constructors.R
\name{taxonomy_table}
\alias{taxonomy_table}
\title{User constructor for taxonomy_table s4 class.}
\usage{
taxonomy_table(taxa)
}
\arguments{
\item{taxa}{character matrix with taxonomic classification.}
}
\description{
User constructor to create an object belonging to formal s4 class taxonomy_table
avoiding the new function.
}
