% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-coercion.R
\name{to_tbl_graph}
\alias{to_tbl_graph}
\alias{to_tbl_graph,mgnet-method}
\alias{to_tbl_graph,mgnets-method}
\title{Coerce mgnet / mgnets objects to a tidygraph tbl_graph}
\usage{
to_tbl_graph(object, selected = TRUE, ...)

\S4method{to_tbl_graph}{mgnet}(object, selected = TRUE, ...)

\S4method{to_tbl_graph}{mgnets}(object, selected = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{mgnet} or \code{mgnets}.}

\item{selected}{Logical or other selector passed to \code{netw()} to
determine which network (or subset of the network) should be used.}

\item{...}{Additional arguments passed to \code{tidygraph::as_tbl_graph()}.}
}
\value{
A \code{tidygraph::tbl_graph} object.
}
\description{
These methods convert \code{mgnet} and \code{mgnets} objects into
\code{tidygraph::tbl_graph} objects, ready to be used with
\code{ggraph::ggraph()}.
}
\details{
For a single \code{mgnet} object, the method wraps the internal
network (as returned by \code{netw()}) into a \code{tbl_graph}.

For a \code{mgnets} object (a collection of \code{mgnet} objects),
all internal graphs are combined via \code{igraph::disjoint_union()},
after adding a \code{mgnet} column to vertices and edges containing
the name (or index) of the originating element. This makes it easy to
facet plots by \code{mgnet} in \code{ggraph}.

The \code{tidygraph} package is listed in \code{Suggests}; an error is
raised at runtime if it is not installed.
}
