% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4h_classes_top.R
\name{top_sample}
\alias{top_sample}
\alias{top_sample,mgnet-method}
\alias{top_sample,mgnetList-method}
\title{Rank Sample Based on Custom Criteria in \code{mgnet} Objects}
\usage{
top_sample(object, expression, by = NULL, decreasing = TRUE)
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{expression}{A dplyr-compatible expression that defines the criteria for
sorting samples. This should involve only the abundance variables \code{abun}, \code{rela},
\code{norm} or grouping functions from \code{dplyr}. The expression should not reference
other data keys directly.}

\item{by}{Optional character string specifying the samples metadata column to group by
before applying the ranking expression. Default is \code{"sample_id"}.}

\item{decreasing}{Logical indicating whether sorting should be in decreasing order.
Default is \code{TRUE}.}
}
\value{
A vector containing the \code{by} column, sorted according to the specified
\code{expression}.
}
\description{
This function sorts and ranks samples based on a specified aggregation or calculation
applied to abundance-related variables (\code{abun}, \code{rela}, \code{norm}) or other grouping
functions. It is designed for \code{mgnet} objects to facilitate focused analyses on
grouped or aggregated data.
}
\details{
\code{top_sample} allows users to specify complex sorting and ranking expressions that
are applied within groups defined by the \code{by} parameter. The function checks if the
provided expression uses only allowed variables (\code{abun}, \code{rela}, \code{norm}) and grouping
functions like \code{n()}, ensuring the expression's appropriateness for the operation.

This function dynamically constructs a data frame from the \code{mgnet} object's slots,
performs the necessary transformations and aggregations, and then sorts the results.
If no \code{abun}, \code{rela}, or \code{norm} keys are needed for the expression, it simply sorts the
sample metadata. Otherwise, it performs left joins to include the necessary abundance data.
}
