% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{top_taxa}
\alias{top_taxa}
\alias{top_taxa,mgnet,character,character,function-method}
\alias{top_taxa,list,ANY,ANY,ANY-method}
\title{Get top taxa.}
\usage{
top_taxa(object, which, rank, fun)

\S4method{top_taxa}{mgnet,character,character,`function`}(object, which, rank, fun)

\S4method{top_taxa}{list,ANY,ANY,ANY}(object, which, rank, fun)
}
\arguments{
\item{object}{mgnet or list.}

\item{which}{the comparison can be made on absolute values, on relative abundances
or on clr transformed data (possible values are "abs","rel","clr").}

\item{rank}{taxonomic level choosen.}

\item{fun}{function choosen for the comparison. It must receve as input a vector
and return a number (mean, median, max, etc...).}
}
\description{
Sort the top taxa at taxonomic level and function choosen
}
