% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{ungroup_link}
\alias{ungroup_link}
\alias{ungroup_link,mgnet-method}
\alias{ungroup_link,mgnets-method}
\title{Clear link grouping}
\usage{
ungroup_link(object)

\S4method{ungroup_link}{mgnet}(object)

\S4method{ungroup_link}{mgnets}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
The same object, with no active link grouping.
}
\description{
Remove the \code{link_groups} attribute from an \code{mgnet} or \code{mgnets} object.
}
