% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_group.R
\name{ungroup_meta}
\alias{ungroup_meta}
\title{Ungroup Meta in an mgnet Object}
\usage{
ungroup_meta(object)
}
\arguments{
\item{object}{An \code{mgnet} object.}
}
\value{
The \code{mgnet} object without any grouping information.
}
\description{
Clears the grouping attributes from the mgnet object, effectively removing any existing meta grouping.
This action resets the context for operations in \code{mutate_meta}, \code{filter_meta},
making them operate without any grouping constraints.
}
