% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_group.R
\name{ungroup_taxa}
\alias{ungroup_taxa}
\title{Ungroup Taxa in an mgnet Object}
\usage{
ungroup_taxa(object)
}
\arguments{
\item{object}{An \code{mgnet} object.}
}
\value{
The \code{mgnet} object without any grouping information.
}
\description{
Clears the grouping attributes from the mgnet object, effectively removing any existing taxa grouping.
This action resets the context for operations in \code{mutate_taxa}, \code{filter_taxa}, and \code{mutate_netw},
making them operate without any grouping constraints.
}
