% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_classes_unite.R
\name{unite_taxa}
\alias{unite_taxa}
\alias{unite_taxa,mgnet-method}
\alias{unite_taxa,mgnetList-method}
\title{Unite Multiple Taxa Metadata Columns into One in \code{mgnet} or \code{mgnetList} Objects}
\usage{
unite_taxa(object, col, ..., sep = "_", remove = TRUE)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object containing the metadata in which columns are to be united.}

\item{col}{Character string specifying the name of the new column that will be created by uniting existing columns.}

\item{...}{Columns to be united into the new column. These should be specified unquoted and separated by commas.}

\item{sep}{A string specifying the separator to be used to concatenate the values of the columns being united.
Default is an underscore ("_").}

\item{remove}{A logical value indicating whether the original columns should be removed after uniting.
Default is \code{TRUE}.}
}
\value{
Returns the \code{mgnet} or \code{mgnetList} object with updated metadata where specified columns are united into
a new single column. For \code{mgnetList}, the operation is performed for each \code{mgnet} object within the list.
}
\description{
Combines multiple columns from the metadata (\code{taxa}) of \code{mgnet} or \code{mgnetList} objects into a single column,
creating a new column by concatenating the contents of specified columns separated by a specified separator.
}
\details{
The \code{unite_taxa} function is a wrapper around \code{tidyr::unite} tailored for handling the metadata of \code{mgnet} objects.
It allows for the consolidation of information spread across multiple columns into a single column, which can be useful
for creating unique identifiers or combining categorical variables.

For \code{mgnetList} objects, the function applies the unite operation to each \code{mgnet} object contained within the list,
ensuring that all metadata across the list is consistently processed.
}
\seealso{
\code{\link[tidyr::unite]{tidyr}} for more details on the underlying unite function.
\code{\link[mgnet]{mgnet}} and \code{\link[mgnet]{mgnetList}} for details on the data structures.
}
