% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{validate_required_groups}
\alias{validate_required_groups}
\title{Validate Required Groups Variables in Expressions}
\usage{
validate_required_groups(object, groups_required, sample_or_taxa)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{groups_required}{A character vector of variable names.}

\item{sample_or_taxa}{A character string, either "sample" or "taxa", indicating the context.}
}
\value{
the controlled grouping variables
}
\description{
This internal function checks that all groups variables required in the object
(either in abundance data or metadata).
}
\keyword{internal}
