% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2e_classes_setters.R
\name{abun<-}
\alias{abun<-}
\alias{abun<-,mgnet-method}
\alias{abun<-,mgnetList-method}
\title{Set Abundance Data}
\usage{
abun(object) <- value
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{value}{The new abundance data to be set.}
}
\value{
The \code{mgnet} or \code{mgnetList} object with updated abundance data.
}
\description{
This setter function allows you to update the abundance data for \code{mgnet} objects
and each \code{mgnet} object within a \code{mgnetList}. The abundance data must be a numeric matrix
for \code{mgnet} objects. For \code{mgnetList} objects, the abundance data should be a named list
of numeric matrices corresponding to each \code{mgnet} object within the list.
}
