% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{abundance}
\alias{abundance}
\alias{abundance,mg,missing-method}
\alias{abundance,mg-method,missing}
\alias{abundance,mg,character-method}
\alias{abundance,mg-method,character}
\title{Get abundances at choosen rank.}
\usage{
abundance(object)

\S4method{abundance}{mg,missing}(object)

\S4method{abundance}{mg,character}(object, rank)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Retrieves the abundances of data at choosen taxonomy rank. Abundance at
higher rank are returned as sums of abundance of elements with the same
classification.
}
