% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency-tools.R
\name{adjacency_edge_density}
\alias{adjacency_edge_density}
\title{Get adjacency making an edge density threshold}
\usage{
adjacency_edge_density(x = NULL, r = NULL, method = "pearson", th = 0.05)
}
\arguments{
\item{x}{matrix of data.}

\item{r}{correlation matrix.}

\item{method}{(default "pearson") correlation type, possible choices are "pearson",
"spearman" or "kendall" (default spearman).}

\item{th}{(default 0.05) edge density choosen.}
}
\description{
Retrieves the adjacency matrix starting from the abundance
or the correlation matrix using as criteria the edge density of the resulting
network. The function uses the \code{\link{cor}} from stats package.
}
