% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_adjacency.R
\name{adjacency_p_adjust}
\alias{adjacency_p_adjust}
\title{Get Adjacency with Corrected P-values}
\usage{
adjacency_p_adjust(
  x = NULL,
  r = NULL,
  n = NULL,
  method = "pearson",
  adjust = "holm",
  alpha = 0.05
)
}
\arguments{
\item{x}{A numeric matrix of data, where each row represents a sample and each column represents a variable/taxon.
This parameter is required if \code{r} and \code{n} are not provided.}

\item{r}{A pre-computed correlation matrix. If provided, \code{n} must also be specified.}

\item{n}{The number of observations used to compute the correlation matrix. Required if \code{r} is provided.}

\item{method}{Character string specifying the correlation method ("pearson", "spearman", "kendall").
Default is "pearson".}

\item{adjust}{Character string specifying the method for p-value adjustment ("holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none"). Default is "holm".}

\item{alpha}{Numeric value specifying the significance level for the adjusted p-values. Default is 0.05.}
}
\value{
An adjacency matrix derived from the correlation matrix, where elements are set to zero
if their corresponding adjusted p-value is greater than \code{alpha}.
}
\description{
Generates an adjacency matrix from abundance data or a pre-computed correlation matrix,
applying a threshold on the correlation matrix based on adjusted p-values with multiple
test corrections. This function is useful for constructing networks from correlation matrices
by considering statistically significant correlations.
}
