% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Rearrange data in higher taxonomic level}
\usage{
aggregate_taxa(obj, rank)
}
\arguments{
\item{obj}{mgnet object}

\item{rank}{taxonomic level choosen}
}
\description{
A short description...
}
