% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities.R
\name{are_list_assign}
\alias{are_list_assign}
\title{Check list in assign methods for mgnetList}
\usage{
are_list_assign(object, value)
}
\arguments{
\item{object}{mgnetList object.}

\item{value}{List intended to be assigned to the mgnetList.}
}
\description{
Internal function to check if the list provided as a value in assign methods
is coherent with the mgnetList object.
}
\keyword{internal}
