% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{are_selected_links}
\alias{are_selected_links}
\alias{are_selected_links,mgnet-method}
\alias{are_selected_links,mgnets-method}
\title{Check exist a link selection in an \code{mgnet} Object}
\usage{
are_selected_links(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
logical.
}
\description{
Retrieves TRUE/FALSE if links are previously selected by \code{\link[=select_link]{select_link()}}.
}
