% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-communities.R
\name{as_igraph_communities}
\alias{as_igraph_communities}
\title{Convert mgnet into igraph communities.}
\usage{
as_igraph_communities(comm)
}
\arguments{
\item{comm}{object belong to communities class.}
}
\description{
The mgnet communities treat the isolated nodes differently respect
igraph putting all of them in the community labeled as '0'. Contrary igraph
describes the isolated nodes as community of size 1. In conclusion the function
modifies the the labels of the isolated nodes setting increasing number.
}
