% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{comm<-}
\alias{comm<-}
\alias{comm<-,mgnet,communities-method}
\title{Assign a new comm to \code{object}}
\usage{
comm(object) <- value

\S4method{comm}{mgnet,communities}(object) <- value
}
\arguments{
\item{object}{mgnet.}

\item{value}{\code{\link{communities}}}
}
\description{
Assign a new comm to \code{object}
}
