% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{data<-}
\alias{data<-}
\alias{data<-,mgnet,matrix-method}
\title{Assign a new ngs matrix to \code{object}}
\usage{
data(object) <- value

\S4method{data}{mgnet,matrix}(object) <- value
}
\arguments{
\item{object}{mgnet.}

\item{value}{matrix}
}
\description{
Assign a new ngs matrix to \code{object}
}
