% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{meta<-}
\alias{meta<-}
\alias{assign-meta}
\alias{meta<-,mg,data.frame-method}
\alias{meta<-,mg,data.frame}
\title{Assign new samples metadata to \code{object}}
\usage{
meta(object) <- value

\S4method{meta}{mg,data.frame}(object) <- value
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{value}{(Required) \code{\link{data.frame}}}
}
\description{
Assign new samples metadata to \code{object}
}
