% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{netw<-}
\alias{netw<-}
\alias{netw<-,mgnet-method}
\title{Assign a new netw to \code{object}}
\usage{
netw(object) <- value

\S4method{netw}{mgnet}(object) <- value
}
\arguments{
\item{object}{mgnet.}

\item{value}{\code{\link{igraph}} or the equivalent adjacency matrix.}
}
\description{
Assign a new netw to \code{object}
}
