% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{netw<-}
\alias{netw<-}
\alias{netw<-,mg,igraph-method}
\alias{netw<-,mgnet,igraph}
\title{Assign a new netw to \code{object}}
\usage{
netw(object) <- value

\S4method{netw}{mg,igraph}(object) <- value
}
\arguments{
\item{object}{(Required) \code{\link{mgnet-class}}.}

\item{value}{(Required) \code{\link{igraph}}}
}
\description{
Assign a new netw to \code{object}
}
