% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors.R
\docType{methods}
\name{data<-}
\alias{data<-}
\alias{assign-ngs_data}
\alias{data<-,ngs_data,ngs_dataOrMatrixOrNULL-method}
\alias{data<-,ngs_data,ngs_data-method}
\alias{data<-,mg,ngs_dataOrMatrixOrNULL-method}
\alias{data<-,mg,ngs_data-method}
\title{Assign a new ngs matrix to \code{object}}
\usage{
data(object) <- value

\S4method{data}{ngs_data,ngs_dataOrMatrixOrNULL}(object) <- value

\S4method{data}{mg,ngs_dataOrMatrixOrNULL}(object) <- value
}
\arguments{
\item{object}{Required \code{\link{mg-class}} or \code{\link{ngs_data-class}}.}

\item{value}{Required \code{\link{ngs_data-class}} or \code{\link{matrix}}}
}
\description{
Assign a new ngs matrix to \code{object}
}
