% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-show.R
\docType{methods}
\name{meta<-}
\alias{meta<-}
\alias{assign-sample_metadata}
\alias{meta<-,mg,sample_metadata-method}
\alias{meta<-,mg,sample_metadata-method,sample_metadata}
\alias{meta<-,mg,data.frame-method}
\alias{meta<-,mg,sample_metadata-method,data.frame}
\title{Assign new samples metadata to \code{object}}
\usage{
meta(object) <- value

\S4method{meta}{mg,sample_metadata}(object) <- value

\S4method{meta}{mg,data.frame}(object) <- value
}
\arguments{
\item{object}{Required \code{\link{mg-class}}.}

\item{value}{Required \code{\link{sample_metadata-class}} or \code{\link{data.frame}}}
}
\description{
Assign new samples metadata to \code{object}
}
