% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R, R/class-mgnet.R
\docType{methods}
\name{taxa<-}
\alias{taxa<-}
\alias{assign-taxa}
\alias{taxa<-,mg,matrix-method}
\alias{taxa<-,mg,matrix}
\alias{taxa<-,mgnet,matrix-method}
\alias{taxa<-,mgnet,matrix}
\title{Assign new taxonomy table to \code{object}}
\usage{
taxa(object) <- value

\S4method{taxa}{mg,matrix}(object) <- value

\S4method{taxa}{mgnet,matrix}(object) <- value
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{value}{(Required) \code{\link{matrix}}}
}
\description{
Assign new taxonomy table to \code{object}
}
