% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{taxa<-}
\alias{taxa<-}
\alias{taxa<-,mgnet,matrix-method}
\title{Assign new taxonomy table to \code{object}}
\usage{
taxa(object) <- value

\S4method{taxa}{mgnet,matrix}(object) <- value
}
\arguments{
\item{object}{mgnet}

\item{value}{matrix}
}
\description{
Assign new taxonomy table to \code{object}
}
