% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-show.R
\docType{methods}
\name{taxa<-}
\alias{taxa<-}
\alias{assign-taxonomy_table}
\alias{taxa<-,mg,taxonomy_table-method}
\alias{taxa<-,mg,taxonomy_table-method,taxonomy_table}
\alias{taxa<-,mg,matrix-method}
\alias{taxa<-,mg,taxonomy_table-method,matrix}
\title{Assign new taxonomy table to \code{object}}
\usage{
taxa(object) <- value

\S4method{taxa}{mg,taxonomy_table}(object) <- value

\S4method{taxa}{mg,matrix}(object) <- value
}
\arguments{
\item{object}{Required \code{\link{mg-class}}.}

\item{value}{Required \code{\link{taxonomy_table-class}} or \code{\link{data.frame}}}
}
\description{
Assign new taxonomy table to \code{object}
}
