% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_forbidden_expressions}
\alias{check_forbidden_expressions}
\title{Check for Forbidden Functions and Disallowed Variables in Expressions}
\usage{
check_forbidden_expressions(expressions)
}
\arguments{
\item{expressions}{A list of expressions captured by \code{rlang::enquos(...)}.}
}
\value{
Throws an error if any expression contains both a forbidden function and a disallowed variable; otherwise, nothing is returned.
}
\description{
This internal function checks each expression to ensure that forbidden functions
and disallowed variables do not coexist in any expression. If they do, an error is thrown.
}
\keyword{internal}
