% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_CLR_transformations.R
\name{clr}
\alias{clr}
\title{Centered Log-Ratio Transformation}
\usage{
clr(X)
}
\arguments{
\item{X}{A numeric matrix with all elements greater than 0, where rows are samples
and columns are components or taxa.}
}
\value{
A numeric matrix of the same dimension as \code{X}, representing the clr-transformed data.
}
\description{
Performs the Centered Log-Ratio (CLR) transformation on a given dataset (or at least a single vector).
This transformation is a standard approach for analyzing compositional data,
which consists of proportions or percentages summing up to a constant sum.
It transforms the data into a space where standard statistical methods can be applied.
Rows represent samples, and columns represent components or taxa.
}
