% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6a_classes_cluster_signed.R
\name{cluster_signed}
\alias{cluster_signed}
\alias{cluster_signed,igraph-method}
\alias{cluster_signed,mgnet-method}
\alias{cluster_signed,mgnetList-method}
\title{Community Detection in Signed Weighted Graphs}
\usage{
cluster_signed(
  object,
  resistance = 0,
  penalty = 1,
  add.names = FALSE,
  cores = 1
)
}
\arguments{
\item{object}{An \code{igraph} object representing a weighted undirected network or
an \code{mgnet} object containing such a network or an \code{mgnetList}.}

\item{resistance}{(Default: 0) Node resistance to joining communities,
specified as a positive or negative real number. This parameter simulates
a common self-loop resistance across nodes, with the default indicating no
resistance.}

\item{penalty}{(Default: 1) Specifies the relative importance of
the null-case term as a non-negative real number. This coefficient affects
how strongly the algorithm penalizes or favors the formation of communities.}

\item{add.names}{(Default: FALSE) Determines whether vertex names (if available)
are used in the output, associating each node with its community assignment.
When FALSE, community assignments are returned as numerical indices only.}

\item{cores}{Number of cores to use for parallel processing; default is 1 (no parallel execution).
This parameter is only applicable for \code{mgnetList} objects and is ignored for \code{mgnet} objects.}
}
\value{
An object of class \code{communities} as defined in the \code{igraph} package.
This object contains community assignments for each node in the input graph.
Isolated nodes are grouped into a special community labeled '0', rather than
being considered as separate communities of size one.
}
\description{
Implements a wrapper for community detection in signed weighted graphs,
adapting the algorithm developed by Sergio Gomez. This method is designed
to identify dense subgraphs, known as communities, by optimizing a signed
modularity score. The algorithm accommodates weighted undirected graphs,
applying a threshold to identify the most significant correlations as edges.
}
\details{
The function is an adaptation of Sergio Gomez's Radatools community detection
algorithm for signed networks. It identifies community structures by optimizing
a signed version of the modularity score, accounting for the positive and
negative weights of edges. The algorithm is particularly suited for analyzing
networks where relationships can have both positive (affiliative) and negative
(antagonistic) connotations.
}
\references{
Gomez, S. et al. (2009). Analysis of community structure in networks of correlated data. Phys. Rev. E.
}
\seealso{
\code{\link[igraph]{communities}}, for information on handling community objects in \code{igraph}.
}
