% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_color.R
\name{colormap_communities}
\alias{colormap_communities}
\title{Create Color Palette for Communities}
\usage{
colormap_communities(
  n = 20,
  alpha = 1,
  colorspace = "pretty",
  isolated_color = "#FFFFFF"
)
}
\arguments{
\item{n}{A positive integer indicating the number of distinct colors to generate, not exceeding 99.}

\item{alpha}{A numeric value specifying the transparency level of the colors, where 1 is opaque
and 0 is fully transparent.}

\item{colorspace}{Defines the color space for color generation. It can be one of the predefined
character options ('pretty', 'pretty_dark', 'rainbow', 'pastels') or a list specifying
the 'h', 's', and 'l' elements for custom color spaces. For a detailed explanation of
these options, consult the \code{qualpalr} package documentation or enter \code{??qualpal} in the R console.}

\item{isolated_color}{The color assigned to isolated nodes, represented by community '0'.
Defaults to white ("#FFFFFF"). Accepts hexadecimal color codes.}
}
\value{
A named vector of colors in hexadecimal format, with names from '0' to 'n', where '0'
represents isolated nodes. The color for isolated nodes can be customized with the
\code{isolated_color} parameter.
}
\description{
Generates a color palette with a specified number of distinct colors for visualizing communities.
The first color, customizable via the \code{isolated_color} parameter, is designated for isolated nodes,
with subsequent colors generated for other communities. This function uses the \code{qualpalr} package
to create visually distinct colors.
}
