% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utility.R
\name{colormap_communities}
\alias{colormap_communities}
\title{Create color palette for communities}
\usage{
<<<<<<< HEAD
colormap_communities(
  n = 20,
  alpha = 1,
  colorspace = list(h = c(0, 360), s = c(0.25, 1), l = c(0.25, 0.75))
)
}
\arguments{
\item{n}{positive integer indicated the number of distinct color, n<100.}

\item{alpha}{(Optional) Level of trasparency of colors.}

\item{colorspace}{(Optional) Permits to set the color ranges. For more details
see \code{\link{qualpal}}.}
=======
colormap_communities(n = 20)
}
\arguments{
\item{n}{positive integer indicated the number of distinct color.}
>>>>>>> main
}
\description{
User wrapper for distinctColorPalette function of randomcoloR
package. Create n+1 distinct color to associate with communities ID. The
first color labeled as 0 is always white for isolated nodes.
}
