% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-color.R
\name{colormap_communities}
\alias{colormap_communities}
\title{Create color palette for communities}
\usage{
colormap_communities(n = 20, alpha = 1, colorspace = "pretty")
}
\arguments{
\item{n}{positive integer indicated the number of distinct color, n<100.}

\item{alpha}{(Optional) Level of trasparency of colors.}

\item{colorspace}{(Optional) Permits to set the color ranges. For more details
see \code{\link{qualpal}}.}
}
\description{
User wrapper for qualpalr function of qualpalr package.
Create n+1 distinct color to associate with communities ID. The
first color labeled as 0 is always white for isolated nodes.
}
