% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utility.R
\name{colormap_communities}
\alias{colormap_communities}
\title{Create color palette for communities}
\usage{
colormap_communities(n = 20)
}
\arguments{
\item{n}{positive integer indicated the number of distinct color.}
}
\description{
User wrapper for distinctColorPalette function of randomcoloR
package. Create n+1 distinct color to associate with communities ID. The
first color labeled as 0 is always white for isolated nodes.
}
