% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_color.R
\name{colormap_taxonomy}
\alias{colormap_taxonomy}
\title{Create Color Palette for Taxonomy}
\usage{
colormap_taxonomy(
  taxaID,
  distinctColor = 20,
  extraColor = "#BFBFBF",
  alpha = 1,
  colorspace = "pretty"
)
}
\arguments{
\item{taxaID}{A character vector of taxonomic identifiers.}

\item{distinctColor}{An integer specifying the maximum number of distinct colors to generate,
ranging from 1 to 99.}

\item{extraColor}{The color used for taxa identifiers exceeding the \code{distinctColor} limit,
defaults to light gray. Accepts hexadecimal color codes.}

\item{alpha}{A numeric value specifying the transparency level of the colors.}

\item{colorspace}{Defines the color space for color generation. It can be one of the predefined
character options ('pretty', 'pretty_dark', 'rainbow', 'pastels') or a list specifying
the 'h', 's', and 'l' elements for custom color spaces. For a detailed explanation of
these options, see the \code{qualpalr} package documentation or enter \code{??qualpal} in the R console.}
}
\value{
A named vector of hexadecimal color codes, with names corresponding to the taxa IDs.
}
\description{
Generates a distinct color palette for taxonomic classifications. Each unique taxon is assigned
a color up to a specified limit (\code{distinctColor}). Beyond this limit, \code{extraColor} is used for
the remaining taxa. Duplicate taxa identifiers are filtered out to ensure uniqueness.
}
