% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utility.R
\name{colormap_taxonomy}
\alias{colormap_taxonomy}
\title{Create color palette for taxonomy}
\usage{
colormap_taxonomy(taxaID)
}
\arguments{
\item{taxaID}{character vector with all taxonomic classification.}
}
\description{
User wrapper for distinctColorPalette function of randomcoloR
package. Create named vector with taxaID as name and colors as values. If are
present duplicated in the vector the function choice only unique values.
}
