% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utility.R
\name{colormap_taxonomy}
\alias{colormap_taxonomy}
\title{Create color palette for taxonomy}
\usage{
<<<<<<< HEAD
colormap_taxonomy(
  taxaID,
  alpha = 1,
  colorspace = list(h = c(0, 360), s = c(0.25, 1), l = c(0.5, 0.9))
)
}
\arguments{
\item{taxaID}{character vector with all taxonomic classification.}

\item{alpha}{(Optional) Level of trasparency of colors.}

\item{colorspace}{(Optional) Permits to set the color ranges. For more details
see \code{\link{qualpal}}.}
=======
colormap_taxonomy(taxaID)
}
\arguments{
\item{taxaID}{character vector with all taxonomic classification.}
>>>>>>> main
}
\description{
User wrapper for distinctColorPalette function of randomcoloR
package. Create named vector with taxaID as name and colors as values. If are
present duplicated in the vector the function choice only unique values.
}
