% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{comm}
\alias{comm}
\alias{comm,mgnet-method}
\alias{comm,mgnets-method}
\title{Get Community Detection Results}
\usage{
comm(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
The community detection result object stored in the \code{comm} slot for an \code{mgnet} object,
or a list of such objects for an \code{mgnets} object, each representing the community
detection results of a contained \code{mgnet} object.
}
\description{
Retrieves the community detection results stored in the \code{comm} slot of an \code{mgnet} object
or each \code{mgnet} object within an \code{mgnets}. These results are typically derived from network
analysis methods and indicate the grouping or clustering of taxa into communities based on their
network interactions.
}
