% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{comm_id}
\alias{comm_id}
\alias{comm_id,mgnet-method}
\alias{comm_id,mgnetList-method}
\title{Retrieve Community IDs from mgnet or mgnetList Objects}
\usage{
comm_id(object, .fmt = "list")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{.fmt}{The format of the output. Possible values are:
\itemize{
\item \code{"list"}: Returns a named list where each element is a character vector of community IDs.
\item \code{"tbl"}: Returns a tibble with columns 'mgnet', 'taxa_id', and 'comm_id', representing the
source \code{mgnet} object, taxa IDs, and their respective community IDs.
}
The default is \code{"list"}.}
}
\description{
This function retrieves community IDs from an \code{mgnet} or \code{mgnetList} object, providing a flexible output
format. For \code{mgnetList}, the community IDs can be output as either a list or a structured tibble.
}
