% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{commID}
\alias{commID}
\alias{commID,mgnet-method}
\alias{commID,list-method}
\title{Get communities ID.}
\usage{
commID(object)

\S4method{commID}{mgnet}(object)

\S4method{commID}{list}(object)
}
\arguments{
\item{object}{mgnet-class.}
}
\description{
Return vector of character with the community ID for each vertex.
}
