% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_constructCorrCLRNet.R
\name{constructCorrCLRNet}
\alias{constructCorrCLRNet}
\alias{constructCorrCLRNet,mgnet-method}
\alias{constructCorrCLRNet,mgnetList-method}
\title{Construct Correlation Network from CLR-Transformed Data}
\usage{
constructCorrCLRNet(
  object,
  zero_strategy = "const",
  clr_method = NULL,
  cor_method = "pearson",
  thresh_method = NULL,
  thresh_value = NULL,
  padj_method = NULL,
  cores = 1
)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object containing abundance data.}

\item{zero_strategy}{Strategy for handling zeros in the abundance data before CLR transformation.
Options are "unif" for replacing zeros with values from a uniform distribution and "const"
for a constant value replacement. Default is "const".}

\item{clr_method}{Specifies the CLR transformation method to be used. The default behavior
is to check for pre-computed normalized data stored in the \code{norm} slot of the \code{mgnet} object.
If normalized data is available, it uses this stored data unless another method is explicitly specified.
Available methods are:
\itemize{
\item \code{"stored"}: Use pre-computed normalized data from the \code{norm} slot. If no normalized data
is present, an error is raised unless another method is specified.
\item \code{"clr"}: Perform standard CLR (Centered Log Ratio) transformation, suitable for most
compositional data analysis scenarios. This will be used to calculate CLR from the raw abundance
data if the \code{norm} slot is empty and no method is specified.
\item \code{"iclr"}: Perform an interquartile CLR transformation, which can be more robust to outliers.
This method is selected only if explicitly specified.
The default setting automatically uses the \code{norm} slot if available; otherwise, it defaults to
calculating standard CLR data.
}}

\item{cor_method}{Correlation method ("pearson", "spearman", "kendall") used for computing
pairwise correlations between taxa after CLR transformation. Default is "pearson".}

\item{thresh_method}{Method for thresholding the correlation matrix to construct the adjacency
matrix: "absolute" applies an absolute value threshold, "density" targets a specific edge
density, and "p-value" uses p-value adjustment to threshold correlations.}

\item{thresh_value}{Numeric threshold applied according to \code{thresh_method}. For "absolute" and
"p-value", this represents the minimum correlation or adjusted p-value for inclusion in the
network. For "density", this represents the desired edge density.}

\item{padj_method}{Method for adjusting p-values when \code{thresh_method} is "p-value". Options
include "holm", "hochberg", "hommel", "bonferroni", "BH" (Benjamini-Hochberg), "BY"
(Benjamini-Yekutieli), "fdr" (false discovery rate), and "none" (no adjustment). Default is "none".}

\item{cores}{Number of cores to use for parallel processing; default is 1 (no parallel execution).
This parameter is only applicable for \code{mgnetList} objects and is ignored for \code{mgnet} objects.}
}
\value{
The input \code{mgnet} or \code{mgnetList} object with an updated \code{netw} slot containing the
constructed correlation network. The network is represented as an \code{igraph} object for \code{mgnet}
and a list of \code{igraph} objects for \code{mgnetList}.
}
\description{
Constructs a correlation network from compositional data using centered log-ratio (CLR)
transformation followed by correlation analysis. This function is designed to handle
compositional data's inherent constraints by applying CLR transformation, facilitating
meaningful correlation analysis. Users can choose from various zero replacement strategies,
CLR methods, and thresholding techniques to tailor the network construction process to their
specific data and research questions.
}
\details{
The function operates in several steps to transform the abundance data into a correlation
network:
\enumerate{
\item \strong{Zero Replacement:} Addressing the issue of zeros in compositional data, which can skew
CLR transformation and correlation analysis. The chosen \code{zero_strategy} is applied across
all samples.
\item \strong{CLR Transformation:} Normalizing the data using CLR transformation (\code{clr}) or
interquartile CLR (\code{iclr}), making it suitable for correlation analysis by removing
compositional constraints.
\item \strong{Correlation Analysis:} Computing pairwise correlations between taxa using the specified
\code{cor_method}, resulting in a correlation matrix.
\item \strong{Threshold Application:} Constructing the adjacency matrix by applying the selected
threshold method (\code{thresh_method}) and value (\code{thresh_value}). This step determines which
correlations are considered significant enough to constitute edges in the network.
\item \strong{Network Construction:} The adjacency matrix is used to construct a weighted, undirected
correlation network, represented as an \code{igraph} object.
}
}
\seealso{
\code{\link[mgnet]{clr}}, \code{\link[mgnet]{iclr}} for centered-log ratio details.
\code{\link[mgnet]{zero_dealing}} for details on zero replacement strategies.
}
