% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors.R
\docType{methods}
\name{data}
\alias{data}
\alias{data,ANY-method}
\alias{data,ngs_data-method}
\alias{data,mg-method}
\title{Retrieves ngs data.}
\usage{
data(object)

\S4method{data}{ANY}(object)

\S4method{data}{ngs_data}(object)

\S4method{data}{mg}(object)
}
\arguments{
\item{object}{Required \code{\link{mg-class}} or \code{\link{ngs_data-class}}.}
}
\description{
Return the numeric matrix associated with the ngs data from mg or ngs_data
classes.
}
