% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{data}
\alias{data}
\alias{data,mg-method}
\alias{data,mg}
\title{Retrieves ngs data.}
\usage{
data(object)

\S4method{data}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return the numeric matrix associated with the ngs data from mg class.
}
