% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{degree_mgnet}
\alias{degree_mgnet}
\alias{degree_mgnet,mgnet,character,character-method}
\alias{degree_mgnet,mgnet,character,character}
\alias{degree_mgnet,list,character,character-method}
\alias{degree_mgnet-methods,list,character,character}
\alias{degree_mgnet,mgnet,character,missing-method}
\alias{degree_mgnet,mgnet,character,missing}
\alias{degree_mgnet,list,character,missing-method}
\alias{degree_mgnet-methods,list,character,missing}
\title{Signed Vertices Degrees with Communities Info}
\usage{
degree_mgnet(obj, sign, type)

\S4method{degree_mgnet}{mgnet,character,character}(obj, sign, type)

\S4method{degree_mgnet}{list,character,character}(obj, sign, type)

\S4method{degree_mgnet}{mgnet,character,missing}(obj, sign, type)

\S4method{degree_mgnet}{list,character,missing}(obj, sign, type)
}
\arguments{
\item{obj}{mgnet class.}

\item{sign}{character indicates the sign of the edges. Possible values are
"positive","negative","all".}

\item{type}{character with possible values "intra","extra","all".}
}
\description{
Calculates the vertices degree using the information on the
sign of the edges and the communities. The function make possible to distinguish
between positive, negative or both sign of the edges and discern intra or extra
(also both cases) communities edges.
}
