% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6b_classes_network_degree-DESKTOP-1L7CUD3.R,
%   R/6b_classes_network_degree.R
\name{degree_mgnet}
\alias{degree_mgnet}
\alias{degree_mgnet,mgnet-method}
\alias{degree_mgnet,mgnetList-method}
\title{Signed Vertices Degrees with Communities Info}
\usage{
degree_mgnet(obj, sign = "all", type = "all")

degree_mgnet(obj, sign = "all", type = "all")
}
\arguments{
\item{obj}{An mgnet object.}

\item{sign}{Character indicating the sign of the edges to consider.
Possible values: "positive" for positive weights, "negative" for negative weights, or "all" for all weights.}

\item{type}{Character indicating the type of edges to include.
Possible values: "intra" for edges within communities, "extra" for edges between communities, or "all" for all edges.}
}
\value{
A vector of vertex degrees.

A vector of vertex degrees.
}
\description{
Calculates the vertices degree using the information on the sign of the edges and the communities.

Calculates the vertices degree using the information on the sign of the edges and the communities.
}
