% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{depth}
\alias{depth}
\alias{depth,mg-method}
\alias{depth,mg}
\title{Get samples depth.}
\usage{
depth(object)

\S4method{depth}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Retrieves the depth of each sample using the Depth variable in in meta slot.
Depth is automatically generated if not present during the construction of
the mg object.
}
