% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_mgnet_validity.R
\name{.assertNoReservedKeywords}
\alias{.assertNoReservedKeywords}
\title{Internal: Check for Reserved Keywords in Column Names}
\usage{
.assertNoReservedKeywords(obj, errors)
}
\arguments{
\item{obj}{The data object to check, expected to have column names (e.g., a data frame or a matrix with column names).}

\item{errors}{A character vector where error messages will be accumulated.
New error messages are appended if reserved keywords are found as column names.}
}
\value{
A character vector of accumulated error messages, including any new errors found during this check.
}
\description{
This function checks if any of the specified reserved keywords are present as column names in a given object.
If such keywords are found, it appends an error message to the provided list of errors.
This check helps prevent conflicts with internal functionality that relies on these reserved keywords.
}
\keyword{internal}
