% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-link.R
\name{.ensure_link_id}
\alias{.ensure_link_id}
\title{Ensure the graph has a link_id edge attribute}
\usage{
.ensure_link_id(g)
}
\arguments{
\item{g}{An \code{igraph} object.}
}
\value{
The \code{igraph} object, with guaranteed link_id in the network if edges exist.
}
\description{
This function checks whether the "link_id" attribute is defined. If missing,
it generates link IDs using \code{.make_link_id()} and assigns them.
}
\keyword{internal}
