% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-link.R
\name{.make_link_id}
\alias{.make_link_id}
\title{Create link_id values for a graph}
\usage{
.make_link_id(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
A character vector of length equal to the number of edges.
}
\description{
Generates a character vector of link IDs in the form:
"\if{html}{\out{<from>}}--\if{html}{\out{<to>}}--\if{html}{\out{<index>}}".
}
\keyword{internal}
