% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{empty}
\alias{empty}
\alias{empty,mg-method}
\alias{relative,mg}
\title{Check if the object is empty.}
\usage{
empty(object)

\S4method{empty}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Control if all slots have lengths equal to 0 (therefore empty).
}
