% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\name{[,mg-method}
\alias{[,mg-method}
\title{Method extensions to extraction operator for mg object.}
\usage{
\S4method{[}{mg}(x, i, j)
}
\arguments{
\item{x}{See \code{\link[base]{Extract}}, \code{\link{mg}} object.}

\item{i}{See \code{\link[base]{Extract}}, samples indices.}

\item{j}{See \code{\link[base]{Extract}}, taxa indices}
}
\description{
Method extensions to extraction operator for mg object.
}
\seealso{
\code{\link[base]{Extract}}
}
