% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/class-mgnet.R
\name{[,mgnet-method}
\alias{[,mgnet-method}
\title{Method extensions to extraction operator for mg object.}
\usage{
\S4method{[}{mgnet}(x, i, j)
}
\arguments{
\item{x}{See \code{\link[base]{Extract}}, mgnet object.}
=======
% Please edit documentation in R/class-mg.R, R/class-mgnet.R
\name{[,mg-method}
\alias{[,mg-method}
\alias{[,mgnet-method}
\title{Method extensions to extraction operator for mg object.}
\usage{
\S4method{[}{mg}(x, i, j)

\S4method{[}{mgnet}(x, i, j)
}
\arguments{
\item{x}{See \code{\link[base]{Extract}}, \code{\link{mg}} object.}
>>>>>>> main

\item{i}{See \code{\link[base]{Extract}}, samples indices.}

\item{j}{See \code{\link[base]{Extract}}, taxa indices}
}
\description{
Method extensions to extraction operator for mg object.
<<<<<<< HEAD
}
\seealso{
\code{\link[base]{Extract}}
=======

Method extensions to extraction operator for mgnet object.
}
\seealso{
\code{\link[base]{Extract}}

\code{\link[base]{Extract}}
>>>>>>> main
}
