% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-shows.R
\name{[,ngs_data-method}
\alias{[,ngs_data-method}
\alias{[,sample_metadata-method}
\alias{[,taxonomy_table-method}
\alias{[,mg-method}
\title{Method extensions to extraction operator for mgnet objects.}
\usage{
\S4method{[}{ngs_data}(x, i, j)

\S4method{[}{sample_metadata}(x, i, j)

\S4method{[}{taxonomy_table}(x, i, j)

\S4method{[}{mg}(x, i, j)
}
\arguments{
\item{x}{See \code{\link[base]{Extract}}}

\item{i}{See \code{\link[base]{Extract}}}

\item{j}{See \code{\link[base]{Extract}}}
}
\description{
Method extensions to extraction operator for mgnet objects.
}
\seealso{
\code{\link[base]{Extract}}
}
