% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{filter_taxa}
\alias{filter_taxa}
\alias{filter_taxa,mg,list-method}
\alias{filter_taxa,mg,list}
\alias{filter_taxa,list,list-method}
\alias{filter_taxa,list,list}
\title{Filter taxa based on across-sample abundance criteria.}
\usage{
filter_taxa(object, flist)

\S4method{filter_taxa}{mg,list}(object, flist)

\S4method{filter_taxa}{list,list}(object, flist)
}
\arguments{
\item{object}{\code{\link{mg-class}}.}

\item{flist}{\code{\link{list}}. Each element of flist it must be
a function.}
}
\description{
It applies an arbitrary set of functions list as across-sample criteria,
one taxa at a time. The function takes as input a mg object, and returns its
trimmed mg version.
}
