% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{filter_taxa}
\alias{filter_taxa}
\alias{filter_taxa,mg,list,logical-method}
\alias{filter_taxa,mg-method}
\title{Filter taxa based on across-sample abundance criteria.}
\usage{
filter_taxa(object, flist, join.trim)

\S4method{filter_taxa}{mg,list,logical}(object, flist, join.trim)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{flist}{(Required) \code{\link{list}}. Each element of flist it must be
a function.}

\item{trim}{(Optional) Default \code{FALSE}.}
}
\description{
It applies an arbitrary set of functions list as across-sample criteria,
one taxa at a time. The function takes as input a mg object,
and returns its trimmed mg version or a logical vector
indicating whether or not each taxa passed the criteria.
Alternatively, if the \code{"trim"} option is set to \code{FALSE},
it return a logical vector indicating whether or not each taxa passed
the criteria.
}
