% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities_link.R
\name{get_selected_links}
\alias{get_selected_links}
\title{Get Selected Links from an mgnet Object}
\usage{
get_selected_links(object)
}
\arguments{
\item{object}{An \code{mgnet} object.}
}
\value{
Numeric vector of indices representing the selected links.
}
\description{
Retrieves the indices of the links that were previously selected using \code{select_link}.
}
