% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP2}
\alias{HMP2}
\title{Example omic object derived from the Human Microbiome Project}
\format{
An object of class \code{omic} with \code{1112} samples and \code{1953} taxa (OTU in this case).
}
\source{
Human Microbiome Project Phase 2
}
\usage{
data(HMP2)
}
\description{
This object contains OTU counts, sample metadata, and taxonomic information
for a subset of the HMP2 dataset, stored as an \code{omic} class.
}
\keyword{datasets}
