% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6c_mgnet_layout_signed.R
\name{layout_signed}
\alias{layout_signed}
\alias{layout_signed,igraph-method}
\alias{layout_signed,mgnet-method}
\title{Signed-Weighted Graph Layout}
\usage{
layout_signed(object)
}
\arguments{
\item{object}{A network object belonging to either the \code{igraph} or \code{mgnet} class.
The graph must be weighted and contain at least one positive-weighted edge.}
}
\value{
A matrix of vertex coordinates determined by the Fruchterman-Reingold layout algorithm,
considering only positive-weighted edges.
}
\description{
Generates layout coordinates for a graph by considering only positive-weighted edges.
This function is applicable to graphs of the \code{igraph} and \code{mgnet} classes. It emphasizes
the structure formed by positive interactions by ignoring edges with non-positive weights.
}
