% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mgnet.R
\docType{methods}
\name{make_mgnet}
\alias{make_mgnet}
\alias{make_mgnet,mgnet,character,character,character,numeric-method}
\alias{make_mgnet,list,character,character,character,numeric-method}
\title{Build a meta-genomic network.}
\usage{
make_mgnet(mgnet, cor.method, thresh.method, adjust, thresh)

\S4method{make_mgnet}{mgnet,character,character,character,numeric}(mgnet, cor.method, thresh.method, adjust, thresh)

\S4method{make_mgnet}{list,character,character,character,numeric}(mgnet, cor.method, thresh.method, adjust, thresh)
}
\arguments{
\item{mgnet}{mgnet object}

\item{cor.method}{correlation type, possible choices are "pearson","spearman","kendall".}

\item{thresh.method}{threshold method to retrieve the adjacency matrix. Possible
choices are "absolute","density","p-value".}

\item{adjust}{multiple test criteria, possible choices are "holm",
"hochberg","hommel","bonferroni","BH","BY","fdr","none".}

\item{thresh}{numeric indicates the threshold value.}
}
\description{
The function has the purpose of building a meta-genomic network
starting from a log_data slot in object following the following steps:
\itemize{
\item Calculates the pairwise taxa correlation matrix from the log_data slot.
\item Retrieves the adjacency matrix by applying a threshold to the lowest
correlations with three possible criteria 'abosulte','density','p-value'.
\item Constructs the undirected, weighted, and signed network of class igraph.
}
}
