% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-merge.R
\name{merge_mgnet}
\alias{merge_mgnet}
\title{Merge \code{mgnet} or \code{mgnets} objects}
\usage{
merge_mgnet(
  ...,
  abun_fun = NULL,
  rela_fun = NULL,
  norm_fun = NULL,
  meta_fun = NULL,
  taxa_fun = NULL,
  netw_fun = NULL,
  comm_fun = NULL
)
}
\arguments{
\item{...}{\code{mgnet} objects, \code{mgnets} objects, or lists of such objects.}

\item{abun_fun, rela_fun, norm_fun}{Functions used to merge abundance-like slots.}

\item{meta_fun, taxa_fun}{Functions used to merge metadata tables.}

\item{netw_fun, comm_fun}{Functions used to merge network and community slots.}
}
\value{
An object of class \code{mgnet} if the inputs are \code{mgnet} objects, or an object
of class \code{mgnets} if the inputs are \code{mgnets} objects.
}
\description{
Merge multiple \code{mgnet} objects into a single \code{mgnet}, or merge multiple
\code{mgnets} objects name-by-name into a single \code{mgnets}.

Input objects must be all of class \code{mgnet} or all of class \code{mgnets}.
Mixed input is not allowed.

For each slot, the user may supply a merge function. If a merge function is
not provided for a given slot, that slot is left empty in the merged result.
}
