% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-show.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,ANY-method}
\alias{meta,mg-method}
\title{Retrieves sample metadata.}
\usage{
meta(object)

\S4method{meta}{ANY}(object)

\S4method{meta}{mg}(object)
}
\arguments{
\item{object}{Required \code{\link{mg-class}}.}
}
\description{
Return the data.frame associated with the sample metadata with experimental
variables from mg class.
}
\seealso{
\code{\link{sample_metadata}}
}
