% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_classes_getters.R
\name{meta}
\alias{meta}
\alias{meta,mgnet-method}
\alias{meta,mgnetList-method}
\title{Get Sample Metadata Information}
\usage{
meta(object, .fmt = "df")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are:
- "df": Returns the output as a \code{data.frame}.
- "tbl": Returns the output as a \code{tibble}. For \code{mgnet} objects, the row names of
the abundance matrix are converted into a new column named \code{sample_id}, ensuring alignment
with the reserved keyword in \code{mgnet-class}.}
}
\value{
The content of the \code{meta} slot for \code{mgnet} or a list of such contents for \code{mgnetList}.
}
\description{
Retrieves the sample information stored in the \code{meta} slot of an \code{mgnet} object
or for each \code{mgnet} object within an \code{mgnetList}, with the option to format the output as
a \code{data.frame}, \code{tibble}.
}
