% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{meta}
\alias{meta}
\alias{meta,mgnet-method}
\alias{meta,mgnets-method}
\title{Get Sample Metadata Information}
\usage{
meta(object, .fmt = "df", .collapse = FALSE)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are:
- "df": Returns the output as a \code{data.frame}.
- "tbl": Returns the output as a \code{tibble}. For \code{mgnet} objects, the row names of
the abundance matrix are converted into a new column named \code{sample_id}, ensuring alignment
with the reserved keyword in \code{mgnet-class}.}

\item{.collapse}{Logical, only for \code{mgnets}. If TRUE, ignore \code{.fmt} and return
a single tibble with all metas row-bound together and an \code{mgnet} column
indicating the source object. Default: FALSE.}
}
\value{
The content of the \code{meta} slot for \code{mgnet} or a list of such contents for \code{mgnets}.
}
\description{
Retrieves the sample information stored in the \code{meta} slot of an \code{mgnet} object
or for each \code{mgnet} object within an \code{mgnets}, with the option to format the output as
a \code{data.frame}, \code{tibble}.
}
