% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,mg-method}
\alias{meta,mg}
\alias{meta,list-method}
\alias{meta,list}
\title{Retrieves sample metadata.}
\usage{
meta(object)

\S4method{meta}{mg}(object)

\S4method{meta}{list}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return the data.frame associated with the sample metadata with experimental
variables from mg class.
}
