% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-definitions-validators-constructors.R
\docType{class}
\name{mg-class}
\alias{mg-class}
\title{S4 class to manage metagenomics data}
\description{
S4 class to manage metagenomics data
}
\section{Slots}{

\describe{
\item{\code{data}}{ngs data}

\item{\code{meta}}{samples experimental features.}

\item{\code{taxa}}{taxonomic table}
}}

