% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\name{mg}
\alias{mg}
\title{User constructor for mg s4 class.}
\usage{
mg(data, meta, taxa)
}
\arguments{
\item{data}{numeric matrix with all elements >=0.}

\item{meta}{data.frame with experimental variables.}

\item{taxa}{character matrix with taxonomic classification.}
}
\description{
User constructor to create an object belonging to formal s4 class mg
avoiding the new function.
}
