% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{mgmelt}
\alias{mgmelt}
\alias{mgmelt,mgnet-method}
\alias{mgmelt,list-method}
\title{Melt mg object into form suitable for easy casting.}
\usage{
mgmelt(object)

\S4method{mgmelt}{mgnet}(object)

\S4method{mgmelt}{list}(object)
}
\arguments{
\item{object}{mgnet.}
}
\description{
Summarize all elements present of an mgnet object in a long format data frame.
}
