% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{mgmelt}
\alias{mgmelt}
\alias{mgmelt,mg-method}
\alias{mgmelt,list-method}
\alias{mgmelt,list}
\title{Melt mg object into form suitable for easy casting.}
\usage{
mgmelt(object)

\S4method{mgmelt}{mg}(object)

\S4method{mgmelt}{list}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Summarize all elements present in an mg object in a single data frame.
The functioning is similar to melt function of reshape2 package and it will
be useful as a preprocess for ggplot2.
}
