% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{mgmelt}
\alias{mgmelt}
\alias{mgmelt,mgnet-method}
\alias{mgmelt,list-method}
\title{Melt mg object into form suitable for easy casting.}
\usage{
mgmelt(object)

\S4method{mgmelt}{mgnet}(object)

\S4method{mgmelt}{list}(object)
}
\arguments{
\item{object}{mgnet.}
}
\description{
Summarize all elements present in an mg object in a single data frame.
The functioning is similar to melt function of reshape2 package and it will
be useful as a preprocess for ggplot2.
}
