% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{class}
\name{mgnet-class}
\alias{mgnet-class}
\alias{mgnet}
\title{S4 class to manage metagenomic networks}
\description{
S4 class to manage metagenomic networks
}
\section{Slots}{

\describe{
\item{\code{data}}{table of data}

\item{\code{meta}}{sample features}

\item{\code{taxa}}{taxonomic information}

\item{\code{netw}}{undirected, weighted and signed igraph network}

\item{\code{comm}}{graph communities of netw}
}}

