% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{mgnet-class}
\alias{mgnet-class}
\title{Metagenomics Network Class}
\description{
An S4 class designed to manage and analyze metagenomics networks, including
abundance data, taxa lineage, sample and taxa metadata.
}
\section{Slots}{

\describe{
\item{\code{abundance}}{numeric matrix representing the abundance data.}

\item{\code{sampleInfo}}{data.frame containing metadata about samples.}

\item{\code{lineage}}{matrix with the taxonomic lineage of each taxa.}

\item{\code{taxaInfo}}{data.frame with metadata about taxa.}

\item{\code{association}}{matrix representing relationships between taxa, such as correlations or conditional independencies.}

\item{\code{network}}{object belong to 'igraph' class depicting the network of interactions among taxa.}

\item{\code{community}}{A 'communities' object representing detected communities within the network.}
}}

