% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\name{mgnet-class}
\alias{mgnet-class}
\alias{mgnet}
\title{S4 class to manage metagenomic networks}
\usage{
mgnet(
  data = matrix(nrow = 0, ncol = 0),
  meta = data.frame(),
  taxa = matrix(nrow = 0, ncol = 0),
  netw = make_empty_graph(n = 0, directed = FALSE),
  comm = cluster_fast_greedy(make_empty_graph(n = 0, directed = FALSE)),
  mg = new("mg")
)
}
\description{
S4 class to manage metagenomic networks
}
\section{Slots}{

\describe{
\item{\code{data}}{table of data}

\item{\code{meta}}{sample features}

\item{\code{taxa}}{taxonomic information}

\item{\code{netw}}{undirected, weighted and signed igraph network}

\item{\code{comm}}{graph communities of netw}
}}

