% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_mgnet_class.R
\name{mgnet}
\alias{mgnet}
\title{Create an mgnet Object}
\usage{
mgnet(
  abun = matrix(numeric(0), nrow = 0, ncol = 0),
  rela = matrix(numeric(0), nrow = 0, ncol = 0),
  norm = matrix(numeric(0), nrow = 0, ncol = 0),
  meta = data.frame(),
  taxa = data.frame(),
  netw = make_empty_graph(n = 0, directed = FALSE),
  comm = cluster_fast_greedy(make_empty_graph(n = 0, directed = FALSE))
)
}
\arguments{
\item{abun}{Numeric matrix with all elements >=0 representing the abundance data
where rows are samples and columns are taxa (OTUs, Species, ...). Defaults to an empty matrix.}

\item{rela}{Numeric matrix representing the relative abundance of each taxon within each sample.
Defaults to an empty matrix.}

\item{norm}{Numeric matrix of normalized abundance data.
Defaults to an empty matrix.}

\item{meta}{Data.frame containing metadata for samples. Each row should correspond
to a sample and each column to an experimental variable. Defaults to an empty data frame.}

\item{taxa}{Data.frame with additional information on taxa. Each row should correspond
to a taxa. Defaults to an empty data frame.}

\item{netw}{An \code{igraph} object representing a network of taxa interactions. Defaults to an
empty graph.}

\item{comm}{An object storing community detection results from \code{network}. Defaults
to the result of a fast greedy clustering on an empty graph.}
}
\value{
Returns an \code{mgnet} object encapsulating the provided metagenomic data.
If an error occurs during object creation, a custom error message is displayed and
the function execution is stopped.
}
\description{
This function constructs an \code{mgnet} object, encapsulating various types of metagenomic data
including abundances data, sample metadata, taxonomic information, network analysis results, etc.,
providing a comprehensive framework for the analysis of microbial communities and their interactions.
It handles object creation with custom error handling for improved user experience.
}
\examples{
# Assuming correct_matrix is a properly defined matrix and other necessary data are defined
mgnet_obj <- mgnet(abun = otu_HMP2,
                   sample = info_sample_HMP2,
                   taxa = info_taxa_HMP2)
}
