% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_mgnet_class.R
\name{mgnet}
\alias{mgnet}
\title{Create an mgnet Object}
\usage{
mgnet(
  abundance = matrix(numeric(0), nrow = 0, ncol = 0),
  rel_abundance = matrix(numeric(0), nrow = 0, ncol = 0),
  norm_abundance = matrix(numeric(0), nrow = 0, ncol = 0),
  info_sample = data.frame(),
  lineage = matrix(character(0), nrow = 0, ncol = 0),
  info_taxa = data.frame(),
  network = make_empty_graph(n = 0, directed = FALSE),
  community = cluster_fast_greedy(make_empty_graph(n = 0, directed = FALSE))
)
}
\arguments{
\item{abundance}{numeric matrix with all elements >=0 representing the abundance data
where rows are samples and columns are taxa (OTUs, Species, ...). Defaults to an empty matrix.}

\item{rel_abundance}{...}

\item{norm_abundance}{numeric matrix of log-transformed abundance data, structured like \code{abundance}.
Defaults to an empty matrix.}

\item{info_sample}{data.frame containing metadata for samples. Each row should correspond
to a sample and each column to an experimental variable. Defaults to an empty data frame.}

\item{lineage}{character matrix with taxonomic classification for each taxa. Each row
corresponds to a taxa and each column to a rank. Defaults to an empty matrix.}

\item{info_taxa}{data.frame with additional information on taxa. Each row should correspond
to a taxa. Defaults to an empty data frame.}

\item{network}{an \code{igraph} object representing a network of taxa interactions. Defaults to an
empty graph.}

\item{community}{an object storing community detection results from \code{network}. Defaults
to the result of a fast greedy clustering on an empty graph.}
}
\value{
Returns an \code{mgnet} object encapsulating the provided metagenomic data.
If an error occurs during object creation, a custom error message is displayed and
the function execution is stopped.
}
\description{
This function constructs an \code{mgnet} object, encapsulating various types of metagenomic data
including abundance data, sample metadata, taxonomic information, network analysis results, etc.,
providing a comprehensive framework for the analysis of microbial communities and their interactions.
It handles object creation with custom error handling for improved user experience.
}
\examples{
# Assuming correct_matrix is a properly defined matrix and other necessary data are defined
mgnet_obj <- mgnet(abundance = otu_HMP2,
                   info_sample = info_sample_HMP2,
                   lineage = lineage_HMP2)
}
