% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/class-mgnet.R
\name{mgnet}
\alias{mgnet}
\title{User constructor for mgnet s4 class.}
\usage{
mgnet(
  data = matrix(nrow = 0, ncol = 0),
  meta_sample = data.frame(),
  taxa = matrix(nrow = 0, ncol = 0),
  meta_taxa = data.frame(),
  log_data = matrix(nrow = 0, ncol = 0),
  netw = make_empty_graph(n = 0, directed = FALSE),
  comm = cluster_fast_greedy(make_empty_graph(n = 0, directed = FALSE))
)
}
\arguments{
\item{data}{numeric matrix with all elements >=0.}

\item{meta_sample}{data.frame with experimental variables.}

\item{taxa}{character matrix with taxonomic classification.}

\item{meta_taxa}{data.frame with addition info on taxa.}

\item{log_data}{log-transformed data.}

\item{netw}{un-directed weighted igraph network or the associated adjacency matrix.}

\item{comm}{communities object associated to netw.}
}
\description{
User constructor to create an object belonging to formal s4
class mgnet avoiding the new function.
=======
% Please edit documentation in R/1a_mgnet_class.R
\name{mgnet}
\alias{mgnet}
\title{Create an mgnet Object}
\usage{
mgnet(
  abundance = matrix(numeric(0), nrow = 0, ncol = 0),
  rel_abundance = matrix(numeric(0), nrow = 0, ncol = 0),
  norm_abundance = matrix(numeric(0), nrow = 0, ncol = 0),
  info_sample = data.frame(),
  lineage = matrix(character(0), nrow = 0, ncol = 0),
  info_taxa = data.frame(),
  network = make_empty_graph(n = 0, directed = FALSE),
  community = cluster_fast_greedy(make_empty_graph(n = 0, directed = FALSE))
)
}
\arguments{
\item{abundance}{Numeric matrix with all elements >=0 representing the abundance data
where rows are samples and columns are taxa (OTUs, Species, ...). Defaults to an empty matrix.}

\item{rel_abundance}{Numeric matrix representing the relative abundance of each taxon within each sample.
Defaults to an empty matrix.}

\item{norm_abundance}{Numeric matrix of log-transformed abundance data, structured like \code{abundance}.
Defaults to an empty matrix.}

\item{info_sample}{Data.frame containing metadata for samples. Each row should correspond
to a sample and each column to an experimental variable. Defaults to an empty data frame.}

\item{lineage}{Character matrix with taxonomic classification for each taxa. Each row
corresponds to a taxa and each column to a rank. Defaults to an empty matrix.}

\item{info_taxa}{Data.frame with additional information on taxa. Each row should correspond
to a taxa. Defaults to an empty data frame.}

\item{network}{An \code{igraph} object representing a network of taxa interactions. Defaults to an
empty graph.}

\item{community}{An object storing community detection results from \code{network}. Defaults
to the result of a fast greedy clustering on an empty graph.}
}
\value{
Returns an \code{mgnet} object encapsulating the provided metagenomic data.
If an error occurs during object creation, a custom error message is displayed and
the function execution is stopped.
}
\description{
This function constructs an \code{mgnet} object, encapsulating various types of metagenomic data
including abundance data, sample metadata, taxonomic information, network analysis results, etc.,
providing a comprehensive framework for the analysis of microbial communities and their interactions.
It handles object creation with custom error handling for improved user experience.
}
\examples{
# Assuming correct_matrix is a properly defined matrix and other necessary data are defined
mgnet_obj <- mgnet(abundance = otu_HMP2,
                   info_sample = info_sample_HMP2,
                   lineage = lineage_HMP2)
>>>>>>> from_scratch
}
