% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-merge.R
\name{mgnet_collapse}
\alias{mgnet_collapse}
\alias{mgnet_collapse,mgnets-method}
\title{Collapse Specified Combinations within an mgnets}
\usage{
mgnet_collapse(
  object,
  by = NULL,
  abun = NULL,
  rela = NULL,
  norm = NULL,
  meta = NULL,
  taxa = NULL,
  netw = NULL,
  comm = NULL
)
}
\arguments{
\item{object}{An \code{mgnets} containing multiple \code{mgnet} objects.}

\item{by}{A list specifying the combinations of \code{mgnet} object names to be merged. Each element in
the list should be a character vector containing the names of the \code{mgnet} objects to merge.
If NULL, all \code{mgnet} objects in the list are merged into a single \code{mgnet} object.}

\item{abun}{A function to merge the 'abun' (abundance) slots from the selected \code{mgnet} objects, or
NULL if no merging is required for this slot.}

\item{rela}{A function to merge the 'rela' (relative abundance) slots, or NULL if not required.}

\item{norm}{A function to merge the 'norm' (normalized abundance) slots, or NULL if not required.}

\item{meta}{A function to merge the 'meta' (metadata) slots, or NULL if not required.}

\item{taxa}{A function to merge the 'taxa' (taxonmgnet data) slots, or NULL if not required.}

\item{netw}{A function to merge the 'netw' (network) slots, or NULL if not required.}

\item{comm}{A function to merge the 'comm' (community structure) slots, or NULL if not required.}
}
\value{
An \code{mgnets} containing the merged \code{mgnet} objects. Each entry in the returned \code{mgnets}
corresponds to a merged group as defined in the \code{by} parameter. The names of the entries
reflect the groups specified in \code{by} or are derived by concatenating the names of the merged
\code{mgnet} objects with dashes if \code{by} is unnamed.
}
\description{
Merges specific combinations of \code{mgnet} objects within an \code{mgnets} according to user-defined rules,
allowing for targeted merging of subsets within the list. This function facilitates focused analyses
by enabling the aggregation of related data into meaningful groups.
}
\details{
The function checks for the validity of names specified in the \code{by} parameter against the names
of \code{mgnet} objects in the \code{mgnets}. It ensures that all specified groups contain valid \code{mgnet}
object names and that each group's name is unique. This function is useful for scenarios where
specific subsets of data within an \code{mgnets} need to be aggregated based on certain criteria
or experimental conditions.

If \code{by} is NULL, a default merge of all \code{mgnet} objects in the list is performed, resulting in a
single \code{mgnet} object.
}
