% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4f_classes_longer.R
\name{mgnet_longer}
\alias{mgnet_longer}
\alias{mgnet_longer,mgnet-method}
\alias{mgnet_longer,mgnetList-method}
\title{Convert mgnetList Object to Longer Format}
\usage{
mgnet_longer(object, abundance=NULL, relative=NULL, log_abundance=NULL,
       lineage=NULL, info_sample=NULL, info_taxa=NULL)
}
\arguments{
\item{object}{An \code{mgnetList} object.}

\item{abundance}{Logical; if TRUE, includes abundance data.}

\item{relative}{Logical; if TRUE, includes relative abundance data.}

\item{log_abundance}{Logical; if TRUE, includes log-transformed abundance data.}

\item{lineage}{Logical; if TRUE, includes lineage information.}

\item{info_sample}{Logical; if TRUE, includes sample metadata.}

\item{info_taxa}{Logical; if TRUE, includes taxonomic information.}
}
\value{
A \code{tibble} in a long format containing the combined data from each \code{mgnet} object
in the \code{mgnetList}, with an additional column \code{source_mgnet} indicating the source object.
}
\description{
Converts data from each \code{mgnet} object within an \code{mgnetList} into a longer format,
suitable for easy analysis and casting, by combining abundance, sample metadata,
taxonomic information, and other relevant data. A new column, \code{source_mgnet}, is added
to identify the source \code{mgnet} object for each row in the combined data.
}
