% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4f_classes_longer.R
\name{mgnet_longer}
\alias{mgnet_longer}
\alias{mgnet_longer,mgnet-method}
\alias{mgnet_longer,mgnetList-method}
\title{Transform mgnet Object(s) to Longer Format}
\usage{
mgnet_longer(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object. For \code{mgnet}, the method will transform the data
to a longer format by combining information on samples, taxa, abundances,
and metadata. For \code{mgnetList}, the method will apply the transformation to each \code{mgnet}
object and bind the results into a single tibble, with an additional column named \code{mgnet}
indicating the source \code{mgnet} object.}
}
\value{
A tibble in a longer format that includes all the combined information. For \code{mgnetList}
objects, the output is a single tibble where results from all \code{mgnet} objects are bound
together, and an extra \code{mgnet} column identifies the source \code{mgnet} object.
}
\description{
Converts an \code{mgnet} or \code{mgnetList} object into a "longer" format, collecting all possible
information (abundance, relative abundance, normalized abundance, and metadata) into a single
tibble. This method uses a tidyverse approach to gather data into a structure suitable for
further analysis.
}
\details{
This method utilizes specific reserved keywords from the \code{mgnet} class, including:
\itemize{
\item \code{sample_id}: The identifier for each sample.
\item \code{taxa_id}: The identifier for each taxon.
\item \code{comm_id}: The identifier for membership community for each taxon.
\item \code{mgnet}: Added when dealing with \code{mgnetList}, to track the source \code{mgnet} object.
\item \code{abun}: Refers to abundance values in the data.
\item \code{rela}: Refers to relative abundance values in the data.
\item \code{norm}: Refers to normalized abundance values in the data.
\item For \code{mgnet} objects: The method gathers data on samples (\code{sample_id}), taxa (\code{taxa_id}),
and abundance values (e.g., \code{abun}, \code{rela}, \code{norm}) into a longer format, combining sample
metadata and taxa metadata as well.
\item For \code{mgnetList} objects: The transformation is applied to each individual \code{mgnet} in the list.
The resulting longer format data for each \code{mgnet} is then combined into a single tibble, with
an additional column \code{mgnet}, which contains the name of each \code{mgnet} object. This allows
users to easily track which data comes from which \code{mgnet}.
\item If the network community information are available a column \code{comm_id} is
added with the membership for each taxon.
}

The tidyverse functions used include \code{pivot_longer} for reshaping data, and \code{left_join} for
merging sample and taxa information. The \code{mgnet_longer} method ensures that all relevant data
(e.g., abundance, relative abundance, normalized abundance, metadata) is organized in a
consistent and accessible manner.
}
