% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8a_classes_longer.R
\name{mgnet_longer}
\alias{mgnet_longer}
\alias{mgnet_longer,mgnet-method}
\alias{mgnet_longer,mgnetList-method}
\title{Convert mgnetList Object to Longer Format}
\usage{
mgnet_longer(
  object,
  source_mgnet_to,
  abundance = NULL,
  rel_abundance = NULL,
  norm_abundance = NULL,
  lineage = NULL,
  info_sample = NULL,
  info_taxa = NULL,
  community = NULL
)
}
\arguments{
\item{object}{An \code{mgnetList} object containing multiple \code{mgnet} objects to be combined and
transformed into a longer format.}

\item{source_mgnet_to}{A string specifying the column name to be added for identifying the source
\code{mgnet} object within an \code{mgnetList}. This parameter is required for \code{mgnetList} objects and
must be provided; it is not used for single \code{mgnet} objects.}

\item{abundance}{A logical value; if \code{TRUE}, includes abundance data for each taxa in each sample.}

\item{rel_abundance}{A logical value; if \code{TRUE}, includes relative abundance data for each taxa
in each sample, facilitating comparison across samples.}

\item{norm_abundance}{A logical value; if \code{TRUE}, includes normalized abundance data for each taxa,
useful for downstream analyses that require standardized input.}

\item{lineage}{A logical value; if \code{TRUE}, includes lineage information for each taxa, enriching
the dataset with taxonomic context.}

\item{info_sample}{A logical value; if \code{TRUE}, includes sample metadata, such as environmental
conditions or collection details, enhancing interpretability.}

\item{info_taxa}{A logical value; if \code{TRUE}, includes detailed taxonomic information for each taxa,
supporting fine-grained biological insights.}

\item{community}{A logical value; if \code{TRUE}, includes community membership data for each taxa,
enabling analysis of community structure and diversity.}
}
\value{
For \code{mgnetList} objects, returns a \code{tibble} in a long format containing combined data from
each \code{mgnet} object in the list, enriched with specified metadata and abundance information,
and an additional column indicating the source \code{mgnet} object as specified by \code{source_mgnet_to}.
For \code{mgnet} objects, returns a similar \code{tibble} without the source identification column.
}
\description{
This function transforms data from each \code{mgnet} object within an \code{mgnetList} into a longer format,
combining abundance data, sample metadata, taxonomic information, and community memberships into a
single, comprehensive tibble. Each row represents a unique combination of sample and taxa, enriched
with the specified data types. For \code{mgnetList} objects, an additional column is added to identify
the source \code{mgnet} object for each row, facilitating data origin tracking. The \code{source_mgnet_to}
parameter is mandatory for \code{mgnetList} objects to specify this column name.
}
\note{
The function checks for the presence of requested data types in each \code{mgnet} object and
automatically excludes any data types not present or specifically set to \code{FALSE}.
It's essential to ensure that each \code{mgnet} object within the \code{mgnetList} is properly
formatted and contains the necessary data for the requested outputs.
}
