% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_mgnetList_class.R
\docType{class}
\name{mgnetList-class}
\alias{mgnetList-class}
\title{MetaGenomic NETwork List (mgnetList) Class}
\description{
mgnetList is an S4 class designed to hold and manage a collection of mgnet objects,
providing a structured way to handle multiple metagenomic networks. It ensures that all
elements within the list adhere to the mgnet class structure, making it useful for
analyses involving multiple metagenomic datasets. This class allows for consistent and
type-safe manipulation of multiple mgnet objects.
}
\section{Slots}{

\describe{
\item{\code{mgnets}}{A list where each element is an instance of the mgnet class.
If the list is not empty, each mgnet object within the list must be named to facilitate
easy access and management.}
}}

\section{Validity Checks}{

The class includes validity checks to ensure that, if not empty, all elements within the
mgnets slot are instances of the mgnet class and are named. These checks help maintain
integrity and usability, preventing errors during downstream analyses.
}

\seealso{
\code{\link[=mgnet]{mgnet}} for details on the mgnet class and its functionalities.
}
