% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_mgnetList_class.R
\name{mgnetList}
\alias{mgnetList}
\title{Create an mgnetList Object}
\usage{
mgnetList(...)
}
\arguments{
\item{...}{Optional mgnet objects or a list of mgnet objects to be included in the mgnetList.
Providing named mgnet objects enhances manageability and clarity. The list can be
initialized as empty.}
}
\value{
An mgnetList object containing any provided mgnet objects. If not empty,
each mgnet object must be named, ensuring structured and accessible management
of multiple metagenomic networks.
}
\description{
Constructs an mgnetList object to encapsulate and manage a collection of mgnet objects.
This function ensures that all provided mgnet objects are validated and organized into a
structured list, facilitating easy access to multiple metagenomic networks. The constructor
supports the initialization of an empty list, individual mgnet objects, or a list of mgnet objects.
}
