% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnets.R
\name{mgnets-methods}
\alias{mgnets-methods}
\alias{as.list.mgnets}
\alias{coerce,mgnets,list-method}
\alias{length,mgnets-method}
\alias{names,mgnets-method}
\alias{names<-,mgnets,character-method}
\title{General methods for \code{mgnets} objects}
\usage{
\method{as.list}{mgnets}(x, ...)

\S4method{length}{mgnets}(x)

\S4method{names}{mgnets}(x)

\S4method{names}{mgnets,character}(x) <- value
}
\arguments{
\item{x}{An object of class \code{mgnets}.}

\item{...}{Not used.}

\item{value}{A character vector for the replacement form \verb{names<-}.}
}
\value{
\itemize{
\item \code{as.list()} / \code{as(., 'list')}: a list of \code{mgnet} objects. \cr
\item \code{length()}: an integer with the number of elements. \cr
\item \code{names()}: a character vector with element names. \cr
\item \verb{names<-()}: the updated \code{mgnets} object.
}
}
\description{
Basic utilities to interact with \code{mgnets} objects: coerce to list, get length,
get and set names.
}
\section{Methods}{

\describe{
\item{\code{as.list(x)}}{Convert an \code{mgnets} object to a plain list of \code{mgnet} objects.}
\item{\code{as(x, 'list')}}{Coerce the \code{mgnets} object to a list.}
\item{\code{length(x)}}{Return the number of contained \code{mgnet} objects.}
\item{\code{names(x)}}{Return element names.}
\item{\code{names(x) <- value}}{Set element names.}
}
}

