% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnets.R
\name{mgnets}
\alias{mgnets}
\title{Create an \code{mgnets} object}
\usage{
mgnets(...)
}
\arguments{
\item{...}{\code{mgnet} objects, or a \emph{list} of \code{mgnet} objects.}
}
\value{
An object of class \code{mgnets}. If not empty, elements must be named with
unique names.
}
\description{
Build an \code{mgnets} object to encapsulate a collection of \code{mgnet} objects.
Accepts multiple \code{mgnet} arguments or a single \emph{list} of \code{mgnet}.
}
