% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R,
%   R/2c_classes_extractors.R
\name{mgnets}
\alias{mgnets}
\alias{mgnets,mgnetList-method}
\title{Retrieve list of mgnet Objects from an mgnetList}
\usage{
mgnets(object)

mgnets(object)
}
\arguments{
\item{object}{An \code{mgnetList} object from which \code{mgnet} objects are to be retrieved.}
}
\value{
A list of \code{mgnet} objects, where each element is an individual \code{mgnet} object. The elements
are named, corresponding to the names of the \code{mgnet} objects within the \code{mgnetList}, allowing
for easy identification and access. This facilitates individual or batch processing of
metagenomic network data.

A list of \code{mgnet} objects, where each element is an individual \code{mgnet} object. The elements
are named, corresponding to the names of the \code{mgnet} objects within the \code{mgnetList}, allowing
for easy identification and access. This facilitates individual or batch processing of
metagenomic network data.
}
\description{
This method extracts and returns the list of \code{mgnet} objects contained within an \code{mgnetList} object,
offering direct access to the individual \code{mgnet} objects for further analysis or manipulation. It functions
similarly to the \code{\link{as.list}} method for \code{mgnetList} objects, both providing a way to access
the contained \code{mgnet} objects.

This method extracts and returns the list of \code{mgnet} objects contained within an \code{mgnetList} object,
offering direct access to the individual \code{mgnet} objects for further analysis or manipulation. It functions
similarly to the \code{\link{as.list}} method for \code{mgnetList} objects, both providing a way to access
the contained \code{mgnet} objects.
}
\seealso{
\code{\link[=mgnet]{mgnet}} for details on the \code{mgnet} class and functionalities.
\code{\link[=mgnetList]{mgnetList}} for guidance on creating and managing \code{mgnetList} objects.
\code{\link{as.list}} for a similar method that converts an \code{mgnetList} into a list of \code{mgnet} objects,
serving the same purpose as \code{mgnets}.

\code{\link[=mgnet]{mgnet}} for details on the \code{mgnet} class and functionalities.
\code{\link[=mgnetList]{mgnetList}} for guidance on creating and managing \code{mgnetList} objects.
\code{\link{as.list}} for a similar method that converts an \code{mgnetList} into a list of \code{mgnet} objects,
serving the same purpose as \code{mgnets}.
}
