% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4f_classes_mutate_criteria.R
\name{mutate_criteria_taxa}
\alias{mutate_criteria_taxa}
\alias{mutate_criteria_taxa,mgnet-method}
\alias{mutate_criteria_taxa,mgnetList-method}
\title{Mutate Taxa Information based on Abundance Criteria}
\usage{
mutate_criteria_taxa(
  object,
  abundance_criteria = NULL,
  rel_abundance_criteria = NULL,
  norm_abundance_criteria = NULL,
  network_criteria = NULL,
  community_criteria = NULL,
  mgnet_criteria = NULL
)
}
\arguments{
\item{object}{An \code{mgnetList} object.}

\item{abundance_criteria}{A named list of functions to be applied to the abundance data matrix.
Each function should take a single row of abundance data as input and return the value
to be stored in the \code{info_taxa} slot. Default is \code{NULL}.}

\item{rel_abundance_criteria}{A named list of functions to be applied to the relative abundance data matrix.
Each function should take a single row of relative abundance data as input and return
the value to be stored in the \code{info_taxa} slot. Default is \code{NULL}.}

\item{norm_abundance_criteria}{A named list of functions to be applied to the log-transformed abundance data
matrix. Each function should take a single row of log-transformed abundance data as input
and return the value to be stored in the \code{info_taxa} slot. Default is \code{NULL}.}

\item{network_criteria}{A named list of functions to be applied to the network data matrix.
Each function should take a single column of network data as input and return
the value to be stored in the \code{info_taxa} slot. Default is \code{NULL}.}

\item{community_criteria}{A named list of functions to be applied to the community data matrix.
Each function should take a single column of community data as input and return
the value to be stored in the \code{info_taxa} slot. Default is \code{NULL}.}

\item{mgnet_criteria}{A named list of functions to be applied direct on the mgnet object.}
}
\value{
A modified \code{mgnet} object with the \code{info_taxa} slot updated based on the applied mutations.
}
\description{
This function allows for mutating taxa information based on specified criteria applied to
the abundances matrices in an \code{mgnet} object or \code{mgnetList} object.
Custom functions define the criteria for mutation, enabling flexible modification of taxa
information across multiple datasets.
}
\seealso{
\link{filter_criteria_taxa} for taxa-based filtering using criteria.
}
