% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_classes_mutate_netw.R
\name{mutate_netw}
\alias{mutate_netw}
\alias{mutate_netw,mgnet-method}
\alias{mutate_netw,mgnetList-method}
\title{Mutate Network Data in the mgnet Object}
\usage{
mutate_netw(object, ..., group_taxa)
}
\arguments{
\item{object}{An mgnet object.
This function specifically targets the netw and comm slots, which contain network data and community structures, respectively.}

\item{...}{Quoted expressions that specify how to modify the network data.
The expressions can only use netw and comm as variables, which are bound to the corresponding slots in the mgnet object.}

\item{group_taxa}{Optional; a character vector specifying the columns to group data by before transformations.
Grouping ensures that transformations are contextually applied within each subgroup defined by group_taxa.
If you do not wish to group the data leave it empty.
\subsection{Contextual Variables in mgnet:}{
\itemize{
\item \strong{netw}: Refers to the network slot in the mgnet object where network data and interactions are stored.
\item \strong{comm}: Refers to the community slot in the mgnet object that includes community or clustering data.
}
}}
}
\value{
Returns the modified mgnet object with updated netw and comm slots reflecting the applied transformations.
The rest of the object remains unchanged, ensuring that transformations are isolated to the targeted slots.
}
\description{
This method applies transformations specified by expressions to the netw and comm slots within the mgnet object.
These expressions are evaluated within the context of the object or its subsets, enabling tailored manipulations based on network data.
}
