% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_classes_getters.R
\name{netw}
\alias{netw}
\alias{netw,mgnet-method}
\alias{netw,mgnetList-method}
\title{Retrieve Network Graph from mgnet Objects}
\usage{
netw(object, add_vertex_attr = FALSE)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{add_vertex_attr}{A logical indicating whether to attach all available taxa information as attributes of the vertices. Defaults to FALSE.}
}
\value{
An \code{igraph} object containing the network from an \code{mgnet} object, or a list of \code{igraph} objects
from an \code{mgnetList}, each representing the network graph of a contained \code{mgnet} object.
}
\description{
Retrieves the network graph stored in the \code{network} slot of an \code{mgnet} object
or from each \code{mgnet} object within an \code{mgnetList}. This network graph is typically an \code{igraph} object
representing the interactions among different taxa.
}
