% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{netw}
\alias{netw}
\alias{netw,mgnet-method}
\alias{netw,mgnet}
\title{Retrieves netw.}
\usage{
netw(object)

\S4method{netw}{mgnet}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mgnet-class}}.}
}
\description{
Return the igraph un-directed weighted network.
}
