% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_classes_getters.R
\name{network}
\alias{network}
\alias{network,mgnet-method}
\alias{network,mgnetList-method}
\title{Get Network Graph}
\usage{
network(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
The \code{igraph} object stored in the \code{network} slot for an \code{mgnet} object, or a list of \code{igraph} objects
for an \code{mgnetList} object, each representing the network graph of a contained \code{mgnet} object.
}
\description{
Retrieves the network graph information stored in the \code{network} slot of an \code{mgnet} object
or each \code{mgnet} object within an \code{mgnetList}. This network graph is typically an \code{igraph} object
representing the relationships or interactions among different taxa.
}
