% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ngs_data-class}
\alias{ngs_data-class}
\title{The S4 class to store ngs data.}
\description{
The information is saved in a  \code{\link{matrix}} where the rows express the samples and
the columns the taxa.

\describe{
\item{.Data}{Slot containing \code{\link{matrix}} with ngs data.}
}
}
