% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-base-methods.R
\docType{methods}
\name{nsample}
\alias{nsample}
\alias{nsample,ANY-method}
\alias{nsample,ngs_data-method}
\alias{nsample,sample_metadata-method}
\alias{nsample,mg-method}
\title{Retrieves ngs data.}
\usage{
data(object)

\S4method{nsample}{ANY}(object)

\S4method{nsample}{ngs_data}(object)

\S4method{nsample}{sample_metadata}(object)

\S4method{nsample}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}} or \code{\link{ngs_data-class}}.}
}
\description{
Return the numeric matrix associated with the ngs data from mg or ngs_data
classes.
}
