% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{nsample}
\alias{nsample}
\alias{nsample,mgnet-method}
\alias{nsample,mgnetList-method}
\title{Get Number of Samples}
\usage{
nsample(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, an integer representing the number of samples.
For an \code{mgnetList} object, a named vector of integers, each representing the number
of samples in the corresponding \code{mgnet} objects.
}
\description{
Returns an integer indicating the number of samples in an \code{mgnet} object
or a named vector of integers for each \code{mgnet} object within an \code{mgnetList}.
}
\examples{
data(mg, package = "mgnet")
nsample(mg)  

data(mg, package = "mgnet")
nsample(mgl)  

}
