% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{ntaxa}
\alias{ntaxa}
\alias{ntaxa,mg-method}
\alias{ntaxa,mg}
\title{Get number of taxa}
\usage{
ntaxa(object)

\S4method{ntaxa}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return an integer indicating the number of taxa
}
