% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{ntaxa}
\alias{ntaxa}
\alias{ntaxa,mgnet-method}
\alias{ntaxa,mgnetList-method}
\title{Get Number of Taxa}
\usage{
ntaxa(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, an integer representing the number of taxa.
For an \code{mgnetList} object, a list of integers, each representing the number
of taxa in the corresponding \code{mgnet} objects.
}
\description{
Returns an integer indicating the number of taxa in an \code{mgnet} object
or a list of integers for each \code{mgnet} object within an \code{mgnetList}.
}
