% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_classes_getters.R
\name{pull_meta}
\alias{pull_meta}
\alias{pull_meta,mgnet-method}
\alias{pull_meta,mgnetList-method}
\title{Pull Sample Information from mgnet or mgnetList Objects}
\usage{
pull_meta(object, var = -1)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{var}{The name or position of the column to retrieve from the \code{meta} data frame.
If -1 (default), the last column of the data frame is returned. You can specify the column name unquoted due to non-standard evaluation.}
}
\value{
For a single \code{mgnet} object, a vector containing the data from the specified column
of the \code{meta} data frame is returned. For an \code{mgnetList} object, a list of such vectors
is returned, each corresponding to one \code{mgnet} object in the list.
}
\description{
Retrieves specific sample information from the \code{meta} data frame within \code{mgnet} objects,
or each \code{mgnet} object within an \code{mgnetList}. This function simplifies direct access to specific columns of interest
using dynamic column name handling.
}
\details{
This function supports dynamic evaluation of the column name using \code{rlang} for unquoted names.
}
