% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{ranks}
\alias{ranks}
\alias{ranks,mgnet-method}
\alias{ranks,mgnetList-method}
\title{Get Taxonomic Ranks}
\usage{
ranks(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, a character vector of taxonomic ranks.
For an \code{mgnetList} object, a list of character vectors, each representing
the taxonomic ranks in the corresponding \code{mgnet} objects.
}
\description{
Retrieves the taxonomic classification ranks from an \code{mgnet} object or lists
of taxonomic ranks for each \code{mgnet} object within an \code{mgnetList}.
Taxonomic ranks are derived from the column names of the \code{lineage} slot,
representing different levels of taxonomic classification.
}
