% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{relative}
\alias{relative}
\alias{relative,mgnet,missing-method}
\alias{relative,mgnet,character-method}
\alias{relative,list,missing-method}
\alias{relative,list,character-method}
\title{Get relative abundances.}
\usage{
relative(object,rank)

\S4method{relative}{mgnet,missing}(object)

\S4method{relative}{mgnet,character}(object, rank)

\S4method{relative}{list,missing}(object)

\S4method{relative}{list,character}(object, rank)
}
\arguments{
\item{object}{mgnet or list.}

\item{rank}{(Optional)
See \code{\link{save_sample_sum}}.}
}
\description{
Retrieves the relative abundances of data, normalized by their sample sum at
the taxonomic rank choosen. In mgnet object must be save the variable sample_sum
in meta_sample slot (see \code{\link{save_sample_sum}}).
}
