% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{relative}
\alias{relative}
\alias{relative,mg-method}
\title{Get relative abundances.}
\usage{
relative(object)

\S4method{relative}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Retrieves the relative abundances of data, normalized by their depths.
}
