% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_classes_base_methods.R
\name{sample_id}
\alias{sample_id}
\alias{sample_id,mgnet-method}
\alias{sample_id,mgnetList-method}
\title{Get Sample IDs}
\usage{
sample_id(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, a character vector representing the IDs of samples.
For an \code{mgnetList} object, a list of character vectors, each representing the sample IDs
in the corresponding \code{mgnet} objects.
}
\description{
Returns the names of samples as a character vector from an \code{mgnet} object
or lists of sample IDs for each \code{mgnet} object within an \code{mgnetList}.
}
