% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{sample_info}
\alias{sample_info}
\alias{sample_info,mg-method}
\title{Get sample metadata variables.}
\usage{
sample_info(object)

\S4method{sample_info}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return sample metadata variables number as character vector.
}
