% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{save_log_data}
\alias{save_log_data}
\alias{save_log_data,list-method}
\title{Save log-transformed data in mgnet.}
\usage{
save_log_data(object, method = "base")

\S4method{save_log_data}{list}(object, method = "base")
}
\arguments{
\item{object}{mgnet or list}

\item{method}{character with possible values "base" or "inter-quantile".}
}
\description{
Store in mgnet the log-transformed data to dealing with compositional issues.
}
