% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_classes_utilities_link.R
\name{select_link}
\alias{select_link}
\title{Select Links in mgnet Object Based on Conditions}
\usage{
select_link(object, ..., .suffix = c("_1", "_2"))
}
\arguments{
\item{object}{An \code{mgnet} object.}

\item{...}{Conditions to be applied for selecting links, passed as expressions.}

\item{.suffix}{Character vector specifying suffixes for identifying nodes in links.}
}
\value{
Returns the \code{mgnet} object with the \code{selected_links} attribute set to the indices of the filtered links.
}
\description{
This method filters links in the \code{mgnet} object according to specified conditions and stores the indices
of the selected links as an attribute of the object.
}
