% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{select_link}
\alias{select_link}
\alias{select_link,mgnet-method}
\alias{select_link,mgnets-method}
\title{Select Links in an \code{mgnet} Object Based on Conditions}
\usage{
select_link(object, ...)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object containing a network (edges).}

\item{...}{Conditions to be applied for selecting links, passed as expressions.}
}
\value{
The same \code{mgnet} (or \code{mgnets}) object, updated with a \code{selected_links}
attribute containing the numeric indices of the filtered links.
}
\description{
This method \strong{highlights} links in the \code{mgnet} object according to specified
conditions and stores the indices of these selected links in the \code{selected_links}
attribute of the object.

In particular, you can use \code{select_link()} \emph{inside} a pipeline that calls
\code{\link[=mutate_netw]{mutate_netw()}} to focus on only the selected edges in subsequent network operations.
}
\details{
Inside these expressions, you can call the local helper functions documented in
\code{\link[=helper_link]{helper_link()}}, such as \code{from(var)}, \code{to(var)}, or \code{either(expr)}.
These helpers allow you to reference and filter based on node-level metadata from
the “from” or “to” side of each link.
}
\seealso{
\code{\link[=mutate_link]{mutate_link()}} for modifying link attributes,
\code{\link[=get_selected_links]{get_selected_links()}} and \code{\link[=deselect_link]{deselect_link()}} for managing selection,
\code{\link[=helper_link]{helper_link()}} for details on the local helper functions.
}
