% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_sample}
\alias{selection_sample}
\alias{selection_sample,mgnet-method}
\alias{selection_sample,list-method}
\title{Selection a subset of samples from mgnet object.}
\usage{
selection_sample(object,...,condition="AND")

\S4method{selection_sample}{mgnet}(object, ..., condition = "AND")

\S4method{selection_sample}{list}(object, ..., condition = "AND")
}
\arguments{
\item{object}{mgnet.}

\item{...}{arbitrary set of vectors containing the (numeric or name) or logical indices containing the samples to be preserved}

\item{condition}{(Optional default "AND") logical operation between indices.}
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which samples to keep.
}
