% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_sample}
\alias{selection_sample}
\alias{selection_sample,mgnet-method}
\alias{selection_sample,list-method}
\title{Selection a subset of samples from mgnet object.}
\usage{
selection_sample(object,...,condition="AND")

\S4method{selection_sample}{mgnet}(object, ..., condition = "AND")

\S4method{selection_sample}{list}(object, ..., condition = "AND")
}
\arguments{
\item{object}{mgnet.}

\item{...}{arbitrary set of vectors containing the (numeric or name) or logical indices containing the samples to be preserved}

\item{condition}{(Optional default "AND") logical operation between indices.}
=======
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{selection_sample}
\alias{selection_sample}
\alias{selection_sample,mg,vector-method}
\alias{selection_sample,mg,vector}
\alias{selection_sample,list,vector-method}
\alias{selection_sample,list,vector}
\title{Selection a subset of samples from mg object.}
\usage{
selection_sample(object, idx)

\S4method{selection_sample}{mg,vector}(object, idx)

\S4method{selection_sample}{list,vector}(object, idx)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{idx}{(Required) Vector of integer position indices or logical (like to
\code{[} extractor function).}
>>>>>>> main
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which samples to keep.
}
