% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_taxa}
\alias{selection_taxa}
\alias{selection_taxa,mgnet-method}
\alias{selection_taxa,list-method}
\title{Selection a subset of taxa from mgnet object.}
\usage{
selection_taxa(object, ..., condition="AND")

\S4method{selection_taxa}{mgnet}(object, ..., condition = "AND")

\S4method{selection_taxa}{list}(object, ..., condition = "AND")
}
\arguments{
\item{object}{mgnet.}

\item{...}{arbitrary set of vectors containing the (numeric or name) or logical indices containing the samples to be preserved}

\item{condition}{(Optional default "AND") logical operation between indices.}
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which taxa to keep.
}
