% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_taxa}
\alias{selection_taxa}
\alias{selection_taxa,mgnet,vector-method}
\alias{selection_taxa,list,vector-method}
\title{Selection a subset of taxa from mgnet object.}
\usage{
selection_taxa(object, idx)

\S4method{selection_taxa}{mgnet,vector}(object, idx)

\S4method{selection_taxa}{list,vector}(object, idx)
}
\arguments{
\item{object}{mgnet.}

\item{idx}{Vector of integer position indices or logical or names of taxa.}
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which taxa to keep.
}
