% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{selection_taxa}
\alias{selection_taxa}
\alias{selection_taxa,mgnet-method}
\alias{selection_taxa,list-method}
\title{Selection a subset of taxa from mgnet object.}
\usage{
selection_taxa(object, ..., condition="AND",trim="yes")

\S4method{selection_taxa}{mgnet}(object, ..., condition = "AND", trim = "yes")

\S4method{selection_taxa}{list}(object, ..., condition = "AND", trim = "yes")
}
\arguments{
\item{object}{mgnet.}

\item{...}{arbitrary set of vectors of functions. If vectors they must contain
numeric or name or logical indices to be preserved. If functions they must return
indices as just explained.}

\item{condition}{(Optional default "AND") character indicates the logical operation between indices.}

\item{trim}{character with possible values "yes","no","aggregate" that
control the behavior of discarded taxa as:
\itemize{
\item{"yes":}{ remove the discared taxa from obj}
\item{"no":}{ set to zero the abundances of discared taxa.}
\item{"aggregate":}{ merge togheter all discarded taxa summing the abundances and creating a merged taxa info.}
}}
}
\description{
The function takes as input a vector of logical or position
indices to evaluate which taxa to keep.
}
