% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7c_classes_taxa_color.R
\name{set_community_color}
\alias{set_community_color}
\alias{set_community_color,mgnet-method}
\title{Set Color Mapping to Community in an \code{mgnet} Object}
\usage{
set_community_color(
  object,
  n,
  isolated_color = "#FFFFFF",
  alpha = 1,
  colorspace = "pretty",
  color_to,
  distinct_color = TRUE
)
}
\description{
Set Color Mapping to Community in an \code{mgnet} Object
}
\seealso{
\link[mgnet]{colormap_community}
}
