% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7c_classes_lineage_color.R
\name{set_lineage_color}
\alias{set_lineage_color}
\title{Set Color Mapping for a Specified Taxonomic Rank in an mgnet Object}
\usage{
set_lineage_color(
  object,
  rank,
  distinctColor,
  extraColor = "#FFFFFF",
  alpha = 1,
  colorspace = "pretty",
  color_to
)
}
\arguments{
\item{rank}{A character string specifying the taxonomic rank at which to assign colors.
The rank must be a column name within the \code{lineage} data frame of the \code{mgnet} object.}

\item{colorspace}{The color space setting for generating the palette (defaults to "pretty").
Other options include 'pretty_dark', 'rainbow', 'pastels', or a custom list specifying
color space parameters.}

\item{mgnet_obj}{An \code{mgnet} object.}
}
\value{
An updated \code{mgnet} object with color mapping added to the \code{info_taxa}.
}
\description{
This function updates the \code{info_taxa} data frame of an \code{mgnet} object to include a new
column named \code{color}, which contains hexadecimal color codes assigned to each taxon
at the specified taxonomic rank. The colors are generated using the \code{colormap_categorical} function.
}
