% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{strength_mgnet}
\alias{strength_mgnet}
<<<<<<< HEAD
\alias{strength_mgnet,mgnet-method}
\alias{strength_mgnet,list-method}
\title{Signed Vertices Strengths with Communities Info}
\usage{
strength_mgnet(obj, sign = "all", type = "all")

\S4method{strength_mgnet}{mgnet}(obj, sign = "all", type = "all")

\S4method{strength_mgnet}{list}(obj, sign = "all", type = "all")
=======
\alias{strength_mgnet,mgnet,character,character-method}
\alias{strength_mgnet,mgnet,character,character}
\alias{strength_mgnet,list,character,character-method}
\alias{strength_mgnet-methods,list,character,character}
\alias{strength_mgnet,mgnet,character,missing-method}
\alias{strength_mgnet,mgnet,character,missing}
\alias{strength_mgnet,list,character,missing-method}
\alias{strength_mgnet-methods,list,character,missing}
\title{Signed Vertices Strengths with Communities Info}
\usage{
strength_mgnet(obj, sign, type)

\S4method{strength_mgnet}{mgnet,character,character}(obj, sign, type)

\S4method{strength_mgnet}{list,character,character}(obj, sign, type)

\S4method{strength_mgnet}{mgnet,character,missing}(obj, sign, type)

\S4method{strength_mgnet}{list,character,missing}(obj, sign, type)
>>>>>>> main
}
\arguments{
\item{obj}{mgnet class.}

<<<<<<< HEAD
\item{sign}{(default "all") character indicates the sign of the edges. Possible values are
"positive","negative","all".}

\item{type}{(default "all") character with possible values "intra","extra","all".}
=======
\item{sign}{character indicates the sign of the edges. Possible values are
"positive","negative","all".}

\item{type}{character with possible values "intra","extra","all".}
>>>>>>> main
}
\description{
Calculates the vertices strengths using the information on the
sign of the edges and the communities. The function make possible to distinguish
between positive, negative or both sign of the edges and discern intra or extra
(also both cases) communities edges.
}
