% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_extractors.R
\name{[,mgnetList-method}
\alias{[,mgnetList-method}
\title{Subset mgnetList}
\usage{
\S4method{[}{mgnetList}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{i}{Indices or names specifying the \code{mgnet} objects to retain.}

\item{j}{Not used.}

\item{...}{Additional arguments (currently unused).}

\item{drop}{Not used.}
}
\value{
A new \code{mgnetList} object containing the selected \code{mgnet} objects.
}
\description{
This method subsets an \code{mgnetList} object, returning a new \code{mgnetList} containing only the selected \code{mgnet} objects.
}
