% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_classes_extractors.R
\name{[[,mgnetList-method}
\alias{[[,mgnetList-method}
\title{Extract mgnet Objects from mgnetList}
\usage{
\S4method{[[}{mgnetList}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{mgnetList} object.}

\item{i}{Numeric index, vector of numeric indices, name, or vector of names specifying
the \code{mgnet} object(s) to extract. If missing, an error is thrown.}

\item{j}{Ignored in this method, included for compatibility with the generic method.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Depending on the input \code{i}:
- A single \code{mgnet} object if \code{i} specifies a single index or name.
- A simple list of \code{mgnet} objects if \code{i} specifies multiple indices or names.
}
\description{
This method extracts one or more \code{mgnet} objects from an \code{mgnetList} using either numeric indices
or names. When a single index or name is provided, a single \code{mgnet} object is extracted. When multiple
indices or names are provided, a simple list of \code{mgnet} objects is returned.
}
\details{
Use \code{[[} when you need to work directly with the extracted \code{mgnet} object(s). This method allows
for flexibility in extracting and working with individual or multiple \code{mgnet} objects outside
of an \code{mgnetList} structure.
}
\seealso{
\code{\link[base]{Extract}} for base R extraction operations.
}
