% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{taxa}
\alias{taxa}
\alias{taxa,mgnet-method}
\alias{taxa,list-method}
\title{Retrieves taxonomy table.}
\usage{
taxa(object)

\S4method{taxa}{mgnet}(object)

\S4method{taxa}{list}(object)
}
\arguments{
\item{object}{mgnet or list}
}
\description{
Return the matrix associated with taxonomy classification from mg class.
}
