% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-getters-setters-extractors-show.R
\docType{methods}
\name{taxa}
\alias{taxa}
\alias{taxa,ANY-method}
\alias{taxa,mg-method}
\title{Retrieves taxonomy table.}
\usage{
taxa(object)

\S4method{taxa}{ANY}(object)

\S4method{taxa}{mg}(object)
}
\arguments{
\item{object}{Required \code{\link{mg-class}}.}
}
\description{
Return the matrix associated with taxonomy classification from mg class.
}
\seealso{
\code{\link{taxonomy_table}}
}
