% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_classes_getters.R
\name{taxa}
\alias{taxa}
\alias{taxa,mgnet-method}
\alias{taxa,mgnetList-method}
\title{Get Taxa Metadata Information}
\usage{
taxa(object, .fmt = "df")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are:
- "df": Returns the output as a \code{data.frame}.
- "tbl": Returns the output as a \code{tibble}. For \code{mgnet} objects, the the row names of
the abundance matrix are converted into a new column named \code{taxa_id}, ensuring alignment
with the reserved keyword in \code{mgnet-class}.}
}
\value{
The content of the \code{taxa} slot for \code{mgnet} or a list of such contents for \code{mgnetList}.
}
\description{
Retrieves the taxa information stored in the \code{taxa} slot of an \code{mgnet} object
or for each \code{mgnet} object within an \code{mgnetList}, with the option to format the output as
a \code{data.frame}, \code{tibble}, or a combined \code{tibble} for multiple mgnet objects.
}
