% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{taxa}
\alias{taxa}
\alias{taxa,mg-method}
\alias{taxa,mg}
\alias{taxa,list-method}
\alias{taxa,list}
\title{Retrieves taxonomy table.}
\usage{
taxa(object)

\S4method{taxa}{mg}(object)

\S4method{taxa}{list}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return the matrix associated with taxonomy classification from mg class.
}
