% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{taxa_id}
\alias{taxa_id}
\alias{sample_id,mgnet-method}
\alias{sample_id,mgnetList-method}
\title{Get taxa IDs from mgnet or mgnetList Objects}
\usage{
taxa_id(object, .fmt = "list")
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object from which to extract taxa IDs.}
}
\value{
Retrieves the IDs of taxa from an \code{mgnet} object or lists of taxa IDs for each
\code{mgnet} object within an \code{mgnetList}.
}
\description{
Get taxa IDs from mgnet or mgnetList Objects
}
\examples{
data(mg, package = "mgnet")
taxa_id(mgl)  

data(mgl, package = "mgnet")
taxa_id(mgl)  

}
\seealso{
\link{mgnet} and \link{mgnetList} for details on the classes.
}
