% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_classes_base_methods.R
\name{taxa_id}
\alias{taxa_id}
\alias{taxa_id,mgnet-method}
\alias{taxa_id,mgnetList-method}
\title{Get Taxa IDs}
\usage{
taxa_id(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnetList} object.}
}
\value{
For an \code{mgnet} object, a character vector representing the IDs of taxa.
For an \code{mgnetList} object, a list of character vectors, each representing
the taxa IDs in the corresponding \code{mgnet} objects.
}
\description{
Retrieves the IDs of taxa from an \code{mgnet} object or lists of taxa IDs for each
\code{mgnet} object within an \code{mgnetList}. Taxa IDs represent unique identifiers
for the taxa present in the dataset.
}
