% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgnet.R
\docType{methods}
\name{taxa_name}
\alias{taxa_name}
\alias{taxa_name,mgnet,missing-method}
\alias{taxa_name,mgnet,character-method}
\alias{taxa_name,list,missing-method}
\alias{taxa_name,list,character-method}
\title{Get taxa name.}
\usage{
taxa_name(object, rank)

\S4method{taxa_name}{mgnet,missing}(object)

\S4method{taxa_name}{mgnet,character}(object, rank)

\S4method{taxa_name}{list,missing}(object)

\S4method{taxa_name}{list,character}(object, rank)
}
\arguments{
\item{object}{mgnet or list.}

\item{rank}{taxonomic level choosen (if it is not set, the finest taxonomic rank is assumed)}
}
\description{
Get taxa name at choosen rank as character vector.
}
