% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{taxa_name}
\alias{taxa_name}
\alias{taxa_name,mg,missing-method}
\alias{taxa_name,mg,missing}
\alias{taxa_name,mg,character-method}
\alias{taxa_name,mg,character}
\alias{taxa_name,list,missing-method}
\alias{taxa_name,list,missing}
\alias{taxa_name,list,character-method}
\alias{taxa_name,list,character}
\title{Get taxa name.}
\usage{
taxa_name(object, rank)

\S4method{taxa_name}{mg,missing}(object)

\S4method{taxa_name}{mg,character}(object, rank)

\S4method{taxa_name}{list,missing}(object)

\S4method{taxa_name}{list,character}(object, rank)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}

\item{rank}{taxonomic level choosen (if it is not set, the finest taxonomic rank is assumed)}
}
\description{
Get taxa name at choosen rank as character vector.
}
