% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mg.R
\docType{methods}
\name{taxaID}
\alias{taxaID}
\alias{taxaID,mg-method}
\alias{taxaID,mg}
\title{Get taxa ID.}
\usage{
taxaID(object)

\S4method{taxaID}{mg}(object)
}
\arguments{
\item{object}{(Required) \code{\link{mg-class}}.}
}
\description{
Return taxonomy id as a character vector.
}
