% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{ungroup_link}
\alias{ungroup_link}
\alias{ungroup_link,mgnet-method}
\alias{ungroup_link,mgnets-method}
\title{Ungroup Links in an mgnet Object}
\usage{
ungroup_link(object)

\S4method{ungroup_link}{mgnet}(object)

\S4method{ungroup_link}{mgnets}(object)
}
\arguments{
\item{object}{An \code{mgnet} or \code{mgnets} object.}
}
\value{
The same object, with no \code{link_groups} attribute.
}
\description{
Removes any existing link grouping in an \code{mgnet} or \code{mgnets} object, clearing
the \code{link_groups} attribute so that link operations within \code{\link[=mutate_link]{mutate_link()}}
are no longer grouped.
}
