% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_classes_utilities.R
\name{update_sample_sum}
\alias{update_sample_sum}
\alias{update_sample_sum,mgnet-method}
\alias{update_sample_sum,mgnetList-method}
\title{Update Sample Sum in \code{mgnet} Objects}
\usage{
update_sample_sum(object)
}
\arguments{
\item{object}{An object of class \code{mgnet} or a list of \code{mgnet} objects contained within an \code{mgnetList}.}
}
\value{
The modified \code{mgnet} or \code{mgnetList} object with updated \code{sample_sum} in the \code{info_sample} slot.
}
\description{
This function updates the \code{sample_sum} field in the \code{info_sample} slot of an \code{mgnet} object
or each \code{mgnet} object within an \code{mgnetList}. The \code{sample_sum} represents the total count/abundance
for each sample, which is crucial for calculating relative abundances. Note that \code{sample_sum},
along with \code{taxa_id} and \code{sample_id}, is a reserved keyword within the \code{mgnet} class and should
not be used as a column name in the \code{info_sample} or any other slot externally.
}
\details{
For \code{mgnet} objects, the function calculates the row sums of the \code{abundance} slot, storing
these sums in the \code{info_sample} slot under the column \code{sample_sum}. If \code{info_sample} does not exist,
it initializes this slot as a data.frame with the \code{sample_sum} column. For \code{mgnetList} objects,
it iteratively applies this process to each contained \code{mgnet} object.
}
