% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_zero_replacement.R
\name{zero_dealing}
\alias{zero_dealing}
\title{Zero Replacement in Compositional Data}
\usage{
zero_dealing(X, method = "const")
}
\arguments{
\item{X}{A numeric matrix representing the compositional data, where each row represents a sample and each column represents a variable/taxon.}

\item{method}{A character string indicating the zero replacement strategy: "unif" for uniform random values or "const" for a constant value replacement.}
}
\value{
A numeric matrix with the same dimensions as \code{X}, where zeros have been replaced according to the specified method.
}
\description{
This function implements the simplest strategies for zero replacement in compositional datasets.
Zero replacement is a critical step in preprocessing compositional data, as it
allows for the application of compositional data analysis techniques which assume
non-zero data (like for log-ratio transformetions).

The "unif" method replaces zeros with random values drawn from a uniform distribution,
with bounds set to be a small fraction of the minimum non-zero value within each sample.

The "const" method, on the other hand, replaces zeros with a constant value,
specifically 65\% of the minimum non-zero value found in each sample.
}
\section{Discussion}{

It's important to note that these methods are among the simplest and most
straightforward approaches to zero replacement. While they provide a practical
solution for dealing with zeros in compositional data, especially in datasets
with a large number of zeros, they may not always be the best choice for all
types of compositional analyses.
}

\references{
Lubbe S. et al, 2021. Comparison of zero replacement strategies for compositional data with large numbers of zeros. Chemometrics and Intelligent Laboratory Systems. 10.1016/j.chemolab.2021.104248
}
\seealso{
The following packages provide a range of methodologies for addressing zeros
in compositional data, from statistical imputation techniques to advanced machine
learning models, offering suitable options for different types of analysis and data complexity.
\itemize{
\item zCompositions-package
\item robCompositions-package
\item deepImp-package
}
}
