% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero-replacement.R
\name{zero_dealing}
\alias{zero_dealing}
\title{Simplest way to replace zero.}
\usage{
zero_dealing(X, mar = 1, type = "plus")
}
\arguments{
\item{X}{compositional matrix.}

\item{mar}{integer indicating the margin of samples (rows=1 or cols=2).}

\item{type}{character indicating the zero replacement strategy. Only two
possible values are available, plus and subs.}
}
\description{
This function implement the simplest replacement of the zeros
in a compositional data sets.
}
